### 2023-08-01 #5716 @editorialbot
[PRE REVIEW]: FeenoX: a cloud-first finite-element(ish) computational engineering tool
**Submitting author:** @gtheler (<a href="http://orcid.org/0000-0002-4142-4980">German Theler</a>)
**Repository:** https://github.com/seamplex/feenox/
**Branch with paper.md** (empty if default branch):
**Version:** v0.3
