### 2023-08-01 #5716 @kyleniemeyer
Hi @gtheler, while we are getting started here assigning an editor and finding reviewers, there are a few edits to your paper that I'd recommend:

1. First, please take a look at the DOI messages above.
2. Your paper's title is a bit confusing and unclear. As you might notice, all software published by JOSS is free and open source, so "free and no-fee" is not necessary. It's also not clear to me what "no-X" means. Could you revise the title?
3. Similarly, your paper spends a lot of space explaining free and open-source software, but this is not really necessary. Instead, could you revise it to focus on the elements of [what a JOSS paper should contain](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)?

Thanks! The reviewers and editor may also have more-specific recommendations.
