+++ src/pdes/laplace/bulk.c
@@ -27,6 +27,8 @@ int feenox_problem_build_volumetric_gauss_point_laplace(element_t *e, unsigned i
 #ifdef HAVE_PETSC
   
   double wdet = feenox_mesh_compute_w_det_at_gauss(e, q, feenox.pde.mesh->integration);
+  
+  // for this problem B = B_G so it should be the same
 //  gsl_matrix *B = feenox_mesh_compute_B_G_at_gauss(e, q, feenox.pde.mesh->integration);
   gsl_matrix *B = feenox_mesh_compute_B_at_gauss(e, q, feenox.pde.mesh->integration);
   
@@ -45,7 +47,7 @@ int feenox_problem_build_volumetric_gauss_point_laplace(element_t *e, unsigned i
   if (feenox.pde.has_jacobian) {
     // TODO: jacobian
   }
-    
+  
   // mass matrix Ht*rho*cp*H
   if (feenox.pde.has_mass) {
     gsl_matrix *H = feenox_mesh_compute_H_Gc_at_gauss(e->type, q, feenox.pde.mesh->integration);
