### 2023-08-02 aa22687
partial
