+++ src/mesh/fem.c
@@ -22,27 +22,9 @@
 #include "../feenox.h"
 
 // inverts a small-size square matrix
-/*
-The inverse of a 2x2 matrix:
-
-| a11 a12 |-1             |  a22 -a12 |
-| a21 a22 |    =  1/DET * | -a21  a11 |
-
-with DET  =  a11a22-a12a21
-
-The inverse of a 3x3 matrix:
-
-| a11 a12 a13 |-1             |   a33a22-a32a23  -(a33a12-a32a13)   a23a12-a22a13  |
-| a21 a22 a23 |    =  1/DET * | -(a33a21-a31a23)   a33a11-a31a13  -(a23a11-a21a13) |
-| a31 a32 a33 |               |   a32a21-a31a22  -(a32a11-a31a12)   a22a11-a21a12  |
-
-with DET  =  a11(a33a22-a32a23)-a21(a33a12-a32a13)+a31(a23a12-a22a13)
-*/
-
 // TODO: virtual methods linked to the element type?
-gsl_matrix *feenox_mesh_matrix_invert(gsl_matrix *direct) {
+gsl_matrix *feenox_mesh_matrix_invert(gsl_matrix *direct, gsl_matrix *inverse) {
 
-  gsl_matrix *inverse = NULL;
   switch (direct->size1) {
     case 1:
       if (inverse == NULL) {
@@ -162,10 +144,7 @@ template<typename Derived> struct determinant_impl<Derived, 3>
 // build the canonic B_c matrix at an arbitrary location xi
 gsl_matrix *feenox_mesh_compute_B_c(element_type_t *element_type, double *xi) {
   
-  gsl_matrix *B_c = NULL;
-  if (B_c == NULL) {
-    feenox_check_alloc_null(B_c = gsl_matrix_alloc(element_type->dim, element_type->nodes));
-  }
+  gsl_matrix *B_c = gsl_matrix_alloc(element_type->dim, element_type->nodes);
   for (int d = 0; d < element_type->dim; d++) {
     for (int j = 0; j < element_type->nodes; j++) {
       gsl_matrix_set(B_c, d, j, element_type->dhdxi(j, d, xi));
@@ -175,33 +154,19 @@ gsl_matrix *feenox_mesh_compute_B_c(element_type_t *element_type, double *xi) {
   return B_c;
 }
 
-// copmpute the gradient of h with respect to x evaluated at any arbitrary location
-gsl_matrix *feenox_mesh_compute_B(element_t *e, double *xi, gsl_matrix *invJ_ref) {
+// compute the gradient of h with respect to x evaluated at any arbitrary location
+gsl_matrix *feenox_mesh_compute_B(element_t *e, double *xi) {
 
-  gsl_matrix *B = NULL;
-  gsl_matrix *invJ = NULL;
+  gsl_matrix *J = feenox_mesh_compute_J(e, xi);
+  gsl_matrix *invJ = feenox_mesh_matrix_invert(J, NULL);
   
-  if (invJ_ref != NULL) {
-    // if the caller already provided a J{-1} we use that one
-    invJ = invJ_ref;
-    
-  } else {
-    // otherwise compute dxidx = J^{-1}
-    gsl_matrix *J = feenox_mesh_compute_J(e, xi);
-    invJ = feenox_mesh_matrix_invert(J);
-    gsl_matrix_free(J);
-  }
-  
-  if (B == NULL) {
[…]
