+++ doc/joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'FeenoX: a cloud-first free no-fee no-X uniX-like finite-element(ish) computational engineering tool'
+title: 'FeenoX: a cloud-first finite-element(ish) computational engineering tool'
 tags:
   - engineering
   - partial differential equations
@@ -22,57 +22,28 @@ syntax-definition: feenox.xml
 
 [FeenoX](https://www.seamplex.com/feenox) is a computational tool to solve (currently)
 
- * dynamical systems written as sets of ODEs/DAEs, or
- * steady or transient heat conduction problems, or
- * steady or quasi-static thermo-mechanical problems, or
- * modal analysis problems, or
- * core-level steady-state neutronics
+ * [dynamical systems written as sets of ODEs/DAEs](https://seamplex.com/feenox/examples/daes.html), or
+ * [steady or transient heat conduction problems](https://seamplex.com/feenox/examples/thermal.html), or
+ * [steady or quasi-static thermo-mechanical problems](https://seamplex.com/feenox/examples/mechanical.html), or
+ * [modal analysis problems](https://seamplex.com/feenox/examples/modal.html), or
+ * [core-level steady-state neutronics](https://seamplex.com/feenox/examples/neutron_diffusion.html)
  
-in such a way that the input is a near-English text file that defines the problem to be solved.
+in such a way that the input is a near-English text file that defines the problem.
 FeenoX can be seen either as a finite-element(ish) tool with a particular design basis.
-Two of the main features of this design basis are
+Two of the main features of this [design basis](https://seamplex.com/feenox/doc/sds.html) are
 
  #. Simple problems ought to have simple inputs.
  #. There should be a one-to-one correspondence between the problem definition and FeenoX's input file (fig. \ref{le10}).
  
-![The NAFEMS LE10 problem statement @national1990nafems and the corresponding FeenoX input illustrating the one-to-one correspondence between them.\label{le10}](nafems-le10-problem-input.svg)
+![The NAFEMS LE10 problem statement [@national1990nafems] and the corresponding FeenoX input illustrating the one-to-one correspondence between the two.\label{le10}](nafems-le10-problem-input.svg)
 
 # Statement of need
 
-To better illustrate FeenoX’s unfair advantage, let us first consider what the options are when an engineer needs to to write a technical report, paper or document:
+As discussed in the [README](https://seamplex.com/feenox/#why-feenox), FeenoX is---in a certain sense---to desktop FEA programs (like [Code_Aster](https://www.code-aster.org/spip.php?rubrique2) with [Salome-Meca](https://www.code-aster.org/V2/spip.php?article303) or [CalculiX](http://www.calculix.de/) with [PrePoMax](https://prepomax.fs.um.si/)) and libraries (like [MoFEM](http://mofem.eng.gla.ac.uk/mofem/html/) or [Sparselizard](http://sparselizard.org/)) what [Markdown](https://commonmark.org/) is to Word and [(La)TeX](https://en.wikipedia.org/wiki/LaTeX), respectively and _deliberately_.
 
-
-| Feature                | Microsoft Word |  Google Docs  |   Markdown   |   (La)TeX    |
-|:-----------------------|:--------------:|:-------------:|:------------:|:------------:|
-| Aesthetics             |  $\times$      | $\times$      | $\checkmark$ | $\checkmark$ |
-| Other formats          |  $\sim$        | $\sim$        | $\checkmark$ | $\sim$       |
-| Traceability           |  $\times$      | $\sim$        | $\checkmark$ | $\checkmark$ |
-| Mobile-friendliness    |  $\times$      | $\checkmark$  | $\checkmark$ | $\times$     |
[…]
