+++ doc/joss/paper.bib
@@ -17,6 +17,7 @@
   Title = {Efficient Management of Parallelism in Object Oriented Numerical Software Libraries},
   Booktitle = {Modern Software Tools in Scientific Computing},
   Editor = {E. Arge and A. M. Bruaset and H. P. Langtangen},
+  doi = {10.1007/978-1-4612-1986-6_8},
   Pages = {163--202},
   Publisher = {Birkh{\"{a}}user Press},
   Year = {1997}
@@ -29,7 +30,7 @@
    number = "3",
    pages = "351--362",
    year = "2005",
-   doi = "https://doi.org/10.1145/1089014.1089019"
+   doi = "10.1145/1089014.1089019"
 }
 @TechReport{slepc-manual,
    author = "J. E. Roman and C. Campos and L. Dalcin and E. Romero and A. Tomas",
@@ -45,4 +46,18 @@
   url={https://books.google.com.ar/books?id=--qwHAAACAAJ},
   year={1990},
   publisher={NAFEMS}
+}
+
+@article{milonga-openfoam,
+title = {Coupled unstructured fine-mesh neutronics and thermal-hydraulics methodology using open software: A proof-of-concept},
+journal = {Annals of Nuclear Energy},
+volume = {115},
+pages = {173-185},
+year = {2018},
+issn = {0306-4549},
+doi = {10.1016/j.anucene.2018.01.021},
+url = {https://www.sciencedirect.com/science/article/pii/S0306454918300215},
+author = {Vitor Vasconcelos and André Santos and Daniel Campolina and Germán Theler and Claubia Pereira},
+keywords = {Neutronics, Thermal-hydraulics, Coupling, Multi-physics, TRIGA, FOSS, Serpent, Simulation, Software},
+abstract = {The development of a fine mesh coupled neutronics/thermal-hydraulics framework mainly using free open source software is presented. The proposed contributions go in two different directions: one, is the focus on the open software approach development, a concept widely spread in many fields of knowledge but rarely explored in the nuclear engineering field; the second, is the use operating system shared memory as a fast and reliable storage area to couple the computational fluid dynamics (CFD) software OpenFOAM to the free and flexible reactor core analysis code milonga. This concept was applied to model the behavior of a TRIGA-IPR-R1 reactor fuel pin in steady-state mode. The macroscopic cross-sections for the model, a set of two-group cross-sections data, were generated using the Serpent code. The results show that this coupled system gives consistent results, encouraging system further development and its use for complex geometries simulations.}
 }
