+++ src/mesh/fem.c
@@ -144,7 +144,7 @@ template<typename Derived> struct determinant_impl<Derived, 3>
 // build the canonic B_c matrix at an arbitrary location xi
 gsl_matrix *feenox_mesh_compute_B_c(element_type_t *element_type, double *xi) {
   
-  gsl_matrix *B_c = gsl_matrix_alloc(element_type->dim, element_type->nodes);
+  gsl_matrix *B_c = gsl_matrix_calloc(element_type->dim, element_type->nodes);
   for (int d = 0; d < element_type->dim; d++) {
     for (int j = 0; j < element_type->nodes; j++) {
       gsl_matrix_set(B_c, d, j, element_type->dhdxi(j, d, xi));
@@ -161,7 +161,7 @@ gsl_matrix *feenox_mesh_compute_B(element_t *e, double *xi) {
   gsl_matrix *invJ = feenox_mesh_matrix_invert(J, NULL);
   
   gsl_matrix *B_c = feenox_mesh_compute_B_c(e->type, xi);
-  gsl_matrix *B = gsl_matrix_alloc(e->type->dim, e->type->nodes);
+  gsl_matrix *B = gsl_matrix_calloc(e->type->dim, e->type->nodes);
   gsl_blas_dgemm(CblasTrans, CblasNoTrans, 1.0, invJ, B_c, 0.0, B);
   
   gsl_matrix_free(B_c);
@@ -174,12 +174,13 @@ gsl_matrix *feenox_mesh_compute_B(element_t *e, double *xi) {
 
 inline gsl_matrix *feenox_mesh_compute_invJ_at_gauss(element_t *e, unsigned int q, int integration) {
   
+  feenox.pde.cache_J = 1;  
   gsl_matrix ***invJ = (feenox.pde.cache_J) ? &e->invJ : &feenox.pde.invJi;
   if (*invJ == NULL) {
     feenox_check_alloc_null(*invJ = calloc(e->type->gauss[integration].Q, sizeof(gsl_matrix *)));
   }
   if ((*invJ)[q] == NULL) {
-    (*invJ)[q] = gsl_matrix_alloc(e->type->dim, e->type->dim);
+    (*invJ)[q] = gsl_matrix_calloc(e->type->dim, e->type->dim);
   } else if (feenox.pde.cache_J) {
     return (*invJ)[q];
   }
@@ -206,12 +207,12 @@ inline gsl_matrix *feenox_mesh_compute_C(element_t *e) {
     if (e->C != NULL) {
       return e->C;
     } else {
-      C = gsl_matrix_alloc(e->type->dim, e->type->nodes);
+      C = gsl_matrix_calloc(e->type->dim, e->type->nodes);
       e->C = C;
     }
   } else if (C == NULL) {
 */
-    C = gsl_matrix_alloc(e->type->dim, e->type->nodes);
+    C = gsl_matrix_calloc(e->type->dim, e->type->nodes);
 
 // }
  
@@ -240,7 +241,7 @@ inline gsl_matrix *feenox_mesh_compute_J(element_t *e, double *xi) {
   //   }
   // }
 
-  gsl_matrix *J = gsl_matrix_alloc(e->type->dim, e->type->dim);
+  gsl_matrix *J = gsl_matrix_calloc(e->type->dim, e->type->dim);
   gsl_matrix *B_c = feenox_mesh_compute_B_c(e->type, xi);
   gsl_matrix *C = feenox_mesh_compute_C(e);
   gsl_blas_dgemm(CblasNoTrans, CblasNoTrans, 1.0, B_c, C, 0.0, J);
[…]
