+++ src/mesh/fem.c
@@ -165,8 +165,11 @@ gsl_matrix *feenox_mesh_compute_B(element_t *e, double *xi) {
   gsl_blas_dgemm(CblasTrans, CblasNoTrans, 1.0, invJ, B_c, 0.0, B);
   
   gsl_matrix_free(B_c);
+  B_c = NULL;
   gsl_matrix_free(invJ);
+  invJ = NULL;
   gsl_matrix_free(J);
+  J = NULL;
   
   return B;
 }
@@ -174,7 +177,6 @@ gsl_matrix *feenox_mesh_compute_B(element_t *e, double *xi) {
 
 inline gsl_matrix *feenox_mesh_compute_invJ_at_gauss(element_t *e, unsigned int q, int integration) {
   
-  feenox.pde.cache_J = 1;  
   gsl_matrix ***invJ = (feenox.pde.cache_J) ? &e->invJ : &feenox.pde.invJi;
   if (*invJ == NULL) {
     feenox_check_alloc_null(*invJ = calloc(e->type->gauss[integration].Q, sizeof(gsl_matrix *)));
@@ -185,36 +187,20 @@ inline gsl_matrix *feenox_mesh_compute_invJ_at_gauss(element_t *e, unsigned int 
     return (*invJ)[q];
   }
 
-  gsl_matrix ***J = (feenox.pde.cache_J) ? &e->J : &feenox.pde.Ji;
-  if ((*J) == NULL) {
-    feenox_check_alloc_null((*J) = calloc(e->type->gauss[integration].Q, sizeof(gsl_matrix *)));
-  }
-  if (feenox.pde.cache_J == 0 || (*J)[q] == NULL) {
-    (*J)[q] = feenox_mesh_compute_J_at_gauss(e, q, integration);
-  }
-
-  feenox_mesh_matrix_invert((*J)[q], (*invJ)[q]);
+  gsl_matrix *J = feenox_mesh_compute_J_at_gauss(e, q, integration);
+  feenox_mesh_matrix_invert(J, (*invJ)[q]);
   return (*invJ)[q];
 }
 
 // matrix with the coordinates
 inline gsl_matrix *feenox_mesh_compute_C(element_t *e) {
   
-  gsl_matrix *C = NULL;
-  // TODO: I don't like this
-/*
-  if (feenox.pde.cache_B) {
-    if (e->C != NULL) {
-      return e->C;
-    } else {
-      C = gsl_matrix_calloc(e->type->dim, e->type->nodes);
-      e->C = C;
-    }
-  } else if (C == NULL) {
-*/
+  gsl_matrix *C = (feenox.pde.cache_J) ? e->C : feenox.pde.C;
+  if (C == NULL) {
     C = gsl_matrix_calloc(e->type->dim, e->type->nodes);
-
-// }
+  } else if (feenox.pde.cache_J) {
+    return C;
+  }
  
   // TODO: this is not cache friendly, is it?
   for (unsigned int j = 0; j < e->type->nodes; j++) {
@@ -245,15 +231,12 @@ inline gsl_matrix *feenox_mesh_compute_J(element_t *e, double *xi) {
   gsl_matrix *B_c = feenox_mesh_compute_B_c(e->type, xi);
   gsl_matrix *C = feenox_mesh_compute_C(e);
   gsl_blas_dgemm(CblasNoTrans, CblasNoTrans, 1.0, B_c, C, 0.0, J);
-  if (feenox.pde.cache_B == PETSC_FALSE) {
-    feenox_free(C);
-  }
   gsl_matrix_free(B_c);
   
   return J;
 }
 
-
+/*
 inline void mesh_compute_x(element_t *e, double *xi, double *x) {
 
[…]
