+++ src/mesh/fem.c
@@ -273,7 +273,7 @@ inline double feenox_mesh_compute_w_det_at_gauss(element_t *e, unsigned int q, i
 
   double **w = (feenox.pde.cache_J) ? &e->w : &feenox.pde.w;
   if ((*w) == NULL) {
-    feenox_check_alloc((*w) = calloc(e->type->gauss[integration].Q, sizeof(double)));
+    (*w) = calloc(e->type->gauss[integration].Q, sizeof(double));
   } else if (feenox.pde.cache_J) {
     return (*w)[q];
   }
@@ -477,28 +477,23 @@ inline gsl_matrix *feenox_mesh_compute_J_at_gauss(element_t *e, unsigned int q, 
 
 inline double *feenox_mesh_compute_x_at_gauss(element_t *e, unsigned int q, int integration) {
 
-  double *x = NULL;
-  if (feenox.pde.cache_B) {
-    if (e->x == NULL) {
-      e->x = calloc(e->type->gauss[integration].Q, sizeof(double *));
-    }
-    if (e->x[q] == NULL) {
-      x = e->x[q] = calloc(3, sizeof(double));
-    } else {
-      return e->x[q];
-    }
-  } else {
-    x = calloc(3, sizeof(double));
+  double ***x = (feenox.pde.cache_J) ? &e->x : &feenox.pde.x;
+  if ((*x) == NULL) {
+    (*x) = calloc(e->type->gauss[integration].Q, sizeof(double));
+  } else if (feenox.pde.cache_J) {
+    return (*x)[q];
   }
+  
+  (*x)[q] = calloc(3, sizeof(double));
   
   for (unsigned int j = 0; j < e->type->nodes; j++) {
     double h = gsl_matrix_get(e->type->gauss[integration].H_c[q], 0, j);
     for (unsigned int d = 0; d < 3; d++) {
-      e->x[q][d] += h * e->node[j]->x[d];
+      (*x)[q][d] += h * e->node[j]->x[d];
     }
   }
 
-  return x;
+  return (*x)[q];
 }
 
 
@@ -509,23 +504,21 @@ inline gsl_matrix *feenox_mesh_compute_H_Gc_at_gauss(element_type_t *element_typ
     if (G == 1) {
       element_type->H_Gc = element_type->gauss[integration].H_c;
       return element_type->H_Gc[q];
+    } else {
+      element_type->H_Gc = calloc(element_type->gauss[integration].Q, sizeof(gsl_matrix *));
     }
-    element_type->H_Gc = calloc(element_type->gauss[integration].Q, sizeof(gsl_matrix *));
   }
   
-  if (element_type->H_Gc[q] != NULL) {
-    return element_type->H_Gc[q];
-  }
+  if (element_type->H_Gc[q] == NULL) {
+    unsigned int J = element_type->nodes;
+    element_type->H_Gc[q] = gsl_matrix_calloc(G, G*J);
   
-  
-  unsigned int J = element_type->nodes;
-  element_type->H_Gc[q] = gsl_matrix_calloc(G, G*J);
-  
-  // TODO: measure order of loops
-  for (unsigned int j = 0; j < J; j++) {
-    double h = gsl_matrix_get(element_type->gauss[integration].H_c[q], 0, j);
-    for (unsigned int g = 0; g < G; g++) {
-      gsl_matrix_set(element_type->H_Gc[q], g, G*j+g, h);
+    // TODO: measure order of loops
+    for (unsigned int j = 0; j < J; j++) {
[…]
