+++ src/pdes/build.c
@@ -212,22 +212,22 @@ int feenox_problem_build_element_volumetric(element_t *this) {
   }
  
   // compute the indices of the DOFs to ensamble
-  feenox_call(feenox_mesh_compute_dof_indices(this, feenox.pde.mesh));
+  PetscInt *l = feenox_mesh_compute_dof_indices(this, feenox.pde.dofs);
 
   if (feenox.pde.has_stiffness)  {
-    petsc_call(MatSetValues(feenox.pde.K, feenox.pde.elemental_size, this->l, feenox.pde.elemental_size, this->l, gsl_matrix_ptr(feenox.pde.Ki, 0, 0), ADD_VALUES));
+    petsc_call(MatSetValues(feenox.pde.K, feenox.pde.elemental_size, l, feenox.pde.elemental_size, l, gsl_matrix_ptr(feenox.pde.Ki, 0, 0), ADD_VALUES));
   }  
   if (feenox.pde.has_mass)  {
-    petsc_call(MatSetValues(feenox.pde.M, feenox.pde.elemental_size, this->l, feenox.pde.elemental_size, this->l, gsl_matrix_ptr(feenox.pde.Mi, 0, 0), ADD_VALUES));
+    petsc_call(MatSetValues(feenox.pde.M, feenox.pde.elemental_size, l, feenox.pde.elemental_size, l, gsl_matrix_ptr(feenox.pde.Mi, 0, 0), ADD_VALUES));
   }
   if (feenox.pde.has_jacobian_K) {
-    petsc_call(MatSetValues(feenox.pde.JK, feenox.pde.elemental_size, this->l, feenox.pde.elemental_size, this->l, gsl_matrix_ptr(feenox.pde.JKi, 0, 0), ADD_VALUES));
+    petsc_call(MatSetValues(feenox.pde.JK, feenox.pde.elemental_size, l, feenox.pde.elemental_size, l, gsl_matrix_ptr(feenox.pde.JKi, 0, 0), ADD_VALUES));
   }
   if (feenox.pde.has_jacobian_b) {
-    petsc_call(MatSetValues(feenox.pde.Jb, feenox.pde.elemental_size, this->l, feenox.pde.elemental_size, this->l, gsl_matrix_ptr(feenox.pde.Jbi, 0, 0), ADD_VALUES));
+    petsc_call(MatSetValues(feenox.pde.Jb, feenox.pde.elemental_size, l, feenox.pde.elemental_size, l, gsl_matrix_ptr(feenox.pde.Jbi, 0, 0), ADD_VALUES));
   }
   if (feenox.pde.has_rhs) {
-    petsc_call(VecSetValues(feenox.pde.b, feenox.pde.elemental_size, this->l, gsl_vector_ptr(feenox.pde.bi, 0), ADD_VALUES));
+    petsc_call(VecSetValues(feenox.pde.b, feenox.pde.elemental_size, l, gsl_vector_ptr(feenox.pde.bi, 0), ADD_VALUES));
   }
 
 #endif  
@@ -265,21 +265,21 @@ int feenox_problem_build_element_natural_bc(element_t *this, bc_data_t *bc_data)
     feenox_call(bc_data->set_natural(bc_data, this, q));
   }
   
-  feenox_call(feenox_mesh_compute_dof_indices(this, feenox.pde.mesh));
+  PetscInt *l = feenox_mesh_compute_dof_indices(this, feenox.pde.dofs);
   if (feenox.pde.has_rhs == PETSC_TRUE) {
-    petsc_call(VecSetValues(feenox.pde.b, feenox.pde.elemental_size, this->l, gsl_vector_ptr(feenox.pde.bi, 0), ADD_VALUES));
[…]
