+++ src/pdes/neutron_diffusion/bulk.c
@@ -58,11 +58,8 @@ int feenox_problem_build_volumetric_gauss_point_neutron_diffusion(element_t *e, 
 
 #ifdef HAVE_PETSC
   
-//  double w = feenox_mesh_compute_w_at_gauss(e, q, feenox.pde.mesh->integration);
-  
-//  feenox_call(feenox_mesh_compute_wHB_at_gauss(e, q));
-  double *x = feenox_mesh_compute_x_at_gauss_if_needed(e, q, neutron_diffusion.space_XS);
-  material_t *material = feenox_mesh_get_material(e);
+  double *x = feenox_fem_compute_x_at_gauss_if_needed(e, q, neutron_diffusion.space_XS);
+  material_t *material = feenox_fem_get_material(e);
   
   gsl_matrix_set_zero(neutron_diffusion.D_G);
   gsl_matrix_set_zero(neutron_diffusion.R);
