+++ src/pdes/mechanical/bulk.c
@@ -107,13 +107,13 @@ int feenox_problem_build_volumetric_gauss_point_mechanical(element_t *e, unsigne
       // TODO: matrix-vector product
       double wh = e->w[q] * gsl_matrix_get(e->type->gauss[feenox.pde.mesh->integration].H_c[q], 0, j);
       if (mechanical.f_x.defined) {
-        gsl_vector_add_to_element(feenox.pde.bi, offset+0, wh * mechanical.f_x.eval(&mechanical.f_x, e->x[q], material));
+        gsl_vector_add_to_element(feenox.fem.bi, offset+0, wh * mechanical.f_x.eval(&mechanical.f_x, e->x[q], material));
       }  
       if (mechanical.f_y.defined) {
-        gsl_vector_add_to_element(feenox.pde.bi, offset+1, wh * mechanical.f_y.eval(&mechanical.f_y, e->x[q], material));
+        gsl_vector_add_to_element(feenox.fem.bi, offset+1, wh * mechanical.f_y.eval(&mechanical.f_y, e->x[q], material));
       }
       if (mechanical.f_z.defined) {
-        gsl_vector_add_to_element(feenox.pde.bi, offset+2, wh * mechanical.f_z.eval(&mechanical.f_z, e->x[q], material));
+        gsl_vector_add_to_element(feenox.fem.bi, offset+2, wh * mechanical.f_z.eval(&mechanical.f_z, e->x[q], material));
       }  
     }
   }
@@ -122,7 +122,7 @@ int feenox_problem_build_volumetric_gauss_point_mechanical(element_t *e, unsigne
 
   // elemental stiffness B'*C*B
   feenox_call(gsl_blas_dgemm(CblasNoTrans, CblasNoTrans, 1.0, mechanical.C, mechanical.B, 0, mechanical.CB));
-  feenox_call(gsl_blas_dgemm(CblasTrans, CblasNoTrans, e->w[q], mechanical.B, mechanical.CB, 1.0, feenox.pde.Ki));
+  feenox_call(gsl_blas_dgemm(CblasTrans, CblasNoTrans, e->w[q], mechanical.B, mechanical.CB, 1.0, feenox.fem.Ki));
 
   // thermal expansion strain vector
   if (mechanical.thermal_expansion_model != thermal_expansion_model_none) {
@@ -133,7 +133,7 @@ int feenox_problem_build_volumetric_gauss_point_mechanical(element_t *e, unsigne
     mechanical.compute_thermal_strain(e->x[q], e->physical_group != NULL ? e->physical_group->material : NULL);
     
     feenox_call(gsl_blas_dgemv(CblasTrans, 1.0, mechanical.C, mechanical.et, 0, mechanical.Cet));
-    feenox_call(gsl_blas_dgemv(CblasTrans, e->w[q], mechanical.B, mechanical.Cet, 1.0, feenox.pde.bi));
+    feenox_call(gsl_blas_dgemv(CblasTrans, e->w[q], mechanical.B, mechanical.Cet, 1.0, feenox.fem.bi));
   }
   
 #endif
