+++ src/pdes/build.c
@@ -121,25 +121,25 @@ int feenox_problem_build_elemental_objects_allocate(element_t *this) {
   
   feenox_call(feenox_problem_build_elemental_objects_free());
       
-  feenox.pde.n_local_nodes = this->type->nodes;
-  feenox.pde.elemental_size = this->type->nodes * feenox.pde.dofs;
+  feenox.fem.n_local_nodes = this->type->nodes;
+  feenox.fem.elemental_size = this->type->nodes * feenox.pde.dofs;
   
   if (feenox.pde.has_stiffness) {
-    feenox_check_alloc(feenox.pde.Ki = gsl_matrix_calloc(feenox.pde.elemental_size, feenox.pde.elemental_size));
+    feenox_check_alloc(feenox.fem.Ki = gsl_matrix_calloc(feenox.fem.elemental_size, feenox.fem.elemental_size));
   }  
   if (feenox.pde.has_mass) {
-    feenox_check_alloc(feenox.pde.Mi = gsl_matrix_calloc(feenox.pde.elemental_size, feenox.pde.elemental_size));
+    feenox_check_alloc(feenox.fem.Mi = gsl_matrix_calloc(feenox.fem.elemental_size, feenox.fem.elemental_size));
   }
   if (feenox.pde.has_jacobian_K) {
-    feenox_check_alloc(feenox.pde.JKi = gsl_matrix_calloc(feenox.pde.elemental_size, feenox.pde.elemental_size));
+    feenox_check_alloc(feenox.fem.JKi = gsl_matrix_calloc(feenox.fem.elemental_size, feenox.fem.elemental_size));
   }
   if (feenox.pde.has_jacobian_b) {
-    feenox_check_alloc(feenox.pde.Jbi = gsl_matrix_calloc(feenox.pde.elemental_size, feenox.pde.elemental_size));
+    feenox_check_alloc(feenox.fem.Jbi = gsl_matrix_calloc(feenox.fem.elemental_size, feenox.fem.elemental_size));
   }
   if (feenox.pde.has_rhs) {
-    feenox_check_alloc(feenox.pde.bi = gsl_vector_calloc(feenox.pde.elemental_size));
-    if (feenox.pde.vec_f == NULL) {
-      feenox_check_alloc(feenox.pde.vec_f = gsl_vector_calloc(feenox.pde.dofs));
+    feenox_check_alloc(feenox.fem.bi = gsl_vector_calloc(feenox.fem.elemental_size));
+    if (feenox.fem.vec_f == NULL) {
+      feenox_check_alloc(feenox.fem.vec_f = gsl_vector_calloc(feenox.pde.dofs));
     }
   }
   
@@ -154,14 +154,14 @@ int feenox_problem_build_elemental_objects_free(void) {
 
 #ifdef HAVE_PETSC
   
-  if (feenox.pde.n_local_nodes != 0 && feenox.pde.elemental_size != 0) {
-    gsl_matrix_free(feenox.pde.Ki);
-    gsl_matrix_free(feenox.pde.Mi);
-    gsl_vector_free(feenox.pde.bi);
+  if (feenox.fem.n_local_nodes != 0 && feenox.fem.elemental_size != 0) {
+    gsl_matrix_free(feenox.fem.Ki);
+    gsl_matrix_free(feenox.fem.Mi);
+    gsl_vector_free(feenox.fem.bi);
   }
   
-  feenox.pde.n_local_nodes = 0;
-  feenox.pde.elemental_size = 0;
+  feenox.fem.n_local_nodes = 0;
+  feenox.fem.elemental_size = 0;
   
 #endif  
[…]
