+++ src/pdes/laplace/bc.c
@@ -85,7 +85,7 @@ int feenox_problem_bc_set_laplace_derivative(bc_data_t *this, element_t *e, unsi
 //    gsl_matrix *H = feenox_fem_compute_H_Gc_at_gauss(e->type, q, feenox.pde.mesh->integration);
     gsl_matrix *H = e->type->H_Gc[q];
     // mind the positive sign!
-    feenox_call(gsl_blas_dgemm(CblasTrans, CblasNoTrans, +e->w[q]*dderivativedphi, H, H, 1.0, feenox.pde.Jbi));
+    feenox_call(gsl_blas_dgemm(CblasTrans, CblasNoTrans, +e->w[q]*dderivativedphi, H, H, 1.0, feenox.fem.Jbi));
   }
   
 #endif
