+++ src/pdes/fem.c
@@ -21,6 +21,52 @@
  */
 #include "../feenox.h"
 
+
+int feenox_fem_elemental_caches_reset(void) {
+  
+  if (feenox.fem.current_gauss_type != NULL) {
+    for (unsigned int q = 0; q < feenox.fem.current_gauss_type->gauss[feenox.pde.mesh->integration].Q; q++) {
+      if (feenox.fem.Ji != NULL) {
+        gsl_matrix_free(feenox.fem.Ji[q]);
+        feenox.fem.Ji[q] = NULL;
+      }
+
+      if (feenox.fem.invJi != NULL) {
+        gsl_matrix_free(feenox.fem.invJi[q]);
+        feenox.fem.invJi[q] = NULL;
+      }
+
+      if (feenox.fem.Bi != NULL) {
+        gsl_matrix_free(feenox.fem.Bi[q]);
+        feenox.fem.Bi[q] = NULL;
+      }
+
+      if (feenox.fem.B_Gi != NULL) {
+        gsl_matrix_free(feenox.fem.B_Gi[q]);
+        feenox.fem.B_Gi[q] = NULL;
+      }
+    }
+  }
+  
+  feenox_free(feenox.fem.w);
+  
+  gsl_matrix_free(feenox.fem.C);
+  feenox.fem.C = NULL; 
+  
+  feenox_free(feenox.fem.Ji);
+  feenox_free(feenox.fem.invJi);
+  feenox_free(feenox.fem.Bi);
+  feenox_free(feenox.fem.B_Gi);
+  
+  feenox.fem.current_gauss_element_tag = 0;
+  feenox.fem.current_gauss_type = NULL;
+  
+  return FEENOX_OK;
+
+}
+
+
+
 // inverts a small-size square matrix
 // TODO: virtual methods linked to the element type?
 gsl_matrix *feenox_fem_matrix_invert(gsl_matrix *direct, gsl_matrix *inverse) {
@@ -188,11 +234,14 @@ inline gsl_matrix *feenox_fem_compute_invJ_at_gauss(element_t *e, unsigned int q
 
 // matrix with the coordinates
 inline gsl_matrix *feenox_fem_compute_C(element_t *e) {
+  if (feenox.fem.cache_J == 0 && e->type != feenox.fem.current_gauss_type) {
+    feenox_fem_elemental_caches_reset();
+  }
   
   gsl_matrix **C = (feenox.fem.cache_J) ? &e->C : &feenox.fem.C;
   if ((*C) == NULL) {
     (*C) = gsl_matrix_calloc(e->type->dim, e->type->nodes);
-  } else if (feenox.fem.cache_J || e->tag == feenox.fem.current_element_tag) {
+  } else if (feenox.fem.cache_J || e->tag == feenox.fem.current_gauss_element_tag) {
     return (*C);
   }
   
@@ -202,7 +251,7 @@ inline gsl_matrix *feenox_fem_compute_C(element_t *e) {
       gsl_matrix_set((*C), d, j, e->node[j]->x[d]);
     }
   }
-  feenox.fem.current_element_tag = e->tag;
+  feenox.fem.current_gauss_element_tag = e->tag;
   
   return (*C);
 }
@@ -251,16 +300,23 @@ inline material_t *feenox_fem_get_material(element_t *e) {
 
 inline double feenox_fem_compute_w_det_at_gauss(element_t *e, unsigned int q, int integration) {
 
+  if (feenox.fem.cache_J == 0 && e->type != feenox.fem.current_gauss_type) {
+    feenox_fem_elemental_caches_reset();
+  }
+  
   double **w = (feenox.fem.cache_J) ? &e->w : &feenox.fem.w;
   if ((*w) == NULL) {
[…]
