+++ tests/laplace2d.sh
@@ -20,5 +20,6 @@ exitifwrong $?
 answer poisson-square.fee "0.2949 0.2044 -0.2044"
 exitifwrong $?
 
+gmsh -v 0 -2 ${dir}/maze.geo || exit $?
 answerzero maze.fee
 exitifwrong $?
