### 2023-08-04 292c3b4
cleaned up thermal
