+++ src/pdes/laplace/bulk.c
@@ -38,7 +38,7 @@ int feenox_problem_build_volumetric_gauss_point_laplace(element_t *e, unsigned i
   material_t *material = feenox_fem_get_material(e);
   
   // right-hand side
-  double *x = feenox_fem_compute_x_at_gauss_if_needed(e, q, laplace.space_dependent_source || laplace.space_dependent_mass);
+  double *x = feenox_fem_compute_x_at_gauss_if_needed(e, q, feenox.pde.mesh->integration, laplace.space_dependent_source || laplace.space_dependent_mass);
   if (laplace.f.defined) {
     double f = laplace.f.eval(&laplace.f, x, material);
     feenox_call(feenox_problem_rhs_set(e, q, &f));
