+++ src/mesh/mesh.c
@@ -212,13 +212,14 @@ int feenox_instruction_mesh_read(void *arg) {
         for (size_t i = 0; i < physical_group->n_elements; i++) {
           element_t *element = &this->element[physical_group->element[i]];
           for (unsigned int q = 0; q < element->type->gauss[this->integration].Q; q++) {
-            feenox_call(feenox_fem_compute_w_det_at_gauss(element, q, this->integration));
+            double wdet = feenox_fem_compute_w_det_at_gauss(element, q, this->integration);
+            gsl_matrix *H = element->type->gauss[this->integration].H_c[q];
 
-            for (size_t j = 0; j < element->type->nodes; j++) {
-              double wh = element->w[q] * gsl_matrix_get(element->type->gauss[this->integration].H_c[q], 0, j);
-              physical_group->volume += wh ;
-              for (size_t m = 0; m < 3; m++) {
-                physical_group->cog[m] += wh * element->node[j]->x[m];
+            for (unsigned int j = 0; j < element->type->nodes; j++) {
+              double wdeth = wdet * gsl_matrix_get(H, 0, j);
+              physical_group->volume += wdeth ;
+              for (unsigned int d = 0; d < 3; d++) {
+                physical_group->cog[d] += wdeth * element->node[j]->x[d];
               }
             }
           }

[…]
