+++ src/pdes/neutron_diffusion/bc.c
@@ -74,11 +74,13 @@ int feenox_problem_bc_set_neutron_diffusion_vacuum(bc_data_t *this, element_t *e
   
 #ifdef HAVE_PETSC
 
-//  feenox_call(feenox_mesh_compute_wH_at_gauss(e, q));
   feenox_fem_compute_x_at_gauss_if_needed_and_update_var(e, q, feenox.pde.mesh->integration, this->space_dependent);
-  // TODO: convenience call like  feenox_problem_rhs_set()?
   double coeff = (this->expr.items != NULL) ? feenox_expression_eval(&this->expr) : 0.5;
-  feenox_call(gsl_blas_dgemm(CblasTrans, CblasNoTrans, e->w[q] * coeff, e->type->H_Gc[q], e->type->H_Gc[q], 1.0, feenox.fem.Ki));
+  
+  // TODO: convenience call like  feenox_problem_rhs_set()?
+  double wdet = feenox_fem_compute_w_det_at_gauss(e, q, feenox.pde.mesh->integration);
+  gsl_matrix *H = feenox_fem_compute_H_Gc_at_gauss(e->type, q, feenox.pde.mesh->integration);
+  feenox_call(gsl_blas_dgemm(CblasTrans, CblasNoTrans, wdet*coeff, H, H, 1.0, feenox.fem.Ki));
   
 #endif
   
