+++ src/pdes/modal/parser.c
@@ -50,7 +50,27 @@ int feenox_problem_parse_problem_modal(const char *token) {
 int feenox_problem_parse_post_modal(mesh_write_t *mesh_write, const char *token) {
 
   if (strcmp(token, "all") == 0) {
-    ;
+    
+    char *tokens[3] = {NULL, NULL, NULL};
+    for (unsigned int i = 0; i < feenox.pde.nev; i++) {
+      for (unsigned int g = 0; g < 3; g++) {
+        if (g < feenox.pde.dofs) {
+          asprintf(&tokens[g], "%s%d", feenox.pde.unknown_name[g], i+1);
+        } else {
+          tokens[g] = strdup("0");
+        }
+      }
+      
+      char *mode_name = NULL;
+      asprintf(&mode_name, "mode%d", i+1);
+      feenox_call(feenox_add_post_field(mesh_write, 3, tokens, mode_name, field_location_nodes));
+      feenox_free(mode_name);
+    
+      for (unsigned int g = 0; g < 3; g++) {
+        feenox_free(tokens[g]);
+      }
+    }
+
   } else {
     feenox_push_error_message("undefined keyword '%s' for modal WRITE_RESULTS", token);
     return FEENOX_ERROR;

[…]
