### 2023-08-10 4e3a63d
bimetallic strip test
