### 2023-08-11 53e2fc8
mark sigma, tresca, etc. as used when using LINEARIZE
