### 2023-08-11 75ceb0a
stress linearization in a pipe test
