+++ tests/pipe.sh
@@ -14,11 +14,8 @@ checkmumps
 checkgmsh
 
 # run fem parametric cases
-
-# ust10 gives roundoff error in the integral so we skip it
-# for shape in ust10 uct10 sst10 sct10 ssh20 sch20 ssh27 sch27; do
-for shape in uct10 sst10 sct10 ssh20 sch20 ssh27 sch27; do
+for shape in ust10 uct10 sst10 sct10 ssh20 sch20 ssh27 sch27; do
   gmsh -v 0 -3 ${dir}/pipe-${shape}-2-2.geo || exit $?
-  answerzero1 pipe.fee ${shape} 5e-3
+  answerzero1 pipe.fee ${shape} 1e-2
 done
 
