+++ src/pdes/build.c
@@ -398,9 +398,8 @@ int feenox_problem_rhs_set(element_t *e, unsigned int q, double *value) {
   }  
   
   double wdet = feenox_fem_compute_w_det_at_gauss(e, q, feenox.pde.mesh->integration);
-  gsl_matrix *H_Gc = feenox_fem_compute_H_Gc_at_gauss(e->type, q, feenox.pde.mesh->integration);
+  gsl_matrix *H_Gc = feenox_fem_compute_H_Gc_at_gauss(e, q, feenox.pde.mesh->integration);
   gsl_blas_dgemv(CblasTrans, wdet, H_Gc, feenox.fem.vec_f, 1.0, feenox.fem.bi);
-  // the H_Gc does not have to be freed
 #endif
   
   return FEENOX_OK;
