+++ src/pdes/neutron_diffusion/bc.c
@@ -79,7 +79,7 @@ int feenox_problem_bc_set_neutron_diffusion_vacuum(bc_data_t *this, element_t *e
   
   // TODO: convenience call like  feenox_problem_rhs_set()?
   double wdet = feenox_fem_compute_w_det_at_gauss(e, q, feenox.pde.mesh->integration);
-  gsl_matrix *H = feenox_fem_compute_H_Gc_at_gauss(e->type, q, feenox.pde.mesh->integration);
+  gsl_matrix *H = feenox_fem_compute_H_Gc_at_gauss(e, q, feenox.pde.mesh->integration);
   feenox_call(gsl_blas_dgemm(CblasTrans, CblasNoTrans, wdet*coeff, H, H, 1.0, feenox.fem.Ki));
   
 #endif
