+++ src/parser/parser.c
@@ -2772,7 +2772,7 @@ int feenox_parse_bc(void) {
 ///kw_pde+BC+detail they can be added using as many `PHYSICAL_GROUP` keywords as needed. 
 ///kw_pde+BC+detail If at least one `PHYSICAL_GROUP` is given explicitly, then the `BC` name
 ///kw_pde+BC+detail is not used to try to implicitly link it to a physical group in the mesh.
-    } else if (strcasecmp(token, "PHYSICAL_GROUP") == 0) {
+    } else if (strcasecmp(token, "PHYSICAL_GROUP") == 0 || strcasecmp(token, "GROUP") == 0 || strcasecmp(token, "LABEL") == 0) {
       
       bc->has_explicit_groups = 1;
       
