### 2023-08-14 7c529fa
Gamma function, pi example
