### 2023-08-15 ca7b1da
fixed bunny thermal test
