### 2023-08-15 4d6ecc8
less precision for bunny-modal test
