+++ tests/bunny-modal.sh
@@ -26,11 +26,11 @@ fi
 
 # gmsh -v 0 -3 ${dir}/bunny.geo -order 2 -o bunny-2nd.msh || exit $?
 
-answer1 bunny-modal.fee fixed "0.97 1.10 1.57 1.63 2.05 2.28"
+answer1 bunny-modal.fee fixed "1.0 1.1 1.6 1.6 2.1 2.3"
 exitifwrong $?
 
-answer1 bunny-modal.fee free  "1.56 1.67 2.22 2.42 3.26 3.39"
+answer1 bunny-modal.fee free  "1.6 1.7 2.2 2.4 3.3 3.4"
 exitifwrong $?
 
-answer1 bunny-modal.fee rest  "1.56 1.67 2.23 2.43 3.32 3.48"
+answer1 bunny-modal.fee rest  "1.6 1.7 2.2 2.4 3.3 3.5"
 exitifwrong $?
