+++ examples/daes.md
@@ -303,46 +303,28 @@ Implementation of the dynamical system as described in
     Revisited](http://www.cimec.org.ar/ojs/index.php/mc/article/view/3277/3200),
     by Jeremy Theler, Alejandro Clausse and Fabián J. Bonetto (2010).
 
-The original paper was written using the **first** version of the code,
-named mochin. Recall that FeenoX is a *third system effect*.
+> | The original paper was written using the **first** version of the
+>   code, named mochin.
+> | Recall that FeenoX is a [third system
+>   effect](../doc/sds.md#sec:architecture).
 
-![Figure from the paper above.](boiling-2012-figure.png)
+![Figure from the paper above.](boiling-2010-figure.png)
 
 For reference, the non-dimensional equations are
 
 $$
 \begin{aligned}
-0 &= \, -\frac{1}{h_\text{i}(t)} \cdot \frac{d h_\text{i}}{dt} \left( N_1 - n - \frac{1}{2} \right) \Big[\ell_{n}(t) - \ell_{n-1}(t) \Big]  
-+ \frac{1}{2} \left( \frac{d\ell_n}{dt} + \frac{d\ell_{n-1}}{dt} \right)  \quad\quad & \nonumber \\
-& \quad\quad\quad\quad - u_\text{i}(t) - \frac{N_1}{h_\text{i}(t)} \cdot \frac{N_\text{sub}}{N_\text{pch}} \int_{\ell_{n-1}(t)}^{\ell{n}(t)} q(z,t) \, dz \quad\quad\quad\quad\quad \text{for $n=1,\dots,N_1$}
-\\
-0 &=  u_\text{i}(t) - u_\text{e}(t)  + N_\text{sub} \int_{\lambda(t)}^{1} q(z',t) \, dz' 
-\\
-0 &=  \, \rho_\text{e}(t) - \frac{1}{\displaystyle 1 + N_\text{pch} \cdot \eta(t) \cdot \int_{\lambda(t)}^{1} q(z',t) \, dz'}
-\\
-0 &= \lambda(t) - m(t) + \bigint_{\lambda(t)}^{1} \frac{dz}{\displaystyle 1 + N_\text{pch} \cdot \eta(t) \cdot \int_{\lambda(t)}^{z} q(z',t) \, dz'} 
-\\
-0 &= \varphi(t) - \bigint_{\lambda(t)}^{1} \frac{ \displaystyle u_\text{i}(t) + N_\text{sub}  \int_{\lambda(t)}^{z} q(z',t) \, dz'}{\displaystyle 1 + N_\text{pch}  \cdot \eta(t) \cdot \int_{\lambda(t)}^{z} q(z',t) \, dz'} \, dz
-\\
-0 &= \frac{d m(t)}{dt} + \rho_\text{e}(t) \cdot u_\text{e}(t) - u_\text{i}(t)
-\\
-0 &= 
- \frac{d u_\text{i}(t)}{dt} \cdot \lambda(t) + u_\text{i}(t) \cdot \frac{d \lambda(t)}{dt} + \frac{d\varphi(t)}{dt}
- + \rho_\text{e}(t) \, u_\text{e}^2(t) - \rho_\text{i}(t) \, u_\text{i}^2(t)  \\
-& \quad\quad\quad\quad +
-\Lambda \cdot \left[ u_\text{i}^2(t) \cdot \lambda(t) +
[…]
