+++ src/pdes/thermal/init.c
@@ -29,18 +29,18 @@ int feenox_problem_init_parser_thermal(void) {
   // we are FEM
   feenox.mesh.default_field_location = field_location_nodes;
 
-  feenox.pde.init_runtime_particular = feenox_problem_init_runtime_thermal;
-  feenox.pde.bc_parse = feenox_problem_bc_parse_thermal;
+  feenox.pde.init_before_run = feenox_problem_init_runtime_thermal;
+  feenox.pde.parse_bc = feenox_problem_bc_parse_thermal;
   feenox.pde.setup_ksp = feenox_problem_setup_ksp_thermal;
   feenox.pde.setup_pc = feenox_problem_setup_pc_thermal;
 //  feenox.pde.bc_set_dirichlet = feenox_problem_bc_set_thermal_temperature;
-  feenox.pde.build_element_volumetric_gauss_point = feenox_problem_build_volumetric_gauss_point_thermal;
+  feenox.pde.element_build_volumetric_at_gauss = feenox_problem_build_volumetric_gauss_point_thermal;
   feenox.pde.solve_post = feenox_problem_solve_post_thermal;
-  feenox.pde.feenox_problem_gradient_fill = feenox_problem_gradient_fill_thermal;
-  feenox.pde.feenox_problem_gradient_properties_at_element_nodes = feenox_problem_gradient_properties_at_element_nodes_thermal;
-  feenox.pde.feenox_problem_gradient_fluxes_at_node_alloc = feenox_problem_gradient_fluxes_at_node_alloc_thermal;
-  feenox.pde.feenox_problem_gradient_add_elemental_contribution_to_node = feenox_problem_gradient_add_elemental_contribution_to_node_thermal;
-  feenox.pde.feenox_problem_gradient_fill_fluxes = feenox_problem_gradient_fill_fluxes_thermal;
+  feenox.pde.gradient_fill = feenox_problem_gradient_fill_thermal;
+  feenox.pde.gradient_nodal_properties = feenox_problem_gradient_properties_at_element_nodes_thermal;
+  feenox.pde.gradient_alloc_nodal_fluxes = feenox_problem_gradient_fluxes_at_node_alloc_thermal;
+  feenox.pde.gradient_add_elemental_contribution_to_node = feenox_problem_gradient_add_elemental_contribution_to_node_thermal;
+  feenox.pde.gradient_fill_fluxes = feenox_problem_gradient_fill_fluxes_thermal;
   
   // thermal is a scalar problem
   feenox.pde.dofs = 1;
