+++ src/pdes/modal/init.c
@@ -29,12 +29,12 @@ int feenox_problem_init_parser_modal(void) {
   // we are FEM
   feenox.mesh.default_field_location = field_location_nodes;
   
-  feenox.pde.init_runtime_particular = feenox_problem_init_runtime_modal;
-  feenox.pde.bc_parse = feenox_problem_bc_parse_modal;
+  feenox.pde.init_before_run = feenox_problem_init_runtime_modal;
+  feenox.pde.parse_bc = feenox_problem_bc_parse_modal;
   feenox.pde.setup_eps = feenox_problem_setup_eps_modal;
   feenox.pde.setup_ksp = feenox_problem_setup_ksp_modal;
   feenox.pde.setup_pc = feenox_problem_setup_pc_modal;
-  feenox.pde.build_element_volumetric_gauss_point = feenox_problem_build_volumetric_gauss_point_modal;
+  feenox.pde.element_build_volumetric_at_gauss = feenox_problem_build_volumetric_gauss_point_modal;
   feenox.pde.solve_post = feenox_problem_solve_post_modal;
   
   // move symmetry_axis which is a general PDE setting to
