### 2023-08-24 8db7fb9
refactored virtual methods to have only one resolved by autogen
