+++ src/pdes/neutron_diffusion/methods.h
@@ -24,10 +24,10 @@
 
 // neutron_diffusion/parser.c
 extern int feenox_problem_parse_problem_neutron_diffusion(const char *token);
-extern int feenox_problem_parse_post_neutron_diffusion(mesh_write_t *mesh_write, const char *token);
+extern int feenox_problem_parse_write_post_neutron_diffusion(mesh_write_t *mesh_write, const char *token);
 
 // neutron_diffusion/init.c
-extern int feenox_problem_init_parser_neutron_diffusion(void);
+extern int feenox_problem_parse_time_init_neutron_diffusion(void);
 extern int feenox_problem_init_runtime_neutron_diffusion(void);
 #ifdef HAVE_PETSC
 extern int feenox_problem_setup_pc_neutron_diffusion(PC pc);
