+++ src/pdes/modal/init.c
@@ -23,19 +23,25 @@
 #include "modal.h"
 modal_t modal;
 
-int feenox_problem_init_parser_modal(void) {
+int feenox_problem_parse_time_init_modal(void) {
 
 #ifdef HAVE_SLEPC  
-  // we are FEM
-  feenox.mesh.default_field_location = field_location_nodes;
+  // virtual methods
+  feenox.pde.parse_bc = feenox_problem_bc_parse_modal;
+  feenox.pde.parse_write_results = feenox_problem_parse_write_post_modal;
   
   feenox.pde.init_before_run = feenox_problem_init_runtime_modal;
-  feenox.pde.parse_bc = feenox_problem_bc_parse_modal;
+  
   feenox.pde.setup_eps = feenox_problem_setup_eps_modal;
   feenox.pde.setup_ksp = feenox_problem_setup_ksp_modal;
   feenox.pde.setup_pc = feenox_problem_setup_pc_modal;
+  
   feenox.pde.element_build_volumetric_at_gauss = feenox_problem_build_volumetric_gauss_point_modal;
+  
   feenox.pde.solve_post = feenox_problem_solve_post_modal;
+  
+  // we are FEM
+  feenox.mesh.default_field_location = field_location_nodes;
   
   // move symmetry_axis which is a general PDE setting to
   // the mechanically-particular axisymmetric variant
