+++ src/pdes/thermal/methods.h
@@ -24,10 +24,10 @@
 
 // thermal/parser.c
 extern int feenox_problem_parse_problem_thermal(const char *token);
-extern int feenox_problem_parse_post_thermal(mesh_write_t *mesh_write, const char *token);
+extern int feenox_problem_parse_write_post_thermal(mesh_write_t *mesh_write, const char *token);
 
 // thermal/init.c
-extern int feenox_problem_init_parser_thermal(void);
+extern int feenox_problem_parse_time_init_thermal(void);
 extern int feenox_problem_init_runtime_thermal(void);
 #ifdef HAVE_PETSC
 extern int feenox_problem_setup_pc_thermal(PC pc);

[…]
