### 2023-08-25 a066c76
feenox_physical_group_set_material() and azmy example
