+++ examples/neutron_sn.md
@@ -8,9 +8,12 @@ toc: true
 ...
 # Reed's problem
 
-Reed's problem ([William H. Reed (1971) New Difference Schemes for the
-Neutron Transport Equation, Nuclear Science and Engineering, 46:2,
-309-314, DOI: 10.13182/NSE46-309](http://dx.doi.org/10.13182/NSE46-309))
+Reed's problem as in
+
+-   [William H. Reed, New Difference Schemes for the Neutron Transport
+    Equation, Nuclear Science and Engineering, 46:2, 309-314 (1971).
+    DOI: 10.13182/NSE46-309](http://dx.doi.org/10.13182/NSE46-309)
+
 is a common test problem for transport codes. It is comprised of
 heterogeneous materials with strong absorber, vacuum, and scattering
 regions. These regions are valuable to testing different aspects of
@@ -51,5 +54,198 @@ $
 The solutions obtained in FeenoX with $S_2$, $S_4$, $S_6$ and $S_8$ are plotted and compared against and independent solution from <https://www.drryanmc.com/solutions-to-reeds-problem/>.
 
 ![Solution of the Reed’s problem](reed-flux.svg){width=100%}
+
+
+# Azmy's problem
+
+As in
+
+-   [Y. Y. Azmy, The Weighted Diamond-Difference Form of Nodal Transport
+    Methods. NUCLEAR SCIENCE AND ENGINEERING: 98, 29-40 (1988). DOI:
+    10.13182/NSE88-6](https://doi.org/10.13182/NSE88-6)
+
+![Original problem formulation from 1988 paper](azmy-problem.png)
+
+## Second-order complete structured rectangular grid
+
+This example solves the problem using a structured second-order grid. It
+computes the mean flux in each quadrant by integrating $\phi_1$ over
+each physical group using the instruction
+[`INTEGRATE`](https://www.seamplex.com/feenox/doc/feenox-manual.html#integrate).
+
+![Second-order complete structured rectangular
+grid](azmy-structured-mesh.png)
+
+
+```feenox
+DEFAULT_ARGUMENT_VALUE 1 4
+PROBLEM neutron_sn DIM 2 GROUPS 1 SN $1
+
+READ_MESH $0.msh
+
+MATERIAL src S1=1 Sigma_t1=1 Sigma_s1.1=0.5
+MATERIAL abs S1=0 Sigma_t1=2 Sigma_s1.1=0.1
+
+PHYSICAL_GROUP llq MATERIAL src
+PHYSICAL_GROUP lrq MATERIAL abs
+PHYSICAL_GROUP urq MATERIAL abs
+PHYSICAL_GROUP ulq MATERIAL abs
+
+BC mirror   mirror
+BC vacuum   vacuum
+
+SOLVE_PROBLEM
+
+# compute mean values in each quadrant
+INTEGRATE phi1 OVER llq RESULT lower_left_quadrant
+INTEGRATE phi1 OVER lrq RESULT lower_right_quadrant
+INTEGRATE phi1 OVER urq RESULT upper_right_quadrant
+
+PRINT %.3e "LLQ" lower_left_quadrant/(5*5)  "(ref 1.676e+0)"
+PRINT %.3e "LRQ" lower_right_quadrant/(5*5) "(ref 4.159e-2)"
[…]
