+++ src/pdes/laplace/bc.c
@@ -85,7 +85,8 @@ int feenox_problem_bc_set_laplace_derivative(bc_data_t *this, element_t *e, unsi
     double wdet = feenox_fem_compute_w_det_at_gauss(e, q, feenox.pde.mesh->integration);
     gsl_matrix *H_Gc = feenox_fem_compute_H_Gc_at_gauss(e, q, feenox.pde.mesh->integration);
     // mind the positive sign!
-    feenox_call(gsl_blas_dgemm(CblasTrans, CblasNoTrans, +wdet*dderivativedphi, H_Gc, H_Gc, 1.0, feenox.fem.Jbi));
+    feenox_call(feenox_blas_BtB(H_Gc, +wdet*dderivativedphi, feenox.fem.Jbi));
+    
   }
   
 #endif
