+++ src/pdes/modal/bulk.c
@@ -96,8 +96,7 @@ int feenox_problem_build_volumetric_gauss_point_modal(element_t *e, unsigned int
   double wdet = feenox_fem_compute_w_det_at_gauss(e, q, feenox.pde.mesh->integration);
   
   // elemental stiffness B'*C*B
-  feenox_call(gsl_blas_dgemm(CblasNoTrans, CblasNoTrans, 1.0, modal.C, modal.B, 0, modal.CB));
-  feenox_call(gsl_blas_dgemm(CblasTrans, CblasNoTrans, wdet, modal.B, modal.CB, 1.0, feenox.fem.Ki));
+  feenox_call(feenox_blas_BtCB(modal.B, modal.C, modal.CB, wdet, feenox.fem.Ki));
   
   // elemental mass H'*rho*H
   if (modal.rho.uniform == 0) {
@@ -105,7 +104,7 @@ int feenox_problem_build_volumetric_gauss_point_modal(element_t *e, unsigned int
     modal.rho.eval(&modal.rho, x, feenox_fem_get_material(e));
   }
   gsl_matrix *H_Gc = feenox_fem_compute_H_Gc_at_gauss(e, q, feenox.pde.mesh->integration);
-  feenox_call(gsl_blas_dgemm(CblasTrans, CblasNoTrans, wdet * modal.rho.value, H_Gc, H_Gc, 1.0, feenox.fem.Mi));
+  feenox_call(feenox_blas_BtB(H_Gc, wdet * modal.rho.value, feenox.fem.Mi));
   
 #endif
   
