+++ src/pdes/neutron_diffusion/init.c
@@ -40,6 +40,7 @@ int feenox_problem_parse_time_init_neutron_diffusion(void) {
   feenox.pde.setup_ksp = feenox_problem_setup_ksp_neutron_diffusion;
   feenox.pde.setup_pc = feenox_problem_setup_pc_neutron_diffusion;
   
+  feenox.pde.element_build_volumetric = feenox_problem_build_volumetric_neutron_diffusion;
   feenox.pde.element_build_volumetric_at_gauss = feenox_problem_build_volumetric_gauss_point_neutron_diffusion;
   
   feenox.pde.solve_post = feenox_problem_solve_post_neutron_diffusion;
