### 2023-08-27 232ab63
do not require particular init to set global size
