### 2023-08-29 c8ad7d9
azmy-full example
