### 2023-09-02 fc6ef6b
test for hypre and some 2nd-order thermal slabs
