+++ tests/thermal-1d.sh
@@ -38,7 +38,7 @@ exitifwrong $?
 answerzero thermal-slab-transient-mms.fee 2e-3
 exitifwrong $?
 
-answer thermal-slab-uniform-source-dirichlet-neumann.fee "0.100"
+answerzero thermal-slab-uniform-source-dirichlet-neumann.fee
 exitifwrong $?
 
 
