### 2023-09-02 9371c0d
fix for APPEND and test
