### 2023-09-02 0d18f93
do not need to call feenox_problem_define_solutions()
