+++ src/mesh/elements/triang6.c
@@ -33,7 +33,6 @@ int feenox_mesh_triang6_init(void) {
   int j;
   
   element_type = &feenox.mesh.element_types[ELEMENT_TYPE_TRIANGLE6];
-  feenox_check_alloc(element_type->name = strdup("triang6"));
   element_type->id = ELEMENT_TYPE_TRIANGLE6;
   element_type->dim = 2;
   element_type->order = 2;
