+++ tests/mms/thermal/2d/styles.sh
@@ -1,0 +1,27 @@
+
+declare -A pt
+pt['tri3']=4
+pt['tri6']=18
+pt['quad4']=2
+pt['quad8']=16
+pt['quad9']=17
+
+declare -A lw
+lw['tri3']=1
+lw['tri6']=1
+lw['quad4']=1
+lw['quad8']=1
+lw['quad9']=1
+
+declare -A lt
+lt['delaunay']=4
+lt['frontal']=3
+lt['struct']=2
+
+declare -A co
+co['dirichletstruct']="plum"
+co['dirichletfrontal']="rhodamine"
+co['dirichletdelaunay']="salmon"
+co['neumannstruct']="blue"
+co['neumannfrontal']="cyan"
+co['neumanndelaunay']="seagreen"
