### 2023-09-05 fc60b23
keep track of current tag and gauss point for jacobian and weight
