+++ src/pdes/fem.c
@@ -245,7 +245,7 @@ inline gsl_matrix *feenox_fem_compute_invJ_at_gauss(element_t *e, unsigned int q
 
 // matrix with the coordinates
 inline gsl_matrix *feenox_fem_compute_C(element_t *e) {
-  if (feenox.fem.cache_J == 0 && e->type != feenox.fem.current_gauss_type) {
+  if (feenox.fem.cache_J == PETSC_FALSE && e->type != feenox.fem.current_gauss_type) {
     feenox_fem_elemental_caches_reset();
   }
   
@@ -318,7 +318,7 @@ inline material_t *feenox_fem_get_material(element_t *e) {
 
 inline double feenox_fem_compute_w_det_at_gauss(element_t *e, unsigned int q, int integration) {
 
-  if (feenox.fem.cache_J == 0 && e->type != feenox.fem.current_gauss_type) {
+  if (feenox.fem.cache_J == PETSC_FALSE && e->type != feenox.fem.current_gauss_type) {
     feenox_fem_elemental_caches_reset();
   }
   
@@ -326,7 +326,7 @@ inline double feenox_fem_compute_w_det_at_gauss(element_t *e, unsigned int q, in
   if ((*w) == NULL) {
     (*w) = calloc(e->type->gauss[integration].Q, sizeof(double));
     feenox.fem.current_gauss_type = e->type;
-  } else if (feenox.fem.cache_J && (*w)[q] != 0) {
+  } else if (((feenox.fem.current_gauss_element_tag == e->tag && feenox.fem.current_jacobian_gauss_point == q) || feenox.fem.cache_J) && (*w)[q] != 0) {
     return (*w)[q];
   }
   
@@ -335,6 +335,10 @@ inline double feenox_fem_compute_w_det_at_gauss(element_t *e, unsigned int q, in
   // TODO: choose to complain about zero or negative?
   // TODO: choose to take the absolute value or not? put these two as defines
   (*w)[q] = e->type->gauss[integration].w[q] * fabs(feenox_fem_determinant(J));
+  
+  feenox.fem.current_weight_element_tag = e->tag;
+  feenox.fem.current_weight_gauss_point = q;
+  
   return (*w)[q];
 }
 
@@ -488,7 +492,7 @@ inline gsl_matrix *feenox_fem_compute_J_at_gauss_general(element_t *e, unsigned 
 // magic magic magic!
 inline gsl_matrix *feenox_fem_compute_J_at_gauss(element_t *e, unsigned int q, int integration) {
 
-  if (feenox.fem.cache_J == 0 && e->type != feenox.fem.current_gauss_type) {
+  if (feenox.fem.cache_J == PETSC_FALSE && e->type != feenox.fem.current_gauss_type) {
     feenox_fem_elemental_caches_reset();
   }
   
@@ -499,7 +503,7 @@ inline gsl_matrix *feenox_fem_compute_J_at_gauss(element_t *e, unsigned int q, i
   }
   if ((*J)[q] == NULL) {
     (*J)[q] = gsl_matrix_calloc(e->type->dim, e->type->dim);
-  } else if (feenox.fem.cache_J) {
+  } else if ((feenox.fem.current_gauss_element_tag == e->tag && feenox.fem.current_jacobian_gauss_point == q) || feenox.fem.cache_J) {
     return (*J)[q];
   }
   
@@ -531,6 +535,9 @@ inline gsl_matrix *feenox_fem_compute_J_at_gauss(element_t *e, unsigned int q, i
     (*J)[q] = feenox_fem_compute_J_at_gauss_general(e, q, integration, (*J)[q]);
   }
   
+  feenox.fem.current_jacobian_element_tag = e->tag;
+  feenox.fem.current_jacobian_gauss_point = q;
+
   return (*J)[q];
 }
 
