### 2023-09-05 2ba4f18
init tet4 reduced to 1 gauss point and improved normalization in sn
