### 2023-09-09 59552a4
check return value of asprintf
