### 2023-09-09 1279466
increase dirichlet bc factor
