### 2023-09-09 d84f250
azmy-full parametric example
