+++ examples/azmy-clean.sh
@@ -1,0 +1,1 @@
+rm -f azmy*.msh azmy*.dat azmy*.vtk azmy-full-*.ppl azmy-full-*.pdf azmy-angle-*.geo
