### 2023-09-09 4c202cc
weights for s10
