### 2023-09-10 aead681
fixed  s10 weights
