+++ tests/mms/thermal/2d/README.md
@@ -2,35 +2,20 @@
 
 This directory shows how to perform verification of the 2D thermal solver using the Method of Manufactured Solutions.
 
-## Dependencies
-
- * [FeenoX](https://www.seamplex.com/feenox)
- * [Gmsh](http://gmsh.info/)
- * [Maxima](https://maxima.sourceforge.io/)
- 
-## Execution
-
-Run the script `run.sh` (and `clean.sh` if you want to re-run):
+The chosen domain is a square of size one with the lower left corner at $x=0$ and $y=0$ as in
 
 ```
-./clean.sh
-./run.sh
+SetFactory("OpenCASCADE");
+Rectangle(1) = {0, 0, 0, 1, 1, 0};
+
+Physical Surface("bulk", 1) = {1};
+Physical Curve("left", 2) = {4};
+Physical Curve("right", 3) = {2};
+Physical Curve("bottom", 4) = {1};
+Physical Curve("top", 5) = {3};
 ```
 
-You'll see 8 columns:
-
- 1. logarithm of the mesh size $\log(h)$
- 2. logarithm of the $L-2$ error $\log(e_2)$
- 3. logarithm of the $L-\infty$ error $\log(e_\infty)$
- 4. $c$ factor for mesh size $h \propto 1/c$
- 5. number of cells (volumetric elements)
- 6. number of nodes (degrees of freedom)
- 7. total memory in Mb
- 8. wall time in seconds
- 
-### Changing parameteres
- 
-The assumed temperature solution $T(x,y)$ and the conductivity $k(x,y)$ are read from the main FeenoX input `thermal-square.fee`, like
+The manufactured solution $T(x,y)$ and thermal conductivity $k(x,y)$ are given in the FeenoX input file `thermal-square.fee` as
 
 ```feenox
 # MMS data, set T_mms(x) and k_mms(x) as desired
@@ -38,12 +23,36 @@ T_mms(x,y) = 1 + sin(2*x)^2 * cos(3*y)^2
 k_mms(x,y) = 1 + x - 0.5*y
 ```
 
-Feel free to modify the expressions to something you like.
+As explained below, a script `run.sh` will
+ 
+ 1. perform a parametric sweep over different mesh sizes
+ 2. compute the numerical error with respect to the manufactured solution
+ 3. fit the slope of the logarithm of the error with respect to the logarithm of the mesh size
+ 4. create a Markdown table with the slopes
+ 5. create a Pyxplot file to plot the logarithmic errors
 
 
-Maxima is then called with both $T(x,y)$ and $k(x,y)$ to compute the source term $q(x,y)$ and the partial derivatives of the temperature to compute the heat fluxes in both directions for the Neumann BCs.
+## Dependencies
 
-The `run.sh` script sweeps over (you can edit the script to change what it sweeps over):
+The binaries `feenox`, `gmsh` and `maxima` need to be available to be called from the script.
+
+ * [FeenoX](https://www.seamplex.com/feenox)
+ * [Gmsh](http://gmsh.info/)
+ * [Maxima](https://maxima.sourceforge.io/)
+ * [Pyxplot](https://pyxplot.org.uk/) (optional)
+
+
+ 
+## Execution
+
+Run the script `run.sh` (and `clean.sh` if you want to re-run):
+
+```terminal
+./clean.sh
+./run.sh
+```
+
+By default this will run over a few combinations of 
 
  * BCs
    - dirichlet
@@ -58,18 +67,58 @@ The `run.sh` script sweeps over (you can edit the script to change what it sweep
    - struct
    - frontal
    - dealaunay
[…]
