### 2023-09-11 #5716 @Kevin-Mattheus-Moerman
@AnjaliSandip, @chennachaos, @JaroslavHron, @IgorBaratta, @kinnala, @tjhei, @vijaysm, would you be interested in reviewing this JOSS submission entitled: "FeenoX: a cloud-first finite-element(ish) computational engineering tool"?

As you may know JOSS reviews focuses on the the evaluation of the [the software](https://github.com/seamplex/feenox/), and the review of a [short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05716/joss.05716/10.21105.joss.05716.pdf).

Please let me know if you are interested by responding here. Thanks!
