### 2023-09-11 1ae5010
azmy-full parametric script
