+++ examples/azmy-full.sh
@@ -1,34 +1,104 @@
-#!/bin/sh
+#!/bin/bash
 
 thetas="0 5 15 30 45"
-sns="4 6 8 12"
-scales="4 3 2 1"
+sns="4 6 8 10 12"
+cs="4 3 2 1.5 1"
 
-for theta in $thetas; do
- echo "angle = ${theta};" > azmy-angle-${theta}.geo
- for scale in $scales; do
-  gmsh -v 0 -2 azmy-angle-${theta}.geo azmy-full.geo -clscale ${scale} -o azmy-full-${theta}.msh
-  for sn in $sns; do
-   echo $theta $scale $sn
-   feenox azmy-full.fee ${theta} ${sn} ${scale} --progress
-  done
- done
-done
+thetas="0 15 30 45"
 
-# fixed theta & sn, variable scale
+
+color=(grey blue green tan cyan)
+
+# for theta in ${thetas}; do
+#  echo "angle = ${theta};" > azmy-angle-${theta}.geo
+#  for c in ${cs}; do
+#   gmsh -v 0 -2 azmy-angle-${theta}.geo azmy-full.geo -clscale ${c} -o azmy-full-${theta}.msh
+#   for sn in ${sns}; do
+#    echo ${theta} ${c} ${sn}
+#    feenox azmy-full.fee ${theta} ${sn} ${c} --progress
+#    echo mongocho
+#   done
+#  done
+# done
+
+
+
+# fixed theta & sn, variable c
 rm -f azmy-full-fixed-theta-fixed-sn.ppl
-for theta in $thetas; do
- for sn in $sn; do
+for theta in ${thetas}; do
+ for sn in ${sns}; do
  
   echo "set title \"Fixed \$\\\theta=${theta} \\\textdegree\$, \$N=${sn}\$\"" >> azmy-full-fixed-theta-fixed-sn.ppl
-  echo "set output \"azmy-full-${theta}-${sn}.pdf\"" >> azmy-full-fixed-theta-fixed-sn.ppl 
+  echo "set output \"azmy-full-theta-${theta}-sn-${sn}.pdf\"" >> azmy-full-fixed-theta-fixed-sn.ppl 
   echo -n "plot " >> azmy-full-fixed-theta-fixed-sn.ppl
-  for scale in $scales; do
-   echo -n " \"azmy-full-${theta}-${sn}-${scale}.dat\" w l lw 2*$scale ti \"\$c=${scale}\$\"," >> azmy-full-fixed-theta-fixed-sn.ppl
+  i=0
+  for c in ${cs}; do
+   if [[ "${sn}" != "10" || "${c}" != "1" ]] && [[ "${sn}" != "12" || "${c}" != "1.5" ]] && [[ "${sn}" != "12" || "${c}" != "1" ]]; then
+   
+    if [ $i != 0 ]; then
+      echo -n ", " >> azmy-full-fixed-theta-fixed-sn.ppl
+    fi
+   
+    echo -n " \"azmy-full-${theta}-${sn}-${c}.dat\" w l lw 2*${c} color ${color[${i}]} ti \"\$c=${c}\$\"" >> azmy-full-fixed-theta-fixed-sn.ppl
+    let i++
+    
+   fi
   done
-  echo "0" >> azmy-full-fixed-theta-fixed-sn.ppl
   echo  >> azmy-full-fixed-theta-fixed-sn.ppl
  done
 done
 
 
+
+# fixed theta & c, variable sn
+rm -f azmy-full-fixed-theta-fixed-c.ppl
+for theta in ${thetas}; do
+ for c in ${cs}; do
+ 
[…]
