### 2023-09-14 #5846 @editorialbot
[REVIEW]: FeenoX: a cloud-first free no-fee no-X uniX-like finite-element(ish) computational engineering tool
**Submitting author:** @gtheler (<a href="http://orcid.org/0000-0002-4142-4980">German Theler</a>)
**Repository:** https://github.com/seamplex/feenox/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
