+++ tests/mms/thermal/3d/run.sh
@@ -2,58 +2,60 @@
 
 bcs="dirichlet neumann"
 elems="tet4 tet10 hex8 hex20 hex27"
-algos="struct delaunay hxt"
-# cs="12 16 20 24 28 32 36 38 40"
+# algos="struct delaunay hxt"
+cs="4 8 12 16 20 24 28 32 36 38 40"
 
-bcs="dirichlet"
-elems="hex8"
+# bcs="dirichlet"
+# elems="hex8"
 algos="struct"
-cs="12 16 20"
+# cs="16 60"
+# cs="4 8 12 20"
+ 
+
+# set this flag to 1 if you want to create one VTK for each run
+vtk=1
 
 
-# if this flag is true then the error is computed by first writing
-# T(x,y,z) into a vtk, converting it to .msh with meshio and
-# then reading it back in FeenoX
-# if it is false the error is computed directly in a single call to FeenoX
-vtk=0
-
-for i in feenox gmsh maxima meshio tr; do
+for i in feenox gmsh maxima tr m4; do
  if [ -z "$(which ${i})" ]; then
   echo "error: ${i} not installed"
   exit 1
  fi
 done
 
-
+# n = 1 is steady-state uniform conductivity
+# n = 2 is steady-state space-dependent conductivity
+# n = 3 is steady-state temperature-dependent
+# n = 4 is steady-state space & temperature-dependent
 if [ ! -z "$1" ]; then
  n=$1
 else
  n=1
 fi
 
+# for transient we need to change this line
+m4 -Dn=${n} thermal-cube-static.fee > thermal-cube${n}.fee
+
 . styles.sh
 
 
 # compute the appropriate heat source 
 ## first read temperature and conductivity from FeenoX input
-T=$(grep "T_mms(x,y,z) :=" thermal-cube${n}.fee | sed 's/T_mms(x,y,z)//' | sed 's/:=//')
-T_mms=$(grep "T_mms(x,y,z) :=" thermal-cube${n}.fee)
-k_mms=$(grep "k_mms(x,y,z) :=" thermal-cube${n}.fee)
-
-# cat << EOF
-# ${T_mms};
-# ${k_mms};
-# EOF
+T=$(grep "T_mms(x,y,z) =" thermal-cube${n}-T.fee | sed 's/T_mms(x,y,z)//' | sed 's/=//')
+T_mms=$(grep "T_mms(x,y,z) =" thermal-cube${n}-T.fee | sed 's/=/:=/')
+k_mms=$(grep "k_mms(x,y,z) =" thermal-cube${n}-T.fee | sed 's/=/:=/')
 
 ## call maxima to compute q_mms(x)
+## The T(x,y,z) without the _mms is needed if k depends on T because maxima needs to know the T in the derivative of k wrt T
 maxima --very-quiet << EOF > /dev/null
 ${T_mms};
 ${k_mms};
+T(x,y,z) := ${T}; 
 q_mms(x,y,z) := -(diff(k_mms(x,y,z) * diff(T_mms(x,y,z), x), x) + diff(k_mms(x,y,z) * diff(T_mms(x,y,z), y), y) + diff(k_mms(x,y,z) * diff(T_mms(x,y,z), z), z));
 stringout("thermal-cube${n}-q.txt", q_mms(x,y,z));
-stringout("thermal-cube${n}-qx.txt", k_mms(x,y,z) * diff(T_mms(x,y,z), x));
-stringout("thermal-cube${n}-qy.txt", k_mms(x,y,z) * diff(T_mms(x,y,z), y));
[…]
