+++ tests/mms/thermal/README.md
@@ -1,0 +1,7 @@
+# Verification of the thermal solver with MMS
+
+This directory has two subdirectories with scripts to perform verification usign the Method of Manufactured solutions of the thermal solver in FeenoX.
+
+ * [2D](2d)
+ * [3D](3d)
+
