+++ examples/azmy-full.sh
@@ -1,26 +1,23 @@
 #!/bin/bash
 
-thetas="0 5 15 30 45"
-sns="4 6 8 10 12"
-cs="4 3 2 1.5 1"
-
 thetas="0 15 30 45"
-
-
-color=(grey blue green tan cyan)
+cs="4 3 2 1.5 1"
+sns="4 6 8 10 12"
 
 # for theta in ${thetas}; do
 #  echo "angle = ${theta};" > azmy-angle-${theta}.geo
 #  for c in ${cs}; do
 #   gmsh -v 0 -2 azmy-angle-${theta}.geo azmy-full.geo -clscale ${c} -o azmy-full-${theta}.msh
 #   for sn in ${sns}; do
-#    echo ${theta} ${c} ${sn}
-#    feenox azmy-full.fee ${theta} ${sn} ${c} --progress
-#    echo mongocho
+#    if [ ! -e azmy-full-${theta}-${sn}-${c}.dat ]; then
+#      echo ${theta} ${c} ${sn}
+#      feenox azmy-full.fee ${theta} ${sn} ${c} --progress
+#    fi
 #   done
 #  done
 # done
 
+color=(grey blue green tan cyan)
 
 
 # fixed theta & sn, variable c
@@ -33,10 +30,11 @@ for theta in ${thetas}; do
   echo -n "plot " >> azmy-full-fixed-theta-fixed-sn.ppl
   i=0
   for c in ${cs}; do
-   if [[ "${sn}" != "10" || "${c}" != "1" ]] && [[ "${sn}" != "12" || "${c}" != "1.5" ]] && [[ "${sn}" != "12" || "${c}" != "1" ]]; then
+#    if [[ "${sn}" != "10" || "${c}" != "1" ]] && [[ "${sn}" != "12" || "${c}" != "1.5" ]] && [[ "${sn}" != "12" || "${c}" != "1" ]]; then
+   if [ -e azmy-full-${theta}-${sn}-${c}.dat ]; then
    
     if [ $i != 0 ]; then
-      echo -n ", " >> azmy-full-fixed-theta-fixed-sn.ppl
+     echo -n ", " >> azmy-full-fixed-theta-fixed-sn.ppl
     fi
    
     echo -n " \"azmy-full-${theta}-${sn}-${c}.dat\" w l lw 2*${c} color ${color[${i}]} ti \"\$c=${c}\$\"" >> azmy-full-fixed-theta-fixed-sn.ppl
@@ -49,7 +47,6 @@ for theta in ${thetas}; do
 done
 
 
-
 # fixed theta & c, variable sn
 rm -f azmy-full-fixed-theta-fixed-c.ppl
 for theta in ${thetas}; do
@@ -60,10 +57,11 @@ for theta in ${thetas}; do
   echo -n "plot " >> azmy-full-fixed-theta-fixed-c.ppl
   i=0
   for sn in ${sns}; do
-   if [[ "${sn}" != "10" || "${c}" != "1" ]] && [[ "${sn}" != "12" || "${c}" != "1.5" ]] && [[ "${sn}" != "12" || "${c}" != "1" ]]; then
+#    if [[ "${sn}" != "10" || "${c}" != "1" ]] && [[ "${sn}" != "12" || "${c}" != "1.5" ]] && [[ "${sn}" != "12" || "${c}" != "1" ]]; then
+   if [ -e azmy-full-${theta}-${sn}-${c}.dat ]; then
    
     if [ $i != 0 ]; then
-      echo -n ", " >> azmy-full-fixed-theta-fixed-c.ppl
[…]
