+++ src/pdes/thermal/bulk.c
@@ -44,12 +44,11 @@ int feenox_problem_build_volumetric_gauss_point_thermal(element_t *e, unsigned i
     
   if (feenox.pde.has_jacobian) {
     gsl_matrix *local_vec_T = NULL;
-    unsigned int nodes = e->type->nodes;
-    feenox_check_alloc(local_vec_T = gsl_matrix_calloc(nodes, 1));
+    feenox_check_alloc(local_vec_T = gsl_matrix_calloc(e->type->nodes, 1));
     double T = 0;
     double Tj = 0;
     gsl_matrix *H = feenox_fem_compute_H_Gc_at_gauss(e, q, feenox.pde.mesh->integration);
-    for (unsigned int j = 0; j < nodes; j++) {
+    for (unsigned int j = 0; j < e->type->nodes; j++) {
       Tj = feenox_vector_get(feenox.pde.solution[0]->vector_value, e->node[j]->index_dof[0]);
       gsl_matrix_set(local_vec_T, j, 0, Tj);
       T += gsl_matrix_get(H, 0, j) * Tj; 
@@ -58,7 +57,7 @@ int feenox_problem_build_volumetric_gauss_point_thermal(element_t *e, unsigned i
     if (thermal.temperature_dependent_stiffness) {
       double dkdT = feenox_expression_derivative_wrt_function(thermal.k.expr, feenox.pde.solution[0], T);
 
-      gsl_matrix *BtB = gsl_matrix_calloc(nodes, nodes);
+      gsl_matrix *BtB = gsl_matrix_calloc(e->type->nodes, e->type->nodes);
       feenox_call(feenox_blas_BtB(B, 1.0, BtB));
       feenox_call(feenox_blas_PtCB_accum(BtB, local_vec_T, H, NULL, wdet*dkdT, feenox.fem.JKi));
       gsl_matrix_free(BtB);
