### 2023-09-24 bbc18b9
expose mesh node coordinates as mesh_x mesh_y mesh_z
