### 2023-09-25 e75dfe2
fixes to compile without petsc
