+++ src/parser/parser.c
@@ -3084,11 +3084,11 @@ int feenox_parse_problem(void) {
       
 ///kw_pde+PROBLEM+usage [ CACHE_J ]
     } else if (strcasecmp(token, "CACHE_J") == 0) {
-      feenox.fem.cache_J = PETSC_TRUE;
+      feenox.fem.cache_J = 1;
 ///kw_pde+PROBLEM+usage [ CACHE_B ]
     } else if (strcasecmp(token, "CACHE_B") == 0) {
-      feenox.fem.cache_J = PETSC_TRUE;
-      feenox.fem.cache_B = PETSC_TRUE;
+      feenox.fem.cache_J = 1;
+      feenox.fem.cache_B = 1;
       
 ///kw_pde+PROBLEM+detail If the special variable `end_time` is zero, FeenoX solves a static
 ///kw_pde+PROBLEM+detail  problem---although the variable `static_steps` is still honored.
