### 2023-09-26 88e024a
inflation example
