+++ tests/mms/thermal/README.md
@@ -1,6 +1,6 @@
 # Verification of the thermal solver with MMS
 
-This directory has two subdirectories with scripts to perform verification usign the Method of Manufactured solutions of the thermal solver in FeenoX.
+This directory has two subdirectories with scripts to perform verification using the Method of Manufactured solutions of the thermal solver in [FeenoX](https://www.seamplex.com/feenox).
 
  * [2D](2d)
  * [3D](3d)
