+++ examples/basic.md
@@ -145,6 +145,56 @@ $
 
 
 
+# Financial decisions under inflation
+
+You live in a country with a high inflation rate. A retailer offers you
+a good with two payment options:
+
+a.  30% off the full price plus 3 equal montly payments, or
+b.  full price in 12 equal monthly payments
+
+If the inflation rate is small (large), option a (b) wins.
+
+**Question** What is the monthly inflation rate at which the two options
+are equally good (or bad)?
+
+
+```feenox
+# n = number of payments
+# d = discount 
+# r = monthly inflation rate
+present_value(n, d, r) = sum((1-d)/n * (1/(1+r))^(i-1), i, 1, n)
+r0 = root(present_value(12,0,r) - present_value(3,0.3,r), r, 0, 0.2)
+
+PRINTF "the tipping monthly inflation is %g (that means %.1f%% per month and %.1f%% per year)\n" r0 100*r0 100*(1+r0)^12
+
+PRINT
+
+PRINTF "with 12 payments no discount with r = %g you pay %.3f\n"  r0+0.01 present_value(12,0,r0+0.01)
+PRINTF "with  3 payments and 30%% off with r = %g you pay %.3f\n" r0+0.01 present_value(3,0.3,r0+0.01)
+
+PRINT
+
+PRINTF "with 12 payments no discount with r = %g you pay %.3f\n"  r0-0.01 present_value(12,0,r0-0.01)
+PRINTF "with  3 payments and 30%% off with r = %g you pay %.3f\n" r0-0.01 present_value(3,0.3,r0-0.01)
+```
+
+
+```terminal
+$ feenox inflation.fee 
+the tipping monthly inflation is 0.0931703 (that means 9.3% per month and 291.2% per year)
+
+with 12 payments no discount with r = 0.10317 you pay 0.617
+with  3 payments and 30% off with r = 0.10317 you pay 0.637
+
+with 12 payments no discount with r = 0.0831703 you pay 0.669
+with  3 payments and 30% off with r = 0.0831703 you pay 0.648
+$ 
+
+```
+
+
+
 # The logistic map
 
 Plot the asymptotic behavior of the [logistic
