+++ examples/basic.md
@@ -150,8 +150,8 @@ $
 You live in a country with a high inflation rate. A retailer offers you
 a good with two payment options:
 
-a.  30% off the full price plus 3 equal montly payments, or
-b.  full price in 12 equal monthly payments
+a.  30% off the full price plus 3 equal monthly installments, or
+b.  full price in 12 equal monthly installments
 
 If the inflation rate is small (large), option a (b) wins.
 
