+++ tests/nafems/le10/run.sh
@@ -53,7 +53,7 @@ if [ ! -z "$(which feenox)" ]; then
 fi
 
 # --- Sparselizard ----------------------
-if [ -e "sparselizard/sparselizard" ]; then
+if [ -x "./sparselizard-le10" ]; then
   has_sparselizard="yes"
   rm -f sparselizard_mumps_${m}*.dat
   
@@ -129,13 +129,13 @@ fi
 # has_calculix=""
 # has_reflex=""
 
-# TODO: --check
 if [ "x${1}" == "x--check" ]; then
   echo "FeenoX GAMG:  ${has_feenox}"
   echo "FeenoX MUMPS: ${has_feenox_mumps}"
   echo "Sparselizard: ${has_sparselizard}"
   echo "Code Aster:   ${has_aster}"
   echo "CalculiX:     ${has_calculix}"
+  echo "Reflex:       ${has_reflex}"
   exit 0
 fi
 
@@ -156,7 +156,7 @@ EOF
 if [ ! -z "$(which lstopo)" ]; then
   rm -f arch-${m}.svg
   lstopo arch-${m}.svg
-  echo "![](arch-${m}.svg)\\ " >> arch-${m}.md
+  echo "![Architecture from \`lstopo\`](arch-${m}.svg)" >> arch-${m}.md
 fi
 
 
@@ -164,14 +164,14 @@ fi
 for c in $(feenox steps.fee ${min} ${steps}); do
 
   if [ ! -e le10-${m}-${c}.msh ]; then
-    gmsh -3 le10-${m}.geo -clscale ${c} -o le10-${m}-${c}.msh || exit 1
-    gmsh -3 le10-${m}-${c}.msh -setnumber Mesh.SecondOrderIncomplete 1 -order 2 -o le10_2nd-${m}-${c}.msh || exit 1
+    gmsh -v 0 -3 le10-${m}.geo -clscale ${c} -o le10-${m}-${c}.msh || exit 1
+    gmsh -v 0 -3 le10-${m}-${c}.msh -setnumber Mesh.SecondOrderIncomplete 1 -order 2 -o le10_2nd-${m}-${c}.msh || exit 1
   fi
 
   # ---- FeenoX (GAMG) -----------------------------------------
   if [ ! -z "${has_feenox}" ]; then
     if [ ! -e feenox_gamg_${m}-${c}.sigmay ]; then
-      echo "running FeenoX GAMG m = ${m} c = ${c}"
+      echo "FeenoX GAMG m = ${m} c = ${c}"
       ${time} -o feenox_gamg_${m}-${c}.time feenox le10.fee ${m}-${c} > feenox_gamg_${m}-${c}.sigmay
     fi
     
@@ -189,7 +189,7 @@ for c in $(feenox steps.fee ${min} ${steps}); do
   # ---- FeenoX (MUMPS) -----------------------------------------
   if [ ! -z "${has_feenox_mumps}" ]; then
     if [ ! -e feenox_mumps_${m}-${c}.sigmay ]; then
-      echo "running FeenoX MUMPS c = ${c}"
+      echo "FeenoX MUMPS c = ${c}"
       ${time} -o feenox_mumps_${m}-${c}.time ${feenox_mumps} le10.fee ${m}-${c} --mumps > feenox_mumps_${m}-${c}.sigmay
     fi
 
@@ -208,10 +208,8 @@ for c in $(feenox steps.fee ${min} ${steps}); do
   if [ ! -z "${has_sparselizard}" ]; then
   
     if [ ! -e sparselizard_mumps_${m}-${c}.sigmay ]; then
-      echo "Running Sparselizard m = ${m} c = ${c}"
-      cd sparselizard
[…]
