### 2023-09-26 572602e
mpi_size, mpi_rank and PRINT FROM_ALL_RANKS
