### 2023-09-27 ea13a5d
explicit mpi init and PRINTF_ALL
