+++ src/pdes/solve.c
@@ -103,7 +103,7 @@ int feenox_problem_phi_to_solution(Vec phi) {
   petsc_call(VecGetLocalSize(phi_full, &nlocal));
   if (nlocal != feenox.pde.size_global) {
     feenox_push_error_message("internal check of problem size with scatter failed, %d != %d\n", nlocal, feenox.pde.size_global);
-    return FEENOX_OK;
+    return FEENOX_ERROR;
   }
   petsc_call(VecGetArray(phi_full, &phi_full_array));
   
