### 2023-09-28 5174356
removed MPI_INIT and improved PRINTF_ALL
