+++ src/parser/parser.c
@@ -384,13 +384,6 @@ int feenox_parse_line(void) {
       feenox_call(feenox_parse_solve());
       return FEENOX_OK;
 
-///kw+MPI_INIT+desc Explicitly initialize MPI (not needed for regular users)
-///kw+MPI_INIT+usage MPI_INIT
-      // -----  -----------------------------------------------------------
-    } else if (strcasecmp(token, "MPI_INIT") == 0) {
-      feenox_call(feenox_parse_mpi_init());
-      return FEENOX_OK;
-      
 // TODO: move this to a per-physics parser      
 ///kw_pde+LINEARIZE_STRESS+desc Compute linearized membrane and/or bending stresses according to ASME\ VIII Div\ 2 Sec\ 5.
 ///kw_pde+LINEARIZE_STRESS+usage LINEARIZE_STRESS
@@ -3793,25 +3786,6 @@ int feenox_parse_dump(void) {
   
   return FEENOX_OK;
 }
-
-int feenox_parse_mpi_init(void) {
-
-  mpi_init_t *mpi_init = NULL;
-  feenox_check_alloc(mpi_init = calloc(1, sizeof(mpi_init_t)));
-  
-  char *token = NULL;
-  while ((token = feenox_get_next_token(NULL)) != NULL) {
-    if (strcasecmp(token, "HELLO") == 0) {
-      mpi_init->hello = 1;
-    }
-  }
-  
-  LL_APPEND(feenox.mpi_inits, mpi_init);
-  feenox_call(feenox_add_instruction(feenox_instruction_mpi_init, mpi_init));
-  
-  return FEENOX_OK;
-}
-
 
 int feenox_parse_solve(void) {
   
