### 2023-09-29 60aedae
report rank when catching sigint
