+++ dist/petsc.sh
@@ -24,21 +24,20 @@ cd petsc-${petsc_ver}
 export PETSC_DIR=$(pwd)
 export PETSC_ARCH=${PETSC_ARCH}
 
-# f2cblaslapack is not sse-enabled but download-openblas does not work
-# it fails saying that gcc_s is not found at link time
-./configure --with-mpi=0 \
-            --with-cxx=0 \
-            --with-fortran-bindings=0 \
-            --with-c2html=0 \
-            --with-x=0 \
-            --with-debugging=0 \
-            --with-shared-libraries=0 \
-            --download-openblas \
-            --COPTFLAGS="-O3"
-            
-#            --with-fc=0 \
-#             --download-f2cblaslapack \
-            
+# TODO: neither mpi nor mumps work in static
+./configure \
+  --with-mpi=0 \
+  --with-cxx=0 \
+  --with-fortran-bindings=0 \
+  --with-c2html=0 \
+  --with-bison=0 \
+  --with-x=0 \
+  --with-debugging=0 \
+  --with-shared-libraries=0 \
+  --download-openblas \
+  --download-mpich \
+  --COPTFLAGS="-O3" --FOPTFLAGS="-O3"
+           
 
 # export PETSC_ARCH=arch-linux2-serial-mumps-static
 # ./configure --with-mpi=0 --with-cxx=0 --with-fortran-bindings=0 --with-fc=0 --with-c2html=0 --with-x=0 --with-debugging=0 --with-shared-libraries=0 --with-mumps-serial=1 --download-mumps --download-openblas --COPTFLAGS="-O" --FOPTFLAGS="-O3"
