+++ src/pdes/slepc_eps.c
@@ -49,24 +49,6 @@ int feenox_problem_solve_slepc_eigen(void) {
       petsc_call(EPSSetType(feenox.pde.eps, feenox.pde.eps_type));
     }
     
-    // operators depending on formulation
-    if (feenox.pde.eigen_formulation == eigen_formulation_omega) {
-      petsc_call(EPSSetOperators(feenox.pde.eps, feenox.pde.K_bc, feenox.pde.M_bc));
-    } else {  
-      petsc_call(EPSSetOperators(feenox.pde.eps, feenox.pde.M_bc, feenox.pde.K_bc));
-    }
-
-    // this should be faster but it is not
-    // TODO: let the user choose, let the pdes fill in a type variable
-    // petsc_call(EPSSetProblemType(feenox.pde.eps, (feenox.pde.symmetric_K && feenox.pde.symmetric_M) ? EPS_GHEP : EPS_GNHEP));
-  
-    // specify how many eigenvalues (and eigenvectors) to compute.
-    if (feenox.pde.eps_ncv.items != NULL) {
-      petsc_call(EPSSetDimensions(feenox.pde.eps, feenox.pde.nev, (PetscInt)(feenox_expression_eval(&feenox.pde.eps_ncv)), PETSC_DEFAULT));
-    } else {
-      petsc_call(EPSSetDimensions(feenox.pde.eps, feenox.pde.nev, PETSC_DEFAULT, PETSC_DEFAULT));
-    }
-  
     // get the associated spectral transformation
     ST st = NULL;
     petsc_call(EPSGetST(feenox.pde.eps, &st));
@@ -74,13 +56,28 @@ int feenox_problem_solve_slepc_eigen(void) {
       feenox_push_error_message("cannot retrieve spectral transformation object");
       return FEENOX_ERROR;
     }
-    // we need to set the type to create it, otherwise the icntls do not work
-    petsc_call(STSetType(st, (feenox.pde.st_type != NULL) ? feenox.pde.st_type : (feenox.pde.eigen_formulation == eigen_formulation_omega) ? STSINVERT : STSHIFT));
-    petsc_call(STGetOperator(st,NULL));
+    if (feenox.pde.st_type != NULL) {
+      petsc_call(STSetType(st, feenox.pde.st_type));
 
+      // shift and invert offsets
+      petsc_call(STSetShift(st, feenox_var_value(feenox.pde.vars.eps_st_sigma)));
+      petsc_call(STCayleySetAntishift(st, feenox_var_value(feenox.pde.vars.eps_st_nu)));
+    }  
+    
     if (feenox.pde.setup_eps != NULL) {
       feenox_call(feenox.pde.setup_eps(feenox.pde.eps));
-    } 
+    }  
+
+    // tolerances
+    petsc_call(EPSSetTolerances(feenox.pde.eps, feenox_var_value(feenox.pde.vars.eps_tol),
+                                                (PetscInt)feenox_var_value(feenox.pde.vars.eps_max_it)));
+    
+    // operators depending on formulation
+    if (feenox.pde.eigen_formulation == eigen_formulation_omega) {
+      petsc_call(EPSSetOperators(feenox.pde.eps, feenox.pde.K_bc, feenox.pde.M_bc));
+    } else {  
+      petsc_call(EPSSetOperators(feenox.pde.eps, feenox.pde.M_bc, feenox.pde.K_bc));
+    }
     
     // setup the linear solver
     KSP ksp = NULL;
@@ -89,32 +86,29 @@ int feenox_problem_solve_slepc_eigen(void) {
[…]
