### 2023-10-05 b9d3074
restored gamg_threshold
