### 2023-10-06 cb73883
t21 mesh partitioning test
