+++ src/mesh/gmsh.c
@@ -36,8 +36,6 @@ int feenox_mesh_read_gmsh(mesh_t *this) {
 
   char *dummy = NULL;
   char *name = NULL;
-  physical_group_t *physical_group = NULL;
-  geometrical_entity_t *geometrical_entity = NULL;
   size_t node = 0;
   size_t node_index = 0;
   unsigned int version_maj = 0;
@@ -122,8 +120,7 @@ int feenox_mesh_read_gmsh(mesh_t *this) {
         return FEENOX_ERROR;
       }
  
-      unsigned int i = 0;
-      for (i = 0; i < this->n_physical_names; i++) {
+      for (unsigned int i = 0; i < this->n_physical_names; i++) {
 
         int dimension = 0;
         int tag = 0;
@@ -146,6 +143,7 @@ int feenox_mesh_read_gmsh(mesh_t *this) {
           feenox_check_alloc(name = strdup(buffer));
         }
         
+        physical_group_t *physical_group = NULL;
         if ((physical_group = feenox_get_physical_group_ptr(name, this)) == NULL) {
           // create new physical group
           if ((physical_group = feenox_define_physical_group_get_ptr(name, this, dimension, tag)) == NULL) {
@@ -213,8 +211,7 @@ int feenox_mesh_read_gmsh(mesh_t *this) {
       this->surfaces = num_entities[2];
       this->volumes = num_entities[3];
       
-      size_t i = 0;
-      for (i = 0; i < this->points+this->curves+this->surfaces+this->volumes; i++) {
+      for (size_t i  = 0; i < this->points+this->curves+this->surfaces+this->volumes; i++) {
         unsigned int dimension = 0;
         if (i < this->points) {
           dimension = 0;
@@ -226,7 +223,9 @@ int feenox_mesh_read_gmsh(mesh_t *this) {
           dimension = 3;
         }
         
+        geometrical_entity_t *geometrical_entity = NULL;
         feenox_check_alloc(geometrical_entity = calloc(1, sizeof(geometrical_entity_t)));
+        geometrical_entity->dim = dimension;
         feenox_call(feenox_gmsh_read_data_int(this, 1, &geometrical_entity->tag, binary));
         
         double bbox[6] = {0,0,0,0,0,0};
@@ -295,14 +294,13 @@ int feenox_mesh_read_gmsh(mesh_t *this) {
       // number of entities per dimension
       size_t num_entities[4] = {0, 0, 0, 0};
       feenox_call(feenox_gmsh_read_data_size_t(this, 4, num_entities, binary));
-      this->points = num_entities[0];
-      this->curves = num_entities[1];
-      this->surfaces = num_entities[2];
-      this->volumes = num_entities[3];
+      this->partitioned_points = num_entities[0];
+      this->partitioned_curves = num_entities[1];
+      this->partitioned_surfaces = num_entities[2];
+      this->partitioned_volumes = num_entities[3];
       
-      size_t i = 0;
-      for (i = 0; i < this->partitioned_points+this->partitioned_curves+this->partitioned_surfaces+this->partitioned_volumes; i++) {
-        size_t dimension;
+      for (size_t i = 0; i < this->partitioned_points+this->partitioned_curves+this->partitioned_surfaces+this->partitioned_volumes; i++) {
+        unsigned int dimension = 0;
         if (i < this->partitioned_points) {
           dimension = 0;
         } else if (i < this->partitioned_points+this->partitioned_curves) {
@@ -313,10 +311,12 @@ int feenox_mesh_read_gmsh(mesh_t *this) {
           dimension = 3;
         }
         
+        geometrical_entity_t *geometrical_entity = NULL;
[…]
