### 2023-10-06 ea845f8
is_element_local()
