+++ tests/clean.sh
@@ -1,1 +1,1 @@
-cat .gitignore | sed '/^#.*/ d' | sed '/^\s*$/ d' | sed 's/^/rm -rf /' | bash
+cat .gitignore | sed '/^#.*/ d' | sed '/^\s*$/ d' | grep -v Stanford_Bunny.stl | sed 's/^/rm -rf /' | bash

[…]
