+++ src/pdes/neutron_sn/init.c
@@ -89,7 +89,7 @@ int feenox_problem_parse_time_init_neutron_sn(void) {
   // TODO: document from comments
   for (unsigned int m = 0; m < neutron_sn.directions; m++) {
     for (unsigned int g = 0; g < neutron_sn.groups; g++) {
-      feenox_check_minusone(asprintf(&feenox.pde.unknown_name[m * neutron_sn.groups + g], "psi%d.%d", m+1, g+1));
+      feenox_check_minusone(asprintf(&feenox.pde.unknown_name[m * neutron_sn.groups + g], "psi%u.%u", m+1, g+1));
     }
   }
 
