+++ doc/syntax-feenox.tex
@@ -34,6 +34,7 @@ morekeywords={
       DIMENSIONS,
       DIRICHLET_SCALING,
       DUMP,
+      EIGEN_DIRICHLET_ZERO,
       EIGEN_FORMULATION,
       EIGEN_SOLVER,
       ELSE,
@@ -122,6 +123,7 @@ morekeywords={
       PRECONDITIONER,
       PRINT,
       PRINTF,
+      PRINTF_ALL,
       PRINT_FUNCTION,
       PRINT_VECTOR,
       PROBLEM,
@@ -201,6 +203,8 @@ morekeywords={
       Z_MIN,
       ALLOWED,
       AS_PROVIDED,
+      K,
+      M,
       NONE,
       POST,
       SKIP_HEADER_STEP,
@@ -250,8 +254,10 @@ morekeywords={[3]
       max_dt_0,
       min_dt,
       min_dt_0,
-      ncores,
-      ncores_0,
+      mpi_rank,
+      mpi_rank_0,
+      mpi_size,
+      mpi_size_0,
       on_gsl_error,
       on_gsl_error_0,
       on_nan,
@@ -326,6 +332,8 @@ morekeywords={[4]
       memory,
       min,
       mod,
+      mpi_memory_global,
+      mpi_memory_local,
       not,
       prod,
       quasi_random,
