### 2023-10-12 8b9069f
fix to compile with petsc but without slepc
