### 2023-10-14 cd4ece5
explicit fission spectrum
