### 2023-10-16 d7b432d
removed p52 because it keeps giving out of memory
