+++ tests/los-alamos.sh
@@ -73,9 +73,10 @@ gmsh -v 0 -1 ${dir}/la-p51-UAl-2-0-SL.geo || exit $?
 answerzero la-p51-UAl-2-0-SL.fee
 exitifwrong $?
 
-gmsh -v 0 -3 ${dir}/la-p52-UAl-2-0-SP.geo || exit $?
-answerzero la-p52-UAl-2-0-SP.fee
-exitifwrong $?
+# this one fails with out-of-memory in github actions
+# gmsh -v 0 -3 ${dir}/la-p52-UAl-2-0-SP.geo || exit $?
+# answerzero la-p52-UAl-2-0-SP.fee
+# exitifwrong $?
 
 
 answerzero la-p70-URRa-2-1-IN.fee
