### 2023-10-18 416d1fa
non-homogeneous dirichlet bc and fixed space_XS in diffusion
