### 2023-10-19 a8dcaa1
allow LABEL as an alias for PHYSICAL_GROUP in MATERIAL
