### 2023-10-21 5f29de0
added neutron sources to syntax generator
