+++ doc/syntax-feenox.tex
@@ -56,6 +56,7 @@ morekeywords={
       GAUSS,
       GRADIENT,
       GROUP,
+      GROUPS,
       HANDLE_HANGING_NODES,
       HEADER,
       HISTORY,
@@ -96,6 +97,7 @@ morekeywords={
       MODE,
       MODES,
       MOMENT,
+      N,
       NAME,
       NODE,
       NODES,
@@ -153,6 +155,7 @@ morekeywords={
       SHM_OBJECT,
       SIZE,
       SIZES,
+      SN,
       SNES,
       SNES_TYPE,
       SOLVE,
@@ -250,6 +253,8 @@ morekeywords={[3]
       in_transient_last_0,
       j,
       j_0,
+      keff,
+      keff_0,
       max_dt,
       max_dt_0,
       min_dt,
@@ -274,6 +279,8 @@ morekeywords={[3]
       realtime_scale_0,
       report,
       report_0,
+      sn_alpha,
+      sn_alpha_0,
       static_steps,
       static_steps_0,
       step_static,
