+++ doc/keywords.sh
@@ -9,7 +9,7 @@ done
 
 src="../src"
 parsers="../src/parser/parser.c ../src/pdes/neutron_diffusion/parser.c ../src/pdes/neutron_sn/parser.c "
-inits="../src/flow/init.c ../src/pdes/neutron_diffusion/init.c ../src/pdes/neutron_sn/init.c"
+inits="../src/flow/init.c ../src/pdes/init.c ../src/pdes/neutron_diffusion/init.c ../src/pdes/neutron_sn/init.c"
 functions="../src/math/builtin_functions.c ../src/math/builtin_functionals.c ../src/math/builtin_vectorfunctions.c"
 
 UPPER1=$(grep strcasecmp ${parsers}   | sed -r 's/[^"]*("[^"]*")?/ \1/g;s/" +"/\n/g;s/ *"//g' | sort | mawk '$1 in p{next}{p[$1];print}' | mawk '/^[A-Z]/')
