+++ doc/syntax-feenox.tex
@@ -235,6 +235,16 @@ morekeywords={[3]
       dt_0,
       end_time,
       end_time_0,
+      eps_max_it,
+      eps_max_it_0,
+      eps_st_nu,
+      eps_st_nu_0,
+      eps_st_sigma,
+      eps_st_sigma_0,
+      eps_tol,
+      eps_tol_0,
+      gamg_threshold,
+      gamg_threshold_0,
       i,
       i_0,
       infinite,
@@ -255,20 +265,36 @@ morekeywords={[3]
       j_0,
       keff,
       keff_0,
+      ksp_atol,
+      ksp_atol_0,
+      ksp_divtol,
+      ksp_divtol_0,
+      ksp_max_it,
+      ksp_max_it_0,
+      ksp_rtol,
+      ksp_rtol_0,
       max_dt,
       max_dt_0,
+      memory_available,
+      memory_available_0,
       min_dt,
       min_dt_0,
       mpi_rank,
       mpi_rank_0,
       mpi_size,
       mpi_size_0,
+      mumps_icntl_14,
+      mumps_icntl_14_0,
+      nodes_rough,
+      nodes_rough_0,
       on_gsl_error,
       on_gsl_error_0,
       on_nan,
       on_nan_0,
       on_sundials_error,
       on_sundials_error_0,
+      penalty_weight,
+      penalty_weight_0,
       pi,
       pi_0,
       pid,
@@ -281,6 +307,14 @@ morekeywords={[3]
       report_0,
       sn_alpha,
       sn_alpha_0,
+      snes_atol,
+      snes_atol_0,
+      snes_max_it,
+      snes_max_it_0,
+      snes_rtol,
+      snes_rtol_0,
+      snes_stol,
+      snes_stol_0,
       static_steps,
       static_steps_0,
       step_static,
@@ -289,6 +323,12 @@ morekeywords={[3]
       step_transient_0,
       t,
       t_0,
+      total_dofs,
+      total_dofs_0,
+      ts_atol,
+      ts_atol_0,
+      ts_rtol,
+      ts_rtol_0,
       zero,
       zero_0,
 },
