### 2023-10-23 bca0280
allow _ instead of . for scattering and sech
