### 2023-10-23 e4ee9c5
symlinked COPYING to LICENSE
