### 2023-10-23 #5846 @gtheler
> This article presents an open-source cloud-native computational tool, FeenoX, that provides a mathematical framework to solve partial differential equations (PDEs). The software was launched over a year ago with a decent commit history. The software paper is well written. I suggest the following minor revisions.

Thank you very much for your time @AnjaliSandip .
I am very excited about this review.

>     1. Include a file titled “LICENSE” in the repository (not in the sub-folders) with the license information

The license file is called COPYING, following the GNU recommendations.
I have added a symbolic link named LICENSE to the existing license file in commit e4ee9c5f24fb57e4477960ac7829c4b1a5b23586

>     2. Include the installation instructions in the README file

The common installation instructions (i.e. cloning the git repo and compiling) are already in the README at
https://github.com/seamplex/feenox/blob/main/README#L568
There is also a link to the more detailed installation instructions at https://github.com/seamplex/feenox/blob/main/doc/compilation.md
Can you please be more specific about what you would like to see in the README?

>     3. In the README file, combine 1, 2 and 3.  Include installation instructions/quickstart along with download instructions.  Licensing information should be removed from the README file.

What do you mean by "combine 1, 2 and 3"? Do you want me to remove the headers?
The brief download section is already in the README: https://github.com/seamplex/feenox/tree/main#download with a link to more detailed instructions at https://seamplex.com/feenox/download.html

I really would like to keep the licensing information in the README. It is not a repeated paragraph from the license but my own summary of what free software means.

>     4. Include clear guidelines for third parties wishing to contribute to the software.

There are some guidelines here https://seamplex.com/feenox/doc/programming.html compiled from the markdown sources at https://github.com/seamplex/feenox/blob/main/doc/programming.md
Would you like more particular instructions?

>     5. The summary in the software paper could be revised for a diverse, non-specialist audience.

Fair enough. Any suggestion about which parts should be improved?

I am looking forward to getting this work published.
Again, thank you very much.
>
>
> Once the above have been addressed, I support publication.
