### 2023-10-23 #5846 @AnjaliSandip
@gtheler

1. Looks good.
2. This is satisfactory.
3. I suggest the README file be organized in the following manner - (a) About the software, (b) quickstart (or how to use it), © an example with results, (d)licensing, and  (e)clear guidelines for third parties wishing to  Contribute to the software, Report issues or problems with the software and Seek support. You have much of the above, except for how the information is organized.
4. I suggest you explicitly include how third parties can contribute under “Further information”
5. Terms such as ODEs/DAEs without the full form provided, or core-level steady-state neutronic, finite element? Etc.  I will suggest you have a diverse, non-specialist audience read this and make changes based on their feedback.
