### 2023-10-24 3b9f4ad
update readme as per recommendations
