+++ README.md
@@ -9,58 +9,9 @@ number-sections: true
 toc: true
 ...
 
+# About FeenoX
 
-
-
-::: {.only-in-format .html .markdown}
-> **News**
->
-> * [Debian/Ubuntu `.deb` packages available](download.html#debian) for FeenoX version\ 0.3.
->
-> * A cloud-first approach for solving core-level neutron transport. 8a. Reunión Anual GArCyAR 2022, December 2022.
->   - [Recordings](https://www.youtube.com/watch?v=rdZtyMf5m0k)
->   - [Slides in PDF](8a. Reunión Anual GArCyAR 2022)
->   - [Markdown examples sources](https://github.com/gtheler/2022-garcar)
->
->
-> * FeenoX, a cloud-first free and open source finite-element(ish) tool. Science Circle, on-line course. November 2022.
->   - [Recordings](https://youtu.be/EZSjFzJAhbw)
->   - [Markdown examples sources](https://github.com/gtheler/2022-feenox-sci-circle)
->
->   ::::: {.container .text-center .my-5 .ratio .ratio-16x9}
->   <iframe class="embed-responsive-item" src="https://www.youtube.com/embed/EZSjFzJAhbw?rel=0" allowfullscreen></iframe>
->   :::::
->
-> * FeenoX hands-on tutorial. Science Circle, on-line course. October 2022 <https://www.youtube.com/watch?v=b3K2QewI8jE>
->
->   ::::: {.container .text-center .my-5 .ratio .ratio-16x9}
->   <iframe class="embed-responsive-item" src="https://www.youtube.com/embed/b3K2QewI8jE?rel=0" allowfullscreen></iframe>
->   :::::
-> 
-> * Why FeenoX works the way it works (i.e. does not run in Windows):
->
->   ::::: {.container .text-center .my-5 .ratio .ratio-16x9}
->   <iframe class="embed-responsive-item" src="https://www.youtube.com/embed/66WvYTb4pSg?rel=0" allowfullscreen></iframe>
->   :::::
->
-> * [FeenoX tutorials](doc/tutorials):
->   - [Setting up your workspace](doc/tutorials/000-setup)
->   - [Tensile test](doc/tutorials/110-tensile-test)
->   - [Solving mazes with PDES instead of AI](doc/tutorials/120-mazes)
->
-> * FeenoX for nuclear-related problems Presentation, December 2021
->   - [Recording with CCs](https://youtu.be/e8kFmFOsbPk)
->   - [Slides in PDF](https://www.seamplex.com/feenox/doc/2021-brasil.pdf)
->   - [Markdown examples sources](https://github.com/gtheler/2021-brasil)
->
-> * FeenoX Overview Presentation, August 2021
->   - [Recording (audio in Spanish, slides in English)](https://youtu.be/-RJ5qn7E9uE)
->   - [Slides in PDF](https://www.seamplex.com/feenox/doc/2021-feenox.pdf)
->   - [Markdown examples sources](https://github.com/gtheler/2021-presentation)
-:::
-
-
-# Why FeenoX?
+## Why FeenoX?
 
 > If by "Why FeenoX?" you mean "Why is FeenoX named that way?," read the [FAQs](doc/FAQ.md).
 
@@ -73,7 +24,7 @@ Here is ***why FeenoX is different**.
 doc/why.md
 ```
 
-# How is FeenoX different?
+## How is FeenoX different?
 
 FeenoX meets fictitious-yet-plausible [Software Requirement Specifications](https://www.seamplex.com/feenox/doc/srs.html).
 The FeenoX [Software Design Specifications](https://www.seamplex.com/feenox/doc/sds.html) address each requirement of the SRS.
[…]
