+++ doc/downloads.md
@@ -9,7 +9,8 @@ FeenoX is distributed under the terms of the [GNU General Public License version
 |  Github repository                     | <https://github.com/seamplex/feenox/>                 |
 
 
- * Be aware that FeenoX is a backend. It **does not have a GUI**. Read the [documentation](https://seamplex.com/feenox/doc/), especially the [description](https://www.seamplex.com/feenox/doc/feenox-desc.html) and the [FAQs](https://seamplex.com/feenox/doc/FAQ.html). Ask for help on the [GitHub discussions page](https://github.com/seamplex/feenox/discussions).
+ * FeenoX is **cloud first**. It was designed to run on servers.
+ * Be aware that FeenoX **does not have a GUI**. Read the [documentation](https://seamplex.com/feenox/doc/), especially the [description](https://www.seamplex.com/feenox/doc/feenox-desc.html) and the [FAQs](https://seamplex.com/feenox/doc/FAQ.html). Ask for help on the [GitHub discussions page](https://github.com/seamplex/feenox/discussions).
  * Debian/Ubuntu packages are unofficial, i.e. they are not available in `apt` repositories. They contain dynamically-linked binaries and their dependencies are hard-coded for each Debian/Ubuntu release. Make sure you get the right `.deb` for your release (i.e. `bookworm`/`bullseye` for Debian, `kinetic`/`focal` for Ubuntu).
  * Generic GNU/Linux binaries are provided as statically-linked executables for convenience. They do not support MUMPS nor MPI and have only basic optimization flags. Please compile from source for high-end applications. See [detailed compilatation instructions](https://seamplex.com/feenox/doc/compilation.html).
  * Try to avoid Windows as much as you can. The binaries are provided as transitional packages for people that for some reason still use such an outdated, anachronous, awful and invasive operating system. They are compiled with [Cygwin](http://cygwin.com/) and have no support whatsoever. Really, really, **get rid of Windows ASAP**.
