### 2023-10-24 #5846 @gtheler
Dear @AnjaliSandip , I have updated the README as per your recommendations in this commit: https://github.com/seamplex/feenox/commit/3b9f4ad7614e69c93600bbd27fd1c35d4e4d90d8

About point #5, let me see how I can improve the summary in the paper.
Thanks.
