+++ doc/unix.md
@@ -1,5 +1,5 @@
 
-# Rule of Modularity
+# Rule of Modularity {#sec:modularity}
 
 > Developers should build a program out of simple parts connected by well defined interfaces, so problems are local, and parts of the program can be replaced in future versions to support new features. This rule aims to save time on debugging code that is complex, long, and unreadable.
 
@@ -10,14 +10,14 @@
    - SLEPc
 
 
-# Rule of Clarity
+# Rule of Clarity {#sec:clarity}
 
 > Developers should write programs as if the most important communication is to the developer who will read and maintain the program, rather than the computer. This rule aims to make code as readable and comprehensible as possible for whoever works on the code in the future.
 
  * Example two squares in thermal contact.
  * LE10 & LE11: a one-to-one correspondence between the problem text and the FeenoX input.
 
-# Rule of Composition
+# Rule of Composition {#sec:composition}
 
 > Developers should write programs that can communicate easily with other programs. This rule aims to allow developers to break down projects into small, simple programs rather than overly complex monolithic programs.
 
@@ -26,14 +26,14 @@
  * ASCII output is 100% controlled by the user so it can be tailored to suit any other programs’ input needs such as AWK filters to create LaTeX tables.
 
 
-# Rule of Separation
+# Rule of Separation {#sec:separation}
 
 > Developers should separate the mechanisms of the programs from the policies of the programs; one method is to divide a program into a front-end interface and a back-end engine with which that interface communicates. This rule aims to prevent bug introduction by allowing policies to be changed with minimum likelihood of destabilizing operational mechanisms.
 
  * FeenoX does not include a GUI, but it is GUI-friendly.
 
 
-# Rule of Simplicity
+# Rule of Simplicity {#sec:simplicity}
 
 > Developers should design for simplicity by looking for ways to break up program systems into small, straightforward cooperating pieces. This rule aims to discourage developers’ affection for writing “intricate and beautiful complexities” that are in reality bug prone programs.
 
@@ -43,14 +43,14 @@
  * If there is a single material there is no need to link volumes to properties.
 
 
-# Rule of Parsimony
+# Rule of Parsimony {#sec:parsimony}
 
 > Developers should avoid writing big programs. This rule aims to prevent overinvestment of development time in failed or suboptimal approaches caused by the owners of the program’s reluctance to throw away visibly large pieces of work. Smaller programs are not only easier to write, optimize, and maintain; they are easier to delete when deprecated.
 
  * Parametric and/or optimization runs have to be driven from an outer script (Bash, Python, etc.)
  
 
-# Rule of Transparency
+# Rule of Transparency {#sec:transparency}
 
 > Developers should design for visibility and discoverability by writing in a way that their thought process can lucidly be seen by future developers working on the project and using input and output formats that make it easy to identify valid input and correct output. This rule aims to reduce debugging time and extend the lifespan of programs.
 
@@ -58,17 +58,17 @@
  * Makes use of structures and function pointers to give the same functionality as C++’s virtual methods without needing to introduce other complexities that make the code base harder to maintain and to debug.
 
 
-# Rule of Robustness
+# Rule of Robustness {#sec:robustness}
 
 > Developers should design robust programs by designing for transparency and discoverability, because code that is easy to understand is easier to stress test for unexpected conditions that may not be foreseeable in complex programs. This rule aims to help developers build robust, reliable products.
 
 
-# Rule of Representation
+# Rule of Representation {#sec:representation}
 
[…]
