### 2023-10-26 07ffa9e
improved summary of the joss paper
