### 2023-10-26 #5846 @gtheler
Dear @AnjaliSandip I have updated the summary of the paper in this commit: https://github.com/seamplex/feenox/commit/07ffa9ef3e25f7381539a21d003a2dc7d1a42f8b
