### 2023-10-28 #5846 @vijaysm
@gtheler I just finished reading the submitted paper and have couple of suggestions below.

1. "access element shape functions and their derivatives": While I understand what you are trying to convey, it may be better to have a small snippet or example of how the local element assembly or point-wise Gauss assembly is invoked. This might be the kernel of the assembly routine and an example for say homogeneous laplacian here may suffice. This is important since FeenoX uses FE. Additionally, some description of what types of FE basis and orders are supported may be helpful as well.
2. In terms of references, it might be useful to compare the original Fenics/Dolfin suite (https://fenicsproject.org/) that provides somewhat similar functionality to describe the problem as a weak form statement using [UFL](https://fenics.readthedocs.io/projects/ufl/en/latest/#:~:text=The%20Unified%20Form%20Language%20(UFL,notation%20close%20to%20mathematical%20notation.). A detailed comparison may not be needed - but contrasting the approach used in UFL vs FeenoX Software Design Specification may be interesting.
