+++ src/mesh/mesh.c
@@ -229,7 +229,7 @@ int feenox_instruction_mesh_read(void *arg) {
         for (size_t i = 0; i < physical_group->n_elements; i++) {
           element_t *element = &this->element[physical_group->element[i]];
           for (unsigned int q = 0; q < element->type->gauss[this->integration].Q; q++) {
-            double wdet = feenox_fem_compute_w_det_at_gauss(element, q, this->integration);
+            double wdet = feenox_fem_compute_w_det_at_gauss_integration(element, q, this->integration);
             gsl_matrix *H_c = feenox_fem_compute_H_c_at_gauss(element, q, this->integration);
 
             for (unsigned int j = 0; j < element->type->nodes; j++) {

[…]
