### 2023-10-31 c1fe176
suggestions from vijaysm for the joss paper draft
