### 2023-10-31 #5846 @gtheler
Thanks for your suggestions @vijaysm .
I have implemented them in https://github.com/seamplex/feenox/commit/c1fe176c2ad6ee06a02ab6edcbcb50bd955f59b9
