+++ src/pdes/thermal/init.c
@@ -70,9 +70,9 @@ int feenox_problem_parse_time_init_thermal(void) {
 ///pr_thermal+T_0+description The initial condition for the temperature in transient problems.
 ///pr_thermal+T_0+description If not given, a steady-steady computation at $t=0$ is performed.  
 
-///pr_thermal+T_0+description The initial guess for the temperature in steady-state problems.
-///pr_thermal+T_0+description If not given, a uniform distribution equal to the the average
-///pr_thermal+T_0+description of all the temperature appearing in boundary conditions is used.
+///pr_thermal+T_guess+description The initial guess for the temperature in steady-state problems.
+///pr_thermal+T_guess+description If not given, a uniform distribution equal to the the average
+///pr_thermal+T_guess+description of all the temperature appearing in boundary conditions is used.
   
 ///va_thermal+T_max+detail The maximum temperature\ $T_\text{max}$.
   feenox_check_null(thermal.T_max = feenox_define_variable_get_ptr("T_max"));
