### 2023-11-04 #5846 @chennachaos
Hi @Kevin-Mattheus-Moerman,
After carefully going through the paper, repository, documentation and examples available, I believe that the library is not substantial in its extent, capabilities and usefulness.

Because of the way it is organised and too much unnecessary text in many places, it was a quite difficult library to review. I am not entirely sure what the author is trying to convey with the comparisons with Markdown right at the beginning of the description. I acknowledge that the author might have their own philosophy about how FeenoX is different. The use of scripts as inputs is not new.

While the idea of offering *a general mathematical framework to solve PDEs with a bunch of entry points (as C functions) where new types of PDEs (e.g. electromagnetism, fluid mechanics, etc.) can be added to the set of problems FeenoX can solve* is good, development and testing of such solvers requires a substantial effort from the users/developers. Solving PDEs by writing weak forms in FEniCS and FreeFEM is totally different and much easier than developing the stiffness matrices from scratch as needed in this library. While the library contains solver capabilities only for a few basic problems (mostly steady-state/static, transient heat transfer but mostly linear problems) and offers FEM ingredients for further developments, the development of solvers for other PDEs, particularly for nonlinear problems, is not as simple as indicated in the paper. So, in the context of developing new solvers, the library is not as flexible as claimed. Scalability claims are also not justified.

I am not convinced about the statement of need and the usefulness of this library when compared with similar libraries like FEniCS and FreeFEM. I do not recommend the library in its current form for publication in JOSS. I think the library has potential but requires additional developments in solver capabilities and higher-level functionality that can lessen the burden of solver development. The documentation requires some clarity and cleaning; there is too much text. Moreover, there should be examples explaining the development of new solvers rather than input scripts of test cases for existing solvers.
