### 2023-11-05 #5846 @gtheler
Dear @chennachaos

Here is my elaborated answer to your kind review.
If you all allow me, I will comment your review and then summarize all the ideas at the end.

> After carefully going through the paper, repository, documentation and examples available, I believe that the library is not substantial in its extent, capabilities and usefulness.

Thanks for spending weekend time going through the FeenoX website.
I know you already made a choice about publication, I respect it and it is fine.
I just want to get help to convey the ideas behind FennoX that I think make them unique, at least in the free and open-source world as there's no other tool that fulfills the SDS that I know of.
Maybe the failure in communication comes from the fact that I am not a native English speaker (I am from Argentina) because when I talk (in Spanish) with other colleagues (from the nuclear academy and industry) they do understand me and get my points.
So please, I do humbly ask all four of you @Kevin-Mattheus-Moerman  @AnjaliSandip @vijaysm @chennachaos   to tell me where the communication fails.

In particular, about the extent, capabilities and usefulness mentioned.

 * extent: just the fact that it provides a neutron transport solver using the SN method over unstructured grids and with the possibility of using MPI out of the box is a fact that many people in the nuclear industry consider its extent remarkable.

 * capabilities: the non-conformal mesh capabilities studied in https://github.com/gtheler/feenox-non-conformal-mesh-interpolation are similar in terms of results and better in terms of performance than a "commercial" (in the sense of something that people pay to use and that has paid technical support) solution provided by a FEM software company that invoiced 2 BUSD in 2002 (as confirmed by an employee in charge of the non-conformal mesh solution in the company).

 * usefulness: let me quote an email I got from and advanced user

  > FeenoX has releases with a proper tarball! It has `INSTALL`, `./configure` and
  > just compiles. Wow. Yeah, these are free software basics, but majority of
  > the sim software (some discrete circuit simulators included!) I've tried
  > fail on most of these points. So thanks for making quality software!
  >
  > _Tibor 'Igor2' Palinkas, maintainer of the open source PCB editor pcb-rnd_

  and another one-liner when a company in the US asked for orthotropic thermal expansion coefficients (which were added by me for a fee) when I mentioned that the strains were available as functions of space named `dudx`, `dvdx`, ..., `dwdz`, which could be evaluated at any arbirary point in space like `PRINT dvdy(1.23456,2*pi,1/3)`:

  > Well that makes my life super easy.

> Because of the way it is organised and too much unnecessary text in many places, it was a quite difficult library to review.

I agree with the objection about too much text, but it should boil down to the SRS & SDS.
However, here is a first point that triggers a cascade of further misunderstandings.
FeenoX **is not a library**. The SDS explicitly asks that users need no compile anything to use the tool.
FeenoX is an executable that reads one or more input files. It is designed primarily as a end-user tool, not as a developer tool.
Its main focus is industry while at the same time providing a way for academics, researchers, developers (which might be a small fraction of the main users) to do interesting stuff.
I will come back again to the users and developer distinction below.

> I am not entirely sure what the author is trying to convey with the comparisons with Markdown right at the beginning of the description. I acknowledge that the author might have their own philosophy about how FeenoX is different. The use of scripts as inputs is not new.

Let me work backwards. Not only I do know that "the use of scripts as inputs is not new" but I encourage its usage as opposed to the "everything needs a GUI" idea that clashes with the Unix philosophy, which is the philosophy under which FeenoX was developed and it is the one that is "different" with respect to most of the other computational tools out there. Unix philosophy asks explicitly for ASCII formats which can be tracked with Git. They also ask for the "rule of silence" and the "rule of economy" which I'll briefly explain below.
So now to the comparison between Word, TeX and markdown and traditional FEM solvers, FEM libraries and FeenoX.
The analogy only works if somebody used both Word, Markdown and TeX, but I hope we are on the same s about the combination of easy of use and powerfulness between Word and TeX and the fact that Markdown lands in the middle and wins the battle.

In this comparison, traditional industrial point-and-click FEM software would be like Word. They are easy but not so flexible. In the "free" world I would mention CalculiX and CodeAster.
Advanced libraries (i.e. routines that need to be compiled, used and linked to solve problems) would be like TeX. They are very flexible but extremely hard to use. I a thinking in Sparselizard or MoFEM.
The idea is tha FeenoX would sit in the middle. As requested in the SDS, regular users would not need to compile anything. And a simple problem (like the NAFEMS LE10 case shown in the paper) would need a very simple input file.
More complex cases (like the one shown in the examples page of the website) would need more complex input files.

Both CalculiX and CodeAster (and Elmer etc) read input files.
But these input files are way too complicated (and make no sense to me) even for very simple cases.
In https://seamplex.com/feenox/tests/nafems/le10/ I try to explain how I managed to create input files to solve the very simple NAFEMS-LE10 case with these two standard tools. It was an un-needed nightmare, and the resulting input files are not Git-friendly. The input files reported here directly scare me: https://cofea.readthedocs.io/en/latest/benchmarks/001-thick-plate/tested-codes.html
Also, the LE10 test page is outdated but I can show that FeenoX is way faster than the other tools out there for this case.

> While the idea of offering a general mathematical framework to solve PDEs with a bunch of entry points (as C functions) where new types of PDEs (e.g. electromagnetism, fluid mechanics, etc.) can be added to the set of problems FeenoX can solve is good, development and testing of such solvers requires a substantial effort from the users/developers.

I could not agree more. I cannot help feeling my communication skills are way too bad here.
First of all, once again, users and developers are not the same group of people.
FeenoX should be "easy" for "users" solving "simple" problems. More complex problems do not have to be necesarily easy (but easier than other tools).
Second, adding support for a new PDE _will_ need substantial effort from the developer.
I don't think how it cannot need substantial effort. I just wanted to convey that this effort is smaller that creating a whole solver from scratch.
Please, even if the paper is not published, help me out to communicate this concept in a more efficient way.

> Solving PDEs by writing weak forms in FEniCS and FreeFEM is totally different and much easier than developing the stiffness matrices from scratch as needed in this library.

I don't agree. Again, differentiation between users and developers is central here.
Solving PDEs with FEnniCS and FreeFEM is way more complex than with FeenoX.
Once a PDe has been added to Feenox by developers, users don't need to write the matrices. They have to build a proper input file with `PROBLEM thermal`. I can tell you that a significant fraction of "users" do not know what a boundary condition is, let a alone a weak form.

> While the library contains solver capabilities only for a few basic problems (mostly steady-state/static, transient heat transfer but mostly linear problems) and offers FEM ingredients for further developments, the development of solvers for other PDEs, particularly for nonlinear problems, is not as simple as indicated in the paper.

Again, I did not want to say "simple" but "simpler than..."
Regarding the "few basic problems". The mechanical solver can be run in a quasi-static fashion, for instance reading a temperature distribution computed in a different mesh but same geometry solving a non-linear transient heat conduction problem. Use case: computation of ASME linearized stress history needed to evaluate environmentally-assisted fatigue due to thermal transients in piping of nuclear power plants at interfaces between carbon and stainless steel classes. One can also add seismic loads too: https://github.com/seamplex/piping-asme-fatigue

The tutorials also show a nice application of the ellipticity of the Laplace equation, both steady and transient: https://www.linkedin.com/feed/update/urn:li:activity:7117082910381232128/

To complete, just this last week I was thanked by a colleague that wrote his Master's thesis about thermal optimization of a reflector in an integrated reactor. He just had to focus on the python part of the topological optimization algorithm while FeenoX could take care of solving the non-linear heat conduction equation and write the heat fluxes needed to evalaute the cost function in a way that would have been way more complicated using another FEM solver.

> So, in the context of developing new solvers, the library is not as flexible as claimed.

I respectfully disagree. First, it is not a library. Second, it can solve a wide plethora of problems, some of the illustrated in the examples sections.
Third, have you tried to add neutron transport with the method of discrete ordinates to FEniCSx?

> Scalability claims are also not justified.

It is true that parallelization has a lot of room for improvement.
But the fact that FeenoX can solve neutron transport with large N over unstructured grids by splitting the problem into several MPI ranks in different hosts is already a sound scalability claim.

> I am not convinced about the statement of need and the usefulness of this library when compared with similar libraries like FEniCS and FreeFEM.

Maybe here is the main misunderstanding: FeenoX is not comparable to FEniCS nor FreeFEM.
Not even to Sparselizard. If something, to CalculiX and CodeAster.

> I do not recommend the library in its current form for publication in JOSS.

Fair enough. If this paper is not published I will not re-submit it.

> I think the library has potential but requires additional developments in solver capabilities and higher-level functionality that can lessen the burden of solver development.

I agree. Yet, the main point is not about "solver developers" but about "users".

> The documentation requires some clarity and cleaning; there is too much text.

I agree, but it should boil down to the SRS and the SDS.

> Moreover, there should be examples explaining the development of new solvers rather than input scripts of test cases for existing solvers.

Fair enough. The examples are for mostly for "users". First, let me put some numbers: I expect a "developer" to be 1 per 1,000 to 10,000 "users". So I would rather start with examples for the most common type of user, which is industrial users.
Second, any explicit example about how to add a new PDE will be either simple or particular to that PDE. So an actual example would be exactly adding the PDE myself (which is what I am doing with an academic professor that needs to model electrical currents in biological tissue). The instructions are

 1. take a look at the `laplace` directory for a basic elliptic operator
 2. open a new thread in the Github discussions to see how to proceed

The rule of silence says that if Unix tools should not clutter up output with unnecessary output.
The rule of economy says that engineering time is way more expensive than CPU time.
If one combines these two rules, the only plausible solution for a FEM solver is to have 100%-user-defined output, which is what FeenoX has.
Solvers dating back from 1970s have large output files where every single result is written to.
This makes it far more time consuming for the engineer (user) to filter out the important results than to directly spawn new executions that generate only the desired results. This is something I have not seen in industrial solvers (yes of course in libraries).

If you want, we can prepare a test case like the LE10 problem, solve it with different tools and measure both the CPU and memory needed. I am confident that my FeenoX will be one of the top scorers.

Finally, note that currently I am an employee of said top-notch FEA company. Therefore, FeenoX is a weekend project and the goal of this publication was to give the project more visibility so at the end of the day more people (both users and developers) could join me.
If this paper is not published in its present form with further small comments, I will not re-submit it again as I would be writing very much the same claims.

Once again, thank you very much.
