### 2023-11-14 1670a2a
thermal radiation tests
