+++ doc/syntax-feenox.tex
@@ -383,12 +383,21 @@ morekeywords={[4]
       mpi_memory_local,
       not,
       prod,
+      qrng2d_halton,
+      qrng2d_niederreiter,
+      qrng2d_reversehalton,
+      qrng2d_sobol,
+      qrng_halton,
+      qrng_niederreiter,
+      qrng_reversehalton,
+      qrng_sobol,
       quasi_random,
       random,
       random_gauss,
       root,
       round,
       sawtooth_wave,
+      sech,
       sgn,
       sin,
       sinh,

[…]
