### 2023-11-15 #5846 @vijaysm
I have not yet fully finished my review, but the instructions were easy enough to replicate a clean installation in a docker container. I was able to run most of the examples as well without issues. @gtheler

Does FeenoX provide a groundbreaking new methodology or solver? IMHO, probably not. However, I still find that it offers a flexible infrastructure with lua scripts as input, to compute *linear* solutions of Laplace-like problems to slightly more complicated discrete-ordinates treatment of Boltzmann transport in multiple dimensions. While libraries such as FreeFEM or Fenics do offer better and more efficient functionality to match said capabilities, FeenoX's modules may be useful for quick prototyping or computing sensitivity to parameter variations (once a problem is setup apriori). If documented better, it may also be a simple tool to use for students before writing their own computational solver implementations.

The main deficiency that I do see (and agree with @chennachaos here) is that FeenoX does not seem to showcase any solvers for nonlinear systems. Additionally, with PETSc as a dependency, I do not see any details about the underlying parallel solvers, precondition setups in the documentation, or even example usage of nonlinear schemes for any of the problems (with the `--non-linear` option for FeenoX executable). I had to directly pass in command-line arguments (like `--log_view`,  `--ksp_monitor` etc) to understand the solver setup used in PETSc, and modify options further to change the preconditioners etc. This should be better documented as well, especially if the intended audience are users unfamiliar with PETSc options.

I am also confused by the terminology "cloud-first". I assumed this meant that it could be deployed on cloud computing platforms like AWS. However, this is unclear and misleading.

Overall, I believe that it may be beneficial to accept this submission if the author can address the deficiency in the documentation to clarify better the classes of problems supported, usage options and current limitations, and expected behavior when running examples (including ways to improve the solver). I also find that several of the example pages have dead links to images and other resources, which should be fixed.
