### 2023-11-15 #5846 @gtheler
Thanks for your time @vijaysm .

> I have not yet fully finished my review, but the instructions were easy enough to replicate a clean installation in a docker container. I was able to run most of the examples as well without issues. @gtheler

Good. Honestly I never did thay myself because I don't like docker, but that was exactly the spirit of the code.

> Does FeenoX provide a groundbreaking new methodology or solver? IMHO, probably not.

Agree.

> However, I still find that it offers a flexible infrastructure with lua scripts as input, to compute _linear_ solutions of Laplace-like problems to slightly more complicated discrete-ordinates treatment of Boltzmann transport in multiple dimensions. While libraries such as FreeFEM or Fenics do offer better and more efficient functionality to match said capabilities, FeenoX's modules may be useful for quick prototyping or computing sensitivity to parameter variations (once a problem is setup apriori). If documented better, it may also be a simple tool to use for students before writing their own computational solver implementations.

That's exactly what the Word/Markdown/TeX comparison means with GUI solvers/FeenoX/specific FEM libraries.
Thanks for understanding.

> The main deficiency that I do see (and agree with @chennachaos here) is that FeenoX does not seem to showcase any solvers for nonlinear systems.

Agree. So far the thermal solver is the one that fully handles non-linear solves. Non-linearity might arise from
 1. temperature-dependent BCs
 3. temperature-dependent volumetric power sources
 4. temperature-dependent conductivities
The tests directory covers all three cases, although it might be true that there are no examples of these problems:

 * https://github.com/seamplex/feenox/blob/main/tests/convectionNL.fee
 * https://github.com/seamplex/feenox/blob/main/tests/pellet-nonlinear-guess.fee
 * https://github.com/seamplex/feenox/blob/main/tests/pellet-nonlinear-transient-from-initial.fee
 * https://github.com/seamplex/feenox/blob/main/tests/pellet-nonlinear-transient-from-ss.fee
 * https://github.com/seamplex/feenox/blob/main/tests/radiation-as-convection-celsius.fee
 * https://github.com/seamplex/feenox/blob/main/tests/radiation-as-heatflux-celsius.fee
 * https://github.com/seamplex/feenox/blob/main/tests/radiation-as-heatflux-celsius.fee
 * https://github.com/seamplex/feenox/blob/main/tests/radiation-as-heatflux-kelvin.fee
 * https://github.com/seamplex/feenox/blob/main/tests/thermal-slab-convection-as-heat-nosource.fee
 * https://github.com/seamplex/feenox/blob/main/tests/thermal-slab-temperature-nosource.fee
 * https://github.com/seamplex/feenox/blob/main/tests/thermal-slab-transient-mms-capacity-of-T.fee
 * https://github.com/seamplex/feenox/blob/main/tests/thermal-two-squares-material-explicit-temperature.fee
 * https://github.com/seamplex/feenox/blob/main/tests/thermal-two-squares-material-explicit-temperature.fee

Let's do as follows. What if I wrote a tutorial (like the two that are already available at https://seamplex.com/feenox/doc/tutorials/) for the thermal case, including all the non-linear features? Would that showcase enough non-linear cases?

> Additionally, with PETSc as a dependency, I do not see any details about the underlying parallel solvers, precondition setups in the documentation, or even example usage of nonlinear schemes for any of the problems (with the `--non-linear` option for FeenoX executable).

https://www.seamplex.com/feenox/doc/feenox-manual.html#problem

The `--non-linear` option should not be used. FeenoX should detect if a problem is linear or not by itself.
If you ran any of the cases I listed about, all of them should use SNES so `--snes_view` will show something.
In linear cases, `--snes_view` will be empty.
In those cases where one wants to overwrite what it detected, then use `--linear` or `--non-linear`.

> I had to directly pass in command-line arguments (like `--log_view`, `--ksp_monitor` etc) to understand the solver setup used in PETSc, and modify options further to change the preconditioners etc. This should be better documented as well, especially if the intended audience are users unfamiliar with PETSc options.

One can control the basic options for PETSc with the `PROBLEM` keyword: https://www.seamplex.com/feenox/doc/feenox-manual.html#problem
If one does not want to write PETSc options in the command line every time, one can use the `PETSC_OPTIONS` keyword: https://www.seamplex.com/feenox/doc/feenox-manual.html#petsc_options

The usage of PETSc command-line options is encouraged by both FeenoX and the PETSc team.
I don't see what's wrong with that.

> I am also confused by the terminology "cloud-first". I assumed this meant that it could be deployed on cloud computing platforms like AWS. However, this is unclear and misleading.

Fair enough. It would need a whole document for that. Let me sumarize:

 a. FeenoX is the back-end at www.caeplex.com so it proves it can be deployed to AWS by a single-person company (i.e. myself)
 b. the SRS and the SDS do talk a little bit about the idea of "cloud first." It does not mean just that it can run over ssh. It involves CI, deployment, interaction with other tools, I/O, means of reporting status remotely, etc. I can further explain because I now work at a FEM company that acquired a startup company that developed a cloud-based FEM system from scratch (that startup hired me because of what I had done in CAEplex, so I know what I'm talking about).

> Overall, I believe that it may be beneficial to accept this submission if the author can address the deficiency in the documentation to clarify better the classes of problems supported, usage options and current limitations, and expected behavior when running examples (including ways to improve the solver).

Agree.
To do what you propose in a general way is a many man-months job, and my original idea was to publish FeenoX in JOSS as it is so as to give publicity and get more people onboard. The guidelines mentioned that the software did not have to be fully finished to be accepted for publication.

But if you can summarize these deficiencies in a few bullets I can add to the documentation in a few weeks, that would be a start.

> I also find that several of the example pages have dead links to images and other resources, which should be fixed.

Can you please tell me where so I can correct them?
