### 2023-11-21 #5846 @Kevin-Mattheus-Moerman
@gtheler did you implement the changes you hinted at to @vijaysm e.g. the thermal tutorials featuring non-linear applications? Are we waiting for you to implement some requesting changes?
