### 2023-11-21 #5846 @gtheler
No, I am not @Kevin-Mattheus-Moerman . Not sure what the next steps and what the choices are.
I really would like the paper to be published, so I am eager for suggestions.

If someone thinks writing a non-linear thermal tutorial will help me achieve my goal, then I will hapilly do it (not before the weekend, as FeenoX is a free-time project).
