### 2023-11-21 #5846 @vijaysm
> That's exactly what the Word/Markdown/TeX comparison means with GUI solvers/FeenoX/specific FEM libraries.

You should rephrase this in the paper. The comparison of a requirement specification to markdown or TeX is irrelevant. It creates more confusion rather than providing motivations for the FeenoX design.

> Agree. So far the thermal solver is the one that fully handles non-linear solves. Non-linearity might arise from

Fair enough. However, I do not see any examples or mention in the documentation related to how one might go forward with setting up a nonlinear problem. I was able to run the cases you mentioned in the tests folder and verify that SNES solvers are being used.

> The tests directory covers all three cases, although it might be true that there are no examples of these problems:

Create explicit examples that showcase the nonlinear problem solver that addresses all three conditions to describe the features.

> Let's do as follows. What if I wrote a tutorial (like the two that are already available at https://seamplex.com/feenox/doc/tutorials/) for the thermal case, including all the non-linear features? Would that showcase enough non-linear cases?

Yes.

I also want to point out that under the [examples link](https://www.seamplex.com/feenox/doc/feenox-manual.html#examples), **several** hyperlinks are dead. For example, https://www.seamplex.com/feenox/doc/laplace.html#how-to-solve-a-maze-without-ai, https://www.seamplex.com/feenox/doc/thermal.html#thermal-slabs, https://www.seamplex.com/feenox/doc/mechanical.html#parallelepiped-whose-youngs-modulus-is-a-function-of-the-temperature to name a few.

> The --non-linear option should not be used. FeenoX should detect if a problem is linear or not by itself.

The documentation mentions that one can provide `--linear` or `--non-linear` argument to force FeenoX to use a nonlinear or just a linear solver. But in the examples that I tried, forcing a linear problem to be nonlinear did not work as expected. So it is not clear how these options are implemented if they do not exactly force a behavior.

> One can control the basic options for PETSc with the PROBLEM keyword: https://www.seamplex.com/feenox/doc/feenox-manual.html#problem

This is useful. Please make this link more prominent as I did not notice this before. If you are also using a "prefix" for the solver options, then it will be helpful to mention this in the documentation as well.

> The usage of PETSc command-line options is encouraged by both FeenoX and the PETSc team.
> I don't see what's wrong with that.

There is nothing wrong with that, and that is the advantage of PETSc-based codes. However, without clearly specifying how a user needs to provide these options, the flexibility that is available from command-line options will remain under or unutilized.

> a. FeenoX is the back-end at www.caeplex.com so it proves it can be deployed to AWS by a single-person company (i.e. myself)

I did not have the time to experiment with this, but the capability is interesting. This can be highlighted more prominently in the repository README.md to emphasize how FeenoX can use a cloud-based deployment without installation or computational requirement hurdles for users.

> b. the SRS and the SDS do talk a little bit about the idea of "cloud first." It does not mean that it can run over ssh. It involves CI, deployment, interaction with other tools, I/O, means of reporting status remotely, etc. I can further explain because I now work at a FEM company that acquired a startup company that developed a cloud-based FEM system from scratch (that startup hired me because of what I had done in CAEplex, so I know what I'm talking about).

This particular aspect has no direct implication of being "cloud"-friendly. Almost every good library or application that can be deployed on Docker supports good CI, deployment, status reporting, etc. I suggest focusing on CAEPlex to emphasize cloud computing.

I am on travel the rest of this week and may not be able to devote more time. But hopefully, this gives a sense of what I am looking for to complete the review. Please let me know if you have more questions.
