### 2023-11-21 #5846 @vijaysm
From README.md
> Since it is free ([as in freedom](https://www.gnu.org/philosophy/free-sw.en.html)) and open source, contributions to add features (and to fix bugs) are welcome. In particular, each kind of problem supported by FeenoX (thermal, mechanical, modal, etc.) has a subdirectory of source files which can be used as a template to add new problems, as implied in the “community-contributed problems” bullet above (rules of modularity and extensibility). See the [documentation](https://github.com/seamplex/feenox/blob/main/doc) for details about how to contribute.

I suggest fixing the "documentation" link to point to a dedicated page about guidelines to contributing users. Do you support pull requests? How are reviews done for new changes before acceptance? Github workflows are not used for CI and so how do you decide if new changesets are valid? Some details regarding these aspects will be helpful if you intend to receive contributions from new users.
