+++ doc/tutorials/general.md
@@ -1,0 +1,2 @@
+ 1. [Overview: the tensile test case](110-tensile-test)
+ 2. [Fun & games: solving mazes with PDES instead of AI](120-mazes)
