+++ examples/examples.md
@@ -1,73 +1,73 @@
 - Basic mathematics
-  - [Hello World (and Universe)!](basic.md#hello-world-and-universe)
-  - [Ten ways of computing *π*](basic.md#ten-ways-of-computing-pi)
+  - [Hello World (and Universe)!](https://www.seamplex.com/feenox/examples/basic.html#hello-world-and-universe)
+  - [Ten ways of computing *π*](https://www.seamplex.com/feenox/examples/basic.html#ten-ways-of-computing-pi)
   - [Financial decisions under
-  inflation](basic.md#financial-decisions-under-inflation)
-  - [The logistic map](basic.md#the-logistic-map)
-  - [The Fibonacci sequence](basic.md#the-fibonacci-sequence)
+  inflation](https://www.seamplex.com/feenox/examples/basic.html#financial-decisions-under-inflation)
+  - [The logistic map](https://www.seamplex.com/feenox/examples/basic.html#the-logistic-map)
+  - [The Fibonacci sequence](https://www.seamplex.com/feenox/examples/basic.html#the-fibonacci-sequence)
     - [Using the closed-form formula as a
-    function](basic.md#using-the-closed-form-formula-as-a-function)
-    - [Using a vector](basic.md#using-a-vector)
-    - [Solving an iterative problem](basic.md#solving-an-iterative-problem)
+    function](https://www.seamplex.com/feenox/examples/basic.html#using-the-closed-form-formula-as-a-function)
+    - [Using a vector](https://www.seamplex.com/feenox/examples/basic.html#using-a-vector)
+    - [Solving an iterative problem](https://www.seamplex.com/feenox/examples/basic.html#solving-an-iterative-problem)
   - [Computing the derivative of a function as a UNIX
-  filter](basic.md#computing-the-derivative-of-a-function-as-a-unix-filter)
+  filter](https://www.seamplex.com/feenox/examples/basic.html#computing-the-derivative-of-a-function-as-a-unix-filter)
   - [On the evaluation of thermal expansion
-  coefficients](basic.md#on-the-evaluation-of-thermal-expansion-coefficients)
+  coefficients](https://www.seamplex.com/feenox/examples/basic.html#on-the-evaluation-of-thermal-expansion-coefficients)
 - Ordinary Differential Equations & Differential-Algebraic Equations
   - [Lorenz’ attractor—the one with the
-  butterfly](daes.md#lorenz-attractorthe-one-with-the-butterfly)
-  - [The double pendulum](daes.md#the-double-pendulum)
-  - [Vertical boiling channel](daes.md#vertical-boiling-channel)
+  butterfly](https://www.seamplex.com/feenox/examples/daes.html#lorenz-attractorthe-one-with-the-butterfly)
+  - [The double pendulum](https://www.seamplex.com/feenox/examples/daes.html#the-double-pendulum)
+  - [Vertical boiling channel](https://www.seamplex.com/feenox/examples/daes.html#vertical-boiling-channel)
     - [Original Clausse-Lahey formulation with uniform power
-    distribution](daes.md#original-clausse-lahey-formulation-with-uniform-power-distribution)
-    - [Arbitrary power distribution](daes.md#arbitrary-power-distribution)
+    distribution](https://www.seamplex.com/feenox/examples/daes.html#original-clausse-lahey-formulation-with-uniform-power-distribution)
+    - [Arbitrary power distribution](https://www.seamplex.com/feenox/examples/daes.html#arbitrary-power-distribution)
 - Laplace’s equation
[…]
