+++ doc/feenox-manual.md
@@ -50,9 +50,14 @@ compilation.md
 ../examples/examples.md
 ```
 
-# Tutorial
+# Tutorials
 
-To be done.
+See <https://www.seamplex.com/feenox/tutorials>
+
+```{.include shift-heading-level-by=1}
+tutorials/intro.md
+```
+
 
 # Description
 

[…]
