### 2023-11-24 ea8516c
point kinetics examples
