+++ examples/neutron_diffusion.md
@@ -240,8 +240,8 @@ One eight of two bare reactors
 :::
 
 But what happens "in the middle"? That is to say, how does
-$k_\text{eff}$ changes when we morph the cube into a sphere? Enter Gmsh
-& Feenox.
+$k_\text{eff}$ change when we morph the cube into a sphere? Enter Gmsh &
+Feenox.
 
 ::: {#fig:cube-morph-sphere}
 ![75% cube/25% sphere](cubesphere-25.png){width="33%"}
@@ -408,13 +408,13 @@ $$
 We can then compare the numerical $k_\text{eff}$ computed using...
 
 i.  a non-uniform grid with $n+1$ nodes such that there is a node
-    exactly at $x=b$.
+    exactly at the interface $x=a$.
 ii. an uniform grid (mimicking a neutronic code that cannot handle case
     i.) with $n$ uniformly-spaced elements. The element that contains
     point $x=b$ is assigned to a pseudo material $AB$ that linearly
     interpolates the macroscopic cross sections according to where in
     the element the point $x=b$ lies. That is to say, if the element
-    width is 10 and $b=52$ then this $AB$ material will be 20% of
+    width is 10 and $a=52$ then this $AB$ material will be 20% of
     material $A$ and 80% of material $B$.
 
 > The objective of this example is to show that case i. will
