+++ examples/neutron_sn.md
@@ -101,12 +101,12 @@ INTEGRATE phi1 OVER llq RESULT lower_left_quadrant
 INTEGRATE phi1 OVER lrq RESULT lower_right_quadrant
 INTEGRATE phi1 OVER urq RESULT upper_right_quadrant
 
-PRINTF "LLQ = %.3e (ref 1.676e+0)\n" lower_left_quadrant/(5*5)
-PRINTF "LRQ = %.3e (ref 4.159e-2)\n" lower_right_quadrant/(5*5)
-PRINTF "URQ = %.3e (ref 1.992e-3)\n" upper_right_quadrant/(5*5)
+PRINTF "LLQ = %.3e (ref 1.676e+0)" lower_left_quadrant/(5*5)
+PRINTF "LRQ = %.3e (ref 4.159e-2)" lower_right_quadrant/(5*5)
+PRINTF "URQ = %.3e (ref 1.992e-3)" upper_right_quadrant/(5*5)
 
 WRITE_RESULTS
-PRINTF "%g unknowns for S${1}, memory needed = %.1f Gb\n" total_dofs memory()
+PRINTF "%g unknowns for S${1}, memory needed = %.1f Gb" total_dofs memory()
 
 ```
 
@@ -183,7 +183,7 @@ profile9(y) = phi1(9.84375,y)
 PRINT_FUNCTION profile5 profile7 profile9 MIN 0 MAX 10 NSTEPS 100 FILE $0-$1.dat
 
 WRITE_RESULTS
-PRINTF "%g unknowns for S${1}, memory needed = %.1f Gb\n" total_dofs memory()
+PRINTF "%g unknowns for S${1}, memory needed = %.1f Gb" total_dofs memory()
 
 ```
 
@@ -218,35 +218,35 @@ $
 ```
 
 
-![$\psi_{1,1}$ for S$_4$](azmy-4-01.png)
+![$\psi_{1}$ for S$_4$](azmy-4-01.png)
 
-![$\psi_{2,1}$ for S$_4$](azmy-4-02.png)
+![$\psi_{2}$ for S$_4$](azmy-4-02.png)
 
-![$\psi_{3,1}$ for S$_4$](azmy-4-03.png)
+![$\psi_{3}$ for S$_4$](azmy-4-03.png)
 
-![$\psi_{4,1}$ for S$_4$](azmy-4-04.png)
+![$\psi_{4}$ for S$_4$](azmy-4-04.png)
 
-![$\psi_{5,1}$ for S$_4$](azmy-4-05.png)
+![$\psi_{5}$ for S$_4$](azmy-4-05.png)
 
-![$\psi_{6,1}$ for S$_4$](azmy-4-06.png)
+![$\psi_{6}$ for S$_4$](azmy-4-06.png)
 
-![$\psi_{7,1}$ for S$_4$](azmy-4-07.png)
+![$\psi_{7}$ for S$_4$](azmy-4-07.png)
 
-![$\psi_{8,1}$ for S$_4$](azmy-4-08.png)
+![$\psi_{8}$ for S$_4$](azmy-4-08.png)
 
-![$\psi_{9,1}$ for S$_4$](azmy-4-09.png)
+![$\psi_{9}$ for S$_4$](azmy-4-09.png)
 
-![$\psi_{10,1}$ for S$_4$](azmy-4-10.png)
+![$\psi_{10}$ for S$_4$](azmy-4-10.png)
 
-![$\psi_{11,1}$ for S$_4$](azmy-4-11.png)
+![$\psi_{11}$ for S$_4$](azmy-4-11.png)
 
-![$\psi_{12,1}$ for S$_4$](azmy-4-12.png)
+![$\psi_{12}$ for S$_4$](azmy-4-12.png)
 
[…]

[…]
