+++ doc/tutorials/320-thermal/clean.sh
@@ -1,0 +1,1 @@
+cat .gitignore | sed '/^#.*/ d' | sed '/^\s*$/ d' | sed 's/^/rm -rf /' | grep -v mp4 | bash
