+++ doc/tutorials/make.sh
@@ -1,5 +1,5 @@
 #!/bin/bash -e
-for i in . 000-setup 110-tensile-test 120-mazes; do
+for i in . 000-setup 110-tensile-test 120-mazes 320-thermal; do
   ../md2.sh --pdf  ${i}/README
   ../md2.sh --gfm  ${i}/README
   ../md2.sh --html ${i}/README
