+++ doc/tutorials/320-thermal/syntax.tex
@@ -1,0 +1,1 @@
+../../syntax.tex
