+++ doc/tutorials/320-thermal/README.md
@@ -36,7 +36,7 @@ Here you will learn how to solve the heat conduction equation with FeenoX in all
    FeenoX is able to detect this and will switch to a non-linear solver automatically.
    You can check this by passing `--snes_view` in the command line. Here, SNES means "Scalable Non-linear Equation Solvers" in PETSc's jargon. The `--snes_view` option will show some details about the solver. In linear problems, SNES is not used but the KSP (Krylov SubSpace solvers) framework is used instead. Therefore, `--snes_view` is empty for linear problems. 
 
- * We show how to trigger non-linearities by
+ * We trigger non-linearities by
  
     - adding a boundary conditions that depends on $T(\vec{x})$, such as radiation
     - having a conductivity that depends on temperature, which is the case for most materials anyway
@@ -61,9 +61,94 @@ along with proper boundary conditions.
 
 ## Temperature conditions
 
+The simplest heat conduction problem involves a slab geometry with prescribed temperatures at both ends.
+If the conductivity $k$ is uniform, then the solution is a linear interpolation of these two temperatures---independently of the actual value of the conductivity.
+
 ### Single-material slab
 
+Let us create a unitary slab between $x=0$ and $x=1$ with Gmsh using this `slab.geo`:
+
+```{.c include="slab.geo"}
+```
+
+The end at $x=0$ is called `left` and the one at $x=1$ is called `right`.
+So we can ask FeenoX to solve a thermal problem with uniform conductivity $k$ and fixed temperatures at both ends by
+
+ 1. Defining `PROBLEM` as `thermal` and giving either `1d` or `DIM 1`:
+ 
+    ```feenox
+    PROBLEM thermal 1d
+    ```
+    
+    We can check which problems FeenoX can solve by running it with `--pdes`:
+    
+    ```terminal
+    $ feenox --pdes
+    laplace
+    mechanical
+    modal
+    neutron_diffusion
+    neutron_sn
+    thermal
+    $    
+    ```
+ 
+ 2. Setting the special variable `k` to a constant:
+ 
+    ```feenox
+    k = 1
+    ```
+ 
+    The fact that the conductivity is givens as a variable means that 
+     a. There is a single material
+     b. Its conductivity is uniform, i.e. it does not depend on space
+     
+ 3. Giving `T` equal to the desired temperature values after the `BC` definition for both `left` and `right`
+
+    ```feenox
+    BC left  T=0
+    BC right T=1
+    ```
+
+After the instruction `SOLVE_PROBLEM`, the solution $T(x)$ is available as the one-dimensional function `T(x)`.
+We can then
+ 
+ i. print its definition values with `PRINT_FUNCTION`.
+ ii. evaluate it at any arbitrary location `x`. FeenoX will use the shape functions to interpolate the nodal solutions, and/or
+ 
+Here's a working input file `slab-uniform.fee`:
+
+```{.feenox include="slab-uniform.fee"}
+```
+ 
+```terminal
+$ feenox slab-uniform.fee 
+0       0
+1       1
+0.1     0.1
+0.2     0.2
+0.3     0.3
+0.4     0.4
+0.5     0.5
+0.6     0.6
+0.7     0.7
+0.8     0.8
+0.9     0.9
+# the temperature at x=2/3 is   0.666667
+``` 
+
+> **Homework**
+>
+> 1. Check that the solution does not depend on $k$.
+> 2. Change the values of the boundary conditions and check the result is always a linear interpolation.
+ 
 ### Two-material slab
+
+If we have two materials we can give them in two ways:
+
+ 1. Using the `MATERIALS` keyword, or
+ 2. Appending `_materialname` to either a variable or a function of space.
+
 
 ## Heat flux conditions
 
[…]
