### 2023-11-26 77feca2
non-linear conductivity
