+++ doc/tutorials/320-thermal/README.md
@@ -165,6 +165,9 @@ These problems need at least one fixed-temperature (a.k.a. Dirichlet) condition.
 
 If the material properties or the volumetric heat source depends on the solution $T(\vec{x})$, or the boundary conditions depend non-linearly on $T(\vec{x})$ then the problem is non linear. FeenoX's parser can detect these dependencies so it will use a non-linear solver automatically. That is to say, there is no need for the user to tell the solver which kind of problem it needs to solve---which is reasonable. Why would the user have to tell the solver?
 
+As we all know, solving a non-linear system of equations is far more complex than solving linear problems.
+Even more, the most-widely used non-linear scheme, namely the Newton-Raphson method which is the basis of PETSc's SNES framework, involved repeatedly solving a linear system starting from an initial guess. The matrix associated with these linear solves (which changes from iteration to iteration) is called the jacobian matrix. FeenoX builds an appropriate jacobian for each type of non-linearity, ensuring the convergence is as fast as possible. Advanced users might investigate that indeed the jacobian is correct by using the PETSc options `--snes_test_jacobian` and, for smaller problems, `--snes_test_jacobian_view`. Note that these options render the execution far slower, so make sure the mesh is coarse.
+
 
 ## Temperature-dependent heat flux: radiation
 
@@ -221,19 +224,132 @@ If we wanted to get the temperature in Celsius, we could have done:
 ```{.feenox include="radiation-as-heatflux-celsius.fee"}
 ```
 
-
-
 > **Homework**
 >
-> Rewrite the radiation boundary condition as a convection condition. Hint: note that $T^4 - T_\text{ref}^4$ is a difference of squares.
+> 1. Rewrite the radiation boundary condition as a convection condition. Hint: note that $T^4 - T_\text{ref}^4$ is a difference of squares.
 > Look for `radiation-as-convection.fee` in FeenoX's `tests` directory for the answer.
+> 2. Explain why the solver converges even though there are no prescribed temperature conditions.
 
 ## Temperature-dependent conductivity
 
-UO$_2$
+Another general source of non-linearity in engineering problems modeled as PDEs is due to material properties depending on the unknown.
+For heat conduction, this is true when the thermal conductivity depends on the temperature as a certain function $k(T)$.
+In general, this dependency is given either using 
+ 
+  a. an algebraic expression with a correlation to experimental data, or
+  b. a pointwise-defined "table" with the actual experimental data
+ 
+FeenoX can understand both of them.
+In this section we use the former, and in the next section we use the latter.
+Consider a pellet of uranium dioxide as the ones used inside the fuel elements of nuclear power reactors.
+According to ["Thermophysical Properties of Materials For Nuclear Engineering"](https://www.nuclear-power.com/nuclear-engineering/heat-transfer/thermal-conduction/thermal-conductivity/thermal-conductivity-of-uranium-dioxide/), the thermal conductivity of UO$_2$ can be approximated by
 
+$$
+k(\tau) [ \text{W} \cdot \text{m}^{-1} \cdot \text{K}^{-1} ] = \frac{100}{7.5408· + 17.692 \cdot \tau + 3.614 \tau^2} + \frac{6400}{t^{5/2}} \cdot \exp \left( \frac{-16.35}{\tau} \right)
+$$
+where $\tau = T [ \text{K} ] / 1000$.
+
+How do we tell FeenoX to use this correlation?
[…]
