### 2023-11-26 f69adb6
small additions to thermal tutorial
