+++ doc/tutorials/320-thermal/README.md
@@ -50,11 +50,11 @@ Here you will learn how to solve the heat conduction equation with FeenoX in all
  
 # Linear steady-state problems
 
-In this section we are going to ask FeenoX to compute a temperature distribution $T(\vec{x})$ that satisfies 
+In this section we are going to ask FeenoX to compute a temperature distribution $T(\vec{x})$ that satisfies the linear heat conduction equation
 
 $$
 - \text{div} \Big[ k(\vec{x}) \cdot \text{grad} \left[ T(\vec{x}) \right] \Big] = q(\vec{x})
-$$
+$$ {#eq:heat}
 
 along with proper boundary conditions.
 
@@ -163,10 +163,18 @@ These problems need at least one fixed-temperature (a.k.a. Dirichlet) condition.
 
 # Non-linear state-state problems
 
-If the material properties or the volumetric heat source depends on the solution $T(\vec{x})$, or the boundary conditions depend non-linearly on $T(\vec{x})$ then the problem is non linear. FeenoX's parser can detect these dependencies so it will use a non-linear solver automatically. That is to say, there is no need for the user to tell the solver which kind of problem it needs to solve---which is reasonable. Why would the user have to tell the solver?
+If in the heat @eq:heat above the thermal conductivity $k$ or the volumetric heat source $q$ depends on the solution $T(\vec{x})$, or the boundary conditions depend non-linearly on $T(\vec{x})$ then the problem is non linear. FeenoX's parser can detect these dependencies so it will use a non-linear solver automatically. That is to say, there is no need for the user to tell the solver which kind of problem it needs to solve---which is reasonable. Why would the user have to tell the solver?
+
+
 
 As we all know, solving a non-linear system of equations is far more complex than solving linear problems.
-Even more, the most-widely used non-linear scheme, namely the Newton-Raphson method which is the basis of PETSc's SNES framework, involved repeatedly solving a linear system starting from an initial guess. The matrix associated with these linear solves (which changes from iteration to iteration) is called the jacobian matrix. FeenoX builds an appropriate jacobian for each type of non-linearity, ensuring the convergence is as fast as possible. Advanced users might investigate that indeed the jacobian is correct by using the PETSc options `--snes_test_jacobian` and, for smaller problems, `--snes_test_jacobian_view`. Note that these options render the execution far slower, so make sure the mesh is coarse.
+Even more, the most-widely scheme used to solve the non-linear equation $\vec{F}(\vec{u})=0$,namely the Newton-Raphson method which is the basis of [PETSc's SNES framework](https://petsc.org/release/manual/snes/), involves repeatedly solving a linear system starting from an initial guess $\vec{u}_0$:
+
+ 1. Solve $J(\vec{u}_k) \cdot \Delta \vec{u}_k = -\vec{F}(\vec{u}_k)$
+ 2. Update $\vec{u}_{k+1} \leftarrow \vec{u}_{k} + \Delta \vec{u}_{k}$
+
+
+The matrix $J = \vec{F}^{\prime}$ associated with these linear solves (which changes from iteration to iteration) is called the jacobian matrix. FeenoX builds an appropriate jacobian for each type of non-linearity, ensuring the convergence is as fast as possible. Advanced users might investigate that indeed $J(\vec{u})$ is correct by using the PETSc options `--snes_test_jacobian` and, for smaller problems, `--snes_test_jacobian_view`. Note that these options render the execution far slower, so make sure the mesh is coarse.
 
 
 ## Temperature-dependent heat flux: radiation
[…]
