### 2023-11-27 a83b65b
two more thermal tests
