### 2023-11-27 41dd15a
fixed --linear and --non-linear
