+++ doc/tutorials/README.md
@@ -17,31 +17,19 @@ next_title: \#0 Setup your system
 intro.md
 ```
 
+```{.include shift-heading-level-by=1}
+setup.md
+```
  
-# General tutorials
  
 ```{.include shift-heading-level-by=1}
 general.md
 ```
 
+```{.include shift-heading-level-by=1}
+detailed.md
+```
 
-# Detailed functionality
 
- 1. Input files, expressions and command-line arguments
- 2. Static & transient cases
- 3. Functions & functionals
- 4. Vectors & matrices
- 5. Differential-algebraic equations
- 6. Meshes & distributions
-
-# Physics tutorials
- 
- 1. The Laplace equation
- 2. [Heat conduction](320-thermal)
- 3. Linear elasticity
- 4. Modal analysis
- 5. Thermo-mechanical analysis
- 6. Neutron diffusion
- 7. Neutron transport
 
  
