+++ doc/feenox-manual.md
@@ -46,17 +46,29 @@ compilation.md
 
 # Examples
 
+> See <https://www.seamplex.com/feenox/examples> for updated information.
+
+
 ```{.include shift-heading-level-by=1}
 ../examples/examples.md
 ```
 
 # Tutorials
 
-See <https://www.seamplex.com/feenox/tutorials>
+> See <https://www.seamplex.com/feenox/doc/tutorials> for updated information.
 
 ```{.include shift-heading-level-by=1}
-tutorials/intro.md
+tutorials/setup.md
 ```
+ 
+```{.include shift-heading-level-by=1}
+tutorials/general.md
+```
+
+```{.include shift-heading-level-by=1}
+tutorials/detailed.md
+```
+
 
 
 # Description
