+++ doc/tutorials/detailed.md
@@ -1,0 +1,8 @@
+# Detailed functionality
+
+ 1. Input files, expressions and command-line arguments
+ 2. Static & transient cases
+ 3. Functions & functionals
+ 4. Vectors & matrices
+ 5. Differential-algebraic equations
+ 6. Meshes & distributions
