+++ doc/tutorials/setup.md
@@ -1,0 +1,1 @@
+1. [Setting up your workspace](https://www.seamplex.com/feenox/doc/tutorials/000-setup)
