+++ doc/tutorials/320-thermal/README.md
@@ -15,33 +15,43 @@ prev_title: \#2 Solving mazes
 
 Welcome to **FeenoX’s tutorial number three**.
 Here you will learn how to solve the heat conduction equation with FeenoX in all of its flavors: linear and non-linear, static and transient.
+All the files needed to go through the tutorial are available in [FeenoX's Git repository under `doc/tutorials/320-thermal`](https://github.com/seamplex/feenox/tree/main/doc/tutorials/320-thermal). 
 
 
- * We start solving steady-state problems. As long as neither of the
+ * We start solving linear steady-state problems. As long as neither of the
  
     a. material properties, nor
     b. sources
     
-   depend on the temperature $T(\vec{x})$ and 
+   depend on the temperature $T(\mathbf{x})$ and 
    
-    c. the boundary conditions do not depend or depend linearly on $T(\vec{x})$
+    c. the boundary conditions do not depend or depend linearly on $T(\mathbf{x})$
     
    then problem is linear.
-   If these three guys depend on space $\vec{x}$ (but not on $T(\vec{x})$), the problem is still linear no matter how complex it looks like:
+   If these three guys depend on space $\mathbf{x}$ (but not on $T(\mathbf{x})$), the problem is still linear no matter how complex it looks like:
 
    ```{.feenox include="manufactured.fee"}
    ```
 
- * If any of a. b. or c. do depend on the temperature $T(\vec{x})$, then the problem is non-linear.
-   FeenoX is able to detect this and will switch to a non-linear solver automatically.
-   You can check this by passing `--snes_view` in the command line. Here, SNES means "Scalable Non-linear Equation Solvers" in PETSc's jargon. The `--snes_view` option will show some details about the solver. In linear problems, SNES is not used but the KSP (Krylov SubSpace solvers) framework is used instead. Therefore, `--snes_view` is empty for linear problems. 
-
- * We trigger non-linearities by
+ * If any of a. b. or c. do depend on the temperature $T(\mathbf{x})$, then the problem is non-linear.
+   FeenoX is able to detect this and will switch to a non-linear solver _automatically_.
+   Why would the user need to tell the solver if the problem is linear or not?
+   
+ * We can trigger non-linearities by
  
-    - adding a boundary conditions that depends on $T(\vec{x})$, such as radiation
+    - adding a boundary conditions that depends on $T(\mathbf{x})$, such as radiation
     - having a conductivity that depends on temperature, which is the case for most materials anyway
     - using heat sources that are temperature-dependent, where increasing $T$ decreases the source
-   
+
+ * We can check that FeenoX could detect the non-linearities by passing `--snes_monitor` and/or `--snes_view` in the command line. Here, SNES means "Scalable Non-linear Equation Solvers" in PETSc's jargon. The `--snes_view` option will show some details about the solver. In linear problems, SNES is not used but the KSP (Krylov SubSpace solvers) framework is used instead. Therefore, if we used `--snes_view` in a linear problem then FeenoX will complain about an unused command-line option. 
+ 
+ * If, for some reason, the user does not want to have FeenoX to autodetect the non-linearities then she could force the problem type using either
+ 
+   a. the keywords `LINEAR` and `NON_LINEAR` in the `PROBLEM` definition, or
+   b. the command-line options `--linear` and `--non-linear`.
+ 
[…]
