+++ doc/tutorials/general.md
@@ -1,2 +1,4 @@
- 1. [Overview: the tensile test case](110-tensile-test)
- 2. [Fun & games: solving mazes with PDES instead of AI](120-mazes)
+# General tutorials
+
+ 1. [Overview: the tensile test case](https://www.seamplex.com/feenox/doc/tutorials/110-tensile-test)
+ 2. [Fun & games: solving mazes with PDES instead of AI](https://www.seamplex.com/feenox/doc/tutorials/120-mazes)
