+++ doc/tutorials/physics.md
@@ -1,0 +1,9 @@
+# Physics tutorials
+
+ 1. The Laplace equation
+ 2. [Heat conduction](https://www.seamplex.com/feenox/doc/tutorials/320-thermal)
+ 3. Linear elasticity
+ 4. Modal analysis
+ 5. Thermo-mechanical analysis
+ 6. Neutron diffusion
+ 7. Neutron transport
