### 2023-11-27 ae578b8
tests for --linear and --non-linear
