### 2023-11-27 c832d72
fixed links and DAE examples
