### 2023-11-27 #5846 @gtheler
Dear @vijaysm . I have not finished the thermal tutorial but the non-linear part should be there.
Let me know if I'm on the right track.

> > Let's do as follows. What if I wrote a tutorial (like the two that are already available at https://seamplex.com/feenox/doc/tutorials/) for the thermal case, including all the non-linear features? Would that showcase enough non-linear cases?
>
> Yes.

I haven't finished it, but here's what I've written during the weekend:

 * https://www.seamplex.com/feenox/doc/tutorials/320-thermal/
 * https://www.seamplex.com/feenox/doc/tutorials/320-thermal/README.pdf (the same information in PDF)

I still have to fill in the linear and transient parts. Does the non-linear section make sense for you guys?

> I also want to point out that under the [examples link](https://www.seamplex.com/feenox/doc/feenox-manual.html#examples), **several** hyperlinks are dead. For example, https://www.seamplex.com/feenox/doc/laplace.html#how-to-solve-a-maze-without-ai, https://www.seamplex.com/feenox/doc/thermal.html#thermal-slabs, https://www.seamplex.com/feenox/doc/mechanical.html#parallelepiped-whose-youngs-modulus-is-a-function-of-the-temperature to name a few.

Thanks. They were missing a directory in the URL. It should be now fixed.
Anyway, the source of the examples is https://www.seamplex.com/feenox/examples/

> > The --non-linear option should not be used. FeenoX should detect if a problem is linear or not by itself.
>
> The documentation mentions that one can provide `--linear` or `--non-linear` argument to force FeenoX to use a nonlinear or just a linear solver. But in the examples that I tried, forcing a linear problem to be nonlinear did not work as expected. So it is not clear how these options are implemented if they do not exactly force a behavior.

There was a small bug regarding `--linear` and `--non-linear`.  I fixed it and added tests cases. They should now work as expected. The thermal tutorial covers its usage as well.

> > One can control the basic options for PETSc with the PROBLEM keyword: https://www.seamplex.com/feenox/doc/feenox-manual.html#problem
> This is useful. Please make this link more prominent as I did not notice this before. If you are also using a "prefix" for the solver options, then it will be helpful to mention this in the documentation as well.

I mention this link in the tutorial again. There are no prefixed PETSc objects.
