+++ doc/tutorials/320-thermal/README.md
@@ -14,8 +14,14 @@ prev_title: \#2 Solving mazes
 # Summary
 
 Welcome to **FeenoX’s tutorial number three**.
-Here you will learn how to solve the heat conduction equation with FeenoX in all of its flavors: linear and non-linear, static and transient.
+Here you will learn how to solve the heat conduction equation with FeenoX in all of its flavors: 
+
+ a. linear and non-linear,
+ b. static and transient.
+
 All the files needed to go through the tutorial are available in [FeenoX's Git repository under `doc/tutorials/320-thermal`](https://github.com/seamplex/feenox/tree/main/doc/tutorials/320-thermal). 
+Make sure you also check the [heat conduction examples](https://www.seamplex.com/feenox/examples/thermal.html).
+
 
 
  * We start solving linear steady-state problems. As long as neither of the
@@ -33,24 +39,27 @@ All the files needed to go through the tutorial are available in [FeenoX's Git r
    ```{.feenox include="manufactured.fee"}
    ```
 
- * If any of a. b. or c. do depend on the temperature $T(\mathbf{x})$, then the problem is non-linear.
-   FeenoX is able to detect this and will switch to a non-linear solver _automatically_.
+ * If these conditions are not met, then the problem is non-linear.
+   FeenoX is able to detect this and will **automatically** switch to a non-linear solver.
    Why would the user need to tell the solver if the problem is linear or not?
+   If you think it through, it should be the other way round. And that is what FeenoX does.
    
- * We can trigger non-linearities by
+ * Non-linearities can be triggered by either
  
-    - adding a boundary conditions that depends on $T(\mathbf{x})$, such as radiation
-    - having a conductivity that depends on temperature, which is the case for most materials anyway
-    - using heat sources that are temperature-dependent, where increasing $T$ decreases the source
+    - adding a boundary conditions that depends non-linearly on $T(\mathbf{x})$, such as radiation, and/or
+    - having a conductivity that depends on temperature, which is the case for most materials anyway, and/or
+    - using heat sources that are temperature-dependent, where increasing $T$ decreases the source.
 
- * We can check that FeenoX could detect the non-linearities by passing `--snes_monitor` and/or `--snes_view` in the command line. Here, SNES means "Scalable Non-linear Equation Solvers" in PETSc's jargon. The `--snes_view` option will show some details about the solver. In linear problems, SNES is not used but the KSP (Krylov SubSpace solvers) framework is used instead. Therefore, if we used `--snes_view` in a linear problem then FeenoX will complain about an unused command-line option. 
+ * We can check if FeenoX can detect the non-linearities by passing `--snes_monitor` and/or `--snes_view` in the command line. Here, [SNES](https://petsc.org/release/manual/snes/) means "Scalable Non-linear Equation Solvers" in PETSc's jargon. The `--snes_view` option shows some details about the solver. In linear problems, SNES is not used but the [KSP](https://petsc.org/release/manual/ksp/) (Krylov SubSpace solvers) framework is used instead. Therefore, if we used `--snes_view` in a linear problem then FeenoX would complain about an unused command-line option. 
  
[…]
