+++ tests/mms/thermal/README.md
@@ -5,3 +5,4 @@ This directory has two subdirectories with scripts to perform verification using
  * [2D](2d)
  * [3D](3d)
 
+Check out this video for an explanation about MMS: <https://youtu.be/mGNwL8TGijg>
