### 2023-12-03 9e39198
transient thermal tut & fixes
