+++ examples/thermal.yaml
@@ -5,4 +5,5 @@ titleblock: |
  ===============
 lang: en-US
 toc: true 
+abstract: Check out the [thermal tutorial](https://www.seamplex.com/feenox/doc/tutorials/320-thermal/).
 ...
