### 2023-12-05 2580754
block size = dim
