### 2023-12-09 15a8ec4
transient thermal tutorial in 3d
