+++ doc/tutorials/320-thermal/README.md
@@ -11,7 +11,7 @@ prev_link: ../120-mazes
 prev_title: \#2 Solving mazes
 ...
 
-# Summary
+# Foreword {#sec:foreword}
 
 Welcome to **FeenoX’s tutorial number three**.
 Here you will learn how to solve the heat conduction equation with FeenoX in all of its flavors: 
@@ -22,6 +22,10 @@ Here you will learn how to solve the heat conduction equation with FeenoX in all
 All the files needed to go through the tutorial are available in [FeenoX's Git repository under `doc/tutorials/320-thermal`](https://github.com/seamplex/feenox/tree/main/doc/tutorials/320-thermal). 
 Make sure you also check the [heat conduction examples](https://www.seamplex.com/feenox/examples/thermal.html).
 
+> **Heads up**: this tutorial is quite long. For a quicker introduction, check out the [thermal annotated examples](https://seamplex.com/feenox/examples/thermal.html) in [FeenoX webpage](https://seamplex.com/feenox/).
+ 
+
+## Summary
 
 
  * We start solving linear steady-state problems. As long as neither of the
@@ -71,7 +75,7 @@ Make sure you also check the [heat conduction examples](https://www.seamplex.com
     ii. starting from a steady-state solution and changing the boundary conditions over time
     iii. both
     
- * Transient problems Different time stepping 
+ * If the initial condition does not satisfy the fixed temperature conditions, the solver will not converge. But we can be smart and use FeenoX's functions like [`limit`](https://seamplex.com/feenox/doc/feenox-manual.html#limit), [`if`](https://seamplex.com/feenox/doc/feenox-manual.html#if), [`min`](https://seamplex.com/feenox/doc/feenox-manual.html#min), [`max`](https://seamplex.com/feenox/doc/feenox-manual.html#max), etc. to satisfy them at $t=0$ and then quickly take the boundary conditions to their actual value.
  
 # Linear steady-state problems
 
@@ -363,6 +367,8 @@ Note that since there are four different groups holding the same boundary condit
 ```
 
 ![Output of `bunny-thermal.fee`](square-power.png){#fig:square-power}
+
+> **Heads up**: the volumetric source `q` works as any other material property. In multi-material problems. it can be defined using variables or functions where the material name is appended to the name or using the `MATERIAL` keyword.
 
     
 ## Space-dependent properties: manufactured solution
@@ -761,7 +767,7 @@ Results for the non-uniform power case
 In this final section of the tutorial we solve the transient heat conduction equation
 
 $$
-\rho(\mathbf{x}, T) \cdot c_p(\mathbf{x}, T) \cdot \frac{\partial T}{\partial t} - \text{div} \left[ k(\mathbf{x}, T) \cdot \text{grad}{T} \right] = q'''(\mathbf{x}, T)
+\rho(\mathbf{x}, T,t) \cdot c_p(\mathbf{x}, T,t) \cdot \frac{\partial T}{\partial t} - \text{div} \Big[ k(\mathbf{x}, T,t) \cdot \text{grad} \left[ T(\mathbf{x},t)\right] \Big] = q(\mathbf{x}, T,t)
 $$
 
 For this end, we need the product of the density $\rho$ and heat capacity $c_p$.
@@ -781,11 +787,295 @@ If one needs to stop the transient problem before it reaches the prescribed `end
 The initial condition can be given by defining a function `T_0` of space.
 If there is no `T_0` defined, the initial condition is obtained by solving a steady-state problem with `t=0`.
  
[…]
