### 2023-12-09 b8089a7
transient thermal video
