+++ doc/tutorials/320-thermal/README.md
@@ -38,23 +38,25 @@ Make sure you also check the [heat conduction examples](https://www.seamplex.com
     c. the boundary conditions do not depend or depend linearly on $T(\mathbf{x})$
     
    then problem is linear.
-   If these three guys depend on space $\mathbf{x}$ (but not on $T(\mathbf{x})$), the problem is still linear no matter how complex it looks like:
+   If these three guys depend on space $\mathbf{x}$ (but not on $T(\mathbf{x})$), the problem is still linear no matter how complex it looks like. The following problem (whose output should be a small number, which is a "error" measure) discussed in @sec:mms is still linear:
 
    ```{.feenox include="manufactured.fee"}
    ```
 
  * If these conditions are not met, then the problem is non-linear.
    FeenoX is able to detect this and will **automatically** switch to a non-linear solver.
-   Why would the user need to tell the solver if the problem is linear or not?
+   Why would the user need to tell the solver if the problem is linear or not, as in most FEA tools?
    If you think it through, it should be the other way round. And that is what FeenoX does.
    
  * Non-linearities can be triggered by either
  
-    - adding a boundary conditions that depends non-linearly on $T(\mathbf{x})$, such as radiation, and/or
+    - setting a boundary condition that depends non-linearly on $T(\mathbf{x})$, such as radiation, and/or
     - having a conductivity that depends on temperature, which is the case for most materials anyway, and/or
-    - using heat sources that are temperature-dependent, where increasing $T$ decreases the source.
+    - using a heat source that is temperature-dependent, where increasing $T$ decreases the source.
 
- * We can check if FeenoX can detect the non-linearities by passing `--snes_monitor` and/or `--snes_view` in the command line. Here, [SNES](https://petsc.org/release/manual/snes/) means "Scalable Non-linear Equation Solvers" in PETSc's jargon. The `--snes_view` option shows some details about the solver. In linear problems, SNES is not used but the [KSP](https://petsc.org/release/manual/ksp/) (Krylov SubSpace solvers) framework is used instead. Therefore, if we used `--snes_view` in a linear problem then FeenoX would complain about an unused command-line option. 
+ * We can check if FeenoX can detect the non-linearities by using the **advanced** options `--snes_monitor` and/or `--snes_view` in the command line. Here, [SNES](https://petsc.org/release/manual/snes/) means "Scalable Non-linear Equation Solvers" in PETSc's jargon. The `--snes_view` option shows some details about the solver. In linear problems, SNES is not used but the [KSP](https://petsc.org/release/manual/ksp/) (Krylov SubSpace solvers) framework is used instead. Therefore, if we used `--snes_view` in a linear problem then FeenoX would complain about an unused command-line option.
+ 
+   > **Heads up**: these are **advanced** options. If you did not understand the paragraph above, do not worry. You are still good to go.
  
  * If, for some reason, the user does not want to have FeenoX to auto-detect the non-linearities then she could force the problem type using either
  
@@ -64,8 +66,10 @@ Make sure you also check the [heat conduction examples](https://www.seamplex.com
  * Different linear and non-linear solvers (and pre-conditioners) can be chosen either from the command line (e.g. `--snes_type=newtonls`) or from the `PROBLEM` definition (e.g. `NONLINEAR_SOLVER newtonls`).
   Check out the [FeenoX manual section for the keyword `PROBLEM`](https://www.seamplex.com/feenox/doc/feenox-manual.html#problem) for further details.
[…]
