+++ examples/thermal.yaml
@@ -5,5 +5,7 @@ titleblock: |
  ===============
 lang: en-US
 toc: true 
-abstract: Check out the [thermal tutorial](https://www.seamplex.com/feenox/doc/tutorials/320-thermal/).
+abstract: |
+  Compilation of thermal problems.
+  Check out the [heat conduction tutorial](https://www.seamplex.com/feenox/doc/tutorials/320-thermal/) as well.
 ...
