### 2023-12-16 #5846 @gtheler
Dear all, I finished the heat conduction tutorial: https://seamplex.com/feenox/doc/tutorials/320-thermal/

I agree the documentation is not clear enough so I started to re-design the README to point different users, namely

 1. Academics with knowledge of both PDE mathematics and programming
 2. Advanced users with experience in Unix scripting
 3. Newbies

 I'll keep you posted.
