### 2023-12-17 #5846 @gtheler
> @gtheler did you implement the changes you hinted at to @vijaysm e.g. the thermal tutorials featuring non-linear applications? Are we waiting for you to implement some requested changes?

@Kevin-Mattheus-Moerman  the thermal tutorial explaining all the combinations of linear/non-linear and steady/transient capabilities is ready:

https://seamplex.com/feenox/doc/tutorials/320-thermal/
https://seamplex.com/feenox/doc/tutorials/320-thermal/README.pdf

I'm working on re-designing the README and overall documentation.
Suggestions are welcome.
