### 2023-12-17 #5846 @gtheler
> I suggest fixing the "documentation" link to point to a dedicated page about guidelines to contributing users. Do you support pull requests? How are reviews done for new changes before acceptance? Github workflows are not used for CI and so how do you decide if new changesets are valid? Some details regarding these aspects will be helpful if you intend to receive contributions from new users.

FYI @vijaysm , Github workflows **are** used for CI:
https://github.com/seamplex/feenox/blob/main/.github/workflows/compile_and_test.yaml
https://github.com/seamplex/feenox/actions
