### 2023-12-26 6feb599
joss rewrite plan
