+++ doc/downloads.md
@@ -1,4 +1,4 @@
-FeenoX is distributed under the terms of the [GNU General Public License version 3](https://www.gnu.org/licenses/gpl-3.0.en.html) or (at your option) any later version. See [licensing below](#licensing) for details.
+FeenoX is distributed under the terms of the [GNU General Public License version 3](https://www.gnu.org/licenses/gpl-3.0.en.html) or (at your option) any later version.
 
 |                                        |                                                       |
 |----------------------------------------|-------------------------------------------------------|
