+++ README.md
@@ -11,98 +11,25 @@ toc: true
 
 # About FeenoX
 
-## Why FeenoX?
+FeenoX is a cloud-first free and open source tool to solve engineering-related problems using a computer (or many) with a particular design basis.
+See the [examples](https://www.seamplex.com/feenox/examples/) and the [tutorials](https://www.seamplex.com/feenox/doc/tutorials/) to check out what type of problems it can solve.
+Make sure you also read the [FAQs](doc/FAQ.md), including what the name means.
 
-> If by "Why FeenoX?" you mean "Why is FeenoX named that way?," read the [FAQs](doc/FAQ.md).
+Choose your background for further details about the what, how and whys:
+
+ * [Industry Engineer](README-engineers.md)
+ * [Unix Hacker](README-hackers.md)
+ * [Academic Professor](README-academics.md)
 
 
-The world is already full of finite-element programs and every day a grad student creates a new one from scratch.
-So why adding FeenoX to the already-crowded space of FEA tools?
-Here is ***why FeenoX is different**.
-
-```{.include}
-doc/why.md
-```
-
-## How is FeenoX different?
-
-FeenoX meets fictitious-yet-plausible [Software Requirement Specifications](https://www.seamplex.com/feenox/doc/srs.html).
-The FeenoX [Software Design Specifications](https://www.seamplex.com/feenox/doc/sds.html) address each requirement of the SRS.
-Two of the most important design-basis features are that FeenoX is...
-
- #. a cloud-native computational tool (not just cloud _friendly,_ but cloud **first**).
- #. both free (as in freedom) and open source. See [Licensing].
+## As seen on  "The Science Circle"
  
-But the most important idea is that FeenoX provides a general mathematical framework to solve PDEs with a bunch of entry points (as C functions) where new types of PDEs (e.g. electromagnetism, fluid mechanics, etc.) can be added to the set of what FeenoX can solve. FeenoX will provide means to
-
-  - parse the input file, handle command-line arguments, read mesh files, assign variables, evaluate conditionals, write results, etc.
-
-    ```feenox
-    PROBLEM laplace 2D
-    READ_MESH square-$1.msh
-    [...]
-    WRITE_RESULTS FORMAT vtk
-    ```
-    
-  - handle material properties given as algebraic expressions involving pointwise-defined functions of space, temperature, time, etc.
-    
-    ```feenox
-    MATERIAL steel     E=210e3*(1-1e-3*(T(x,y,z)-20))   nu=0.3
-    MATERIAL aluminum  E=69e3                           nu=7/25
-    ```
-    
-  - read problem-specific boundary conditions as algebraic expressions
-    
-    ```feenox
-    sigma = 5.670374419e-8  # W m^2 / K^4 as in wikipedia
-    e = 0.98      # non-dimensional
-    T0 = 1000     # K
-    Tinf = 300    # K
-
-    BC left  T=T0
-    BC right q=sigma*e*(Tinf^4-T(x,y,z)^4)
-    ```
-
-  - access shape functions and its derivatives evaluated eitehr at Gauss points or at arbitrary locations for computing elementary contributions to
-     * stiffness matrix
-     * mass matrix
-     * right-hand side vector
-    
-  - solve the discretized equations using the appropriate [PETSc](https://petsc.org/)/[SLEPc](https://slepc.upv.es/) objects, i.e.
-    * [KSP](https://petsc.org/release/manual/ksp/) for linear static problems
-    * [SNES](https://petsc.org/release/manual/snes/) for non-linear static problems
-    * [TS](https://petsc.org/release/manual/ts/) for transient problems
-    * [EPS](https://slepc.upv.es/documentation/current/docs/manualpages/EPS/index.html) for eigenvalue problems
-
[…]
