+++ README-hackers.md
@@ -1,0 +1,24 @@
+---
+title: FeenoX for Hackers
+titleblock: |
+ FeenoX for Hackers
+ ====================
+lang: en-US
+number-sections: true
+toc: true
+...
+
+## Why
+
+```{.include}
+doc/word-md-tex.md
+```
+
+## How
+
+cloud first
+unix
+
+## What
+
+SRS + SDS

[…]
