+++ README.md
@@ -11,23 +11,34 @@ toc: true
 
 # About FeenoX
 
-FeenoX is a cloud-first free and open source tool to solve engineering-related problems using a computer (or many) with a particular design basis.
-See the [examples](https://www.seamplex.com/feenox/examples/) and the [tutorials](https://www.seamplex.com/feenox/doc/tutorials/) to check out what type of problems it can solve.
-Make sure you also read the [FAQs](doc/FAQ.md), including what the name means.
+FeenoX is a cloud-first free and open source tool to solve engineering-related problems using a computer (or many computers in parallel) with a particular design basis.
 
-Choose your background for further details about the what, how and whys:
+ * See the [examples](https://www.seamplex.com/feenox/examples/) for an overview of the type of problems FeenoX can solve:
 
- * [Industry Engineer](README-engineers.md)
- * [Unix Hacker](README-hackers.md)
- * [Academic Professor](README-academics.md)
+    #. [Basic mathematics](https://seamplex.com/feenox/examples/basic.html)
+    #. [Systems of ODEs/DAEs](https://seamplex.com/feenox/examples/daes.html)
+    #. [Laplace’s equation](https://seamplex.com/feenox/examples/laplace.html)
+    #. [Heat conduction](https://seamplex.com/feenox/examples/thermal.html)
+    #. [Linear elasticity](https://seamplex.com/feenox/examples/mechanical.html)
+    #. [Modal analysis](https://seamplex.com/feenox/examples/modal.html)
+    #. [Neutron diffusion](https://seamplex.com/feenox/examples/neutron_diffusion.html)
+    #. [Neutron $S_N$](https://seamplex.com/feenox/examples/neutron_sn.html)
 
+ * Take the [tutorials](https://www.seamplex.com/feenox/doc/tutorials/) to learn how to solve them yourself.
+ 
+ * Choose your background for further details about the what, how and whys:
+
+   - [Industry Engineer](README-engineers.md)
+   - [Unix Hacker](README-hackers.md)
+   - [Academic Professor](README-academics.md)
+
+ * Make sure you also read the [FAQs](doc/FAQ.md), including the question about what the name means.
 
 ## As seen on  "The Science Circle"
  
 > ::::: {.container .text-center .my-5 .ratio .ratio-16x9}
 > <iframe class="embed-responsive-item" src="https://www.youtube.com/embed/EZSjFzJAhbw?rel=0" allowfullscreen></iframe>
 > :::::
-
 
  
 # Quickstart
@@ -49,32 +60,23 @@ See the [download page](https://seamplex.com/feenox/download.html) for more deta
 
 # Examples
 
-The [examples page](https://seamplex.com/feenox/examples) contains dozens of curated and annotated cases.
-The FeenoX examples are divided by the type of problem they solve:
+ * The [examples page](https://seamplex.com/feenox/examples) contains dozens of curated and annotated cases.
 
- #. [Basic mathematics](https://seamplex.com/feenox/examples/basic.html)
- #. [Systems of ODEs/DAEs](https://seamplex.com/feenox/examples/daes.html)
- #. [Laplace’s equation](https://seamplex.com/feenox/examples/laplace.html)
- #. [Heat conduction](https://seamplex.com/feenox/examples/thermal.html)
- #. [Linear elasticity](https://seamplex.com/feenox/examples/mechanical.html)
- #. [Modal analysis](https://seamplex.com/feenox/examples/modal.html)
- #. [Neutron diffusion](https://seamplex.com/feenox/examples/neutron_diffusion.html)
- #. [Neutron $S_N$](https://seamplex.com/feenox/examples/neutron_sn.html)
[…]
