### 2023-12-28 3fc2751
cleanup main readme
