+++ README.md
@@ -15,15 +15,11 @@ FeenoX is a cloud-first free and open source tool to solve engineering-related p
 
 ::::: {.only-in-format .html }
 ```{=html}
-<div class="btn-group" role="group" aria-label="FeenoX for...">
- <a class="btn btn-lg btn-outline-primary"   href="https://www.seamplex.com/feenox/README-engineers.html" role="button">FeenoX for Engineers</a>
- <a class="btn btn-lg btn-outline-secondary" href="https://www.seamplex.com/feenox/README-hackers.html"   role="button">FeenoX for Hackers</a>
- <a class="btn btn-lg btn-outline-info"      href="https://www.seamplex.com/feenox/README-academics.html" role="button">FeenoX for Academics</a>
-</div>
+<a class="btn btn-lg btn-outline-primary"   href="https://www.seamplex.com/feenox/README-engineers.html" role="button">FeenoX for Industry Engineers</a>
+<a class="btn btn-lg btn-outline-secondary" href="https://www.seamplex.com/feenox/README-hackers.html"   role="button">FeenoX for Unix Hackers</a>
+<a class="btn btn-lg btn-outline-info"      href="https://www.seamplex.com/feenox/README-academics.html" role="button">FeenoX for Academic Professors</a>
 ```
 :::::
-
-::::: {.not-in-format .html }
 
 Choose your background for further details about the what, how and whys:
 
@@ -31,59 +27,73 @@ Choose your background for further details about the what, how and whys:
  - [Unix Hacker](README-hackers.md)
  - [Academic Professor](README-academics.md)
 
-:::::
 
 ## Extents
 
- * See the [examples](https://www.seamplex.com/feenox/examples/) for an overview of the type of problems FeenoX can solve:
+### Examples
 
-    - [Basic mathematics](https://seamplex.com/feenox/examples/basic.html)
-    - [Systems of ODEs/DAEs](https://seamplex.com/feenox/examples/daes.html)
-    - [Laplace’s equation](https://seamplex.com/feenox/examples/laplace.html)
-    - [Heat conduction](https://seamplex.com/feenox/examples/thermal.html)
-    - [Linear elasticity](https://seamplex.com/feenox/examples/mechanical.html)
-    - [Modal analysis](https://seamplex.com/feenox/examples/modal.html)
-    - [Neutron diffusion](https://seamplex.com/feenox/examples/neutron_diffusion.html)
-    - [Neutron S<sub>N</sub>](https://seamplex.com/feenox/examples/neutron_sn.html)
+ - [Basic mathematics](https://seamplex.com/feenox/examples/basic.html)
+ - [Systems of ODEs/DAEs](https://seamplex.com/feenox/examples/daes.html)
+ - [Laplace’s equation](https://seamplex.com/feenox/examples/laplace.html)
+ - [Heat conduction](https://seamplex.com/feenox/examples/thermal.html)
+ - [Linear elasticity](https://seamplex.com/feenox/examples/mechanical.html)
+ - [Modal analysis](https://seamplex.com/feenox/examples/modal.html)
+ - [Neutron diffusion](https://seamplex.com/feenox/examples/neutron_diffusion.html)
+ - [Neutron S<sub>N</sub>](https://seamplex.com/feenox/examples/neutron_sn.html)
 
- * Take the [tutorials](https://www.seamplex.com/feenox/doc/tutorials/) to learn how to solve them yourself.
+### Tutorials
+
+
+ 0. [Setting up your workspace](https://www.seamplex.com/feenox/doc/tutorials/000-setup)
[…]
