+++ doc/README.md
@@ -8,22 +8,17 @@ lang: en-US
 
 > FeenoX documentation is released under the terms of the [GNU Free Documentation License v1.3](https://www.gnu.org/licenses/fdl-1.3.html), or any later version.
 
-Every bit of FeenoX documentation is written in [Pandoc](https://pandoc.org/)-flavored Markdown.
+Every bit of FeenoX documentation is written in [Pandoc](https://pandoc.org/)-flavored Markdown in the [`doc`](https://github.com/seamplex/feenox/tree/main/doc) directory of the [FeenoX repository](https://github.com/seamplex/feenox/).
 It is then converted to HTML, PDF, Unix manpage or Texinfo as needed.
 
 
-# FeenoX manual and description
+# Manual and description
 
 
 As per the [GNU Coding Standards](https://www.gnu.org/prep/standards/standards.html#GNU-Manuals), "a manual should serve both as tutorial and reference." Due to the formatting restrictions, the [Texinfo](https://www.gnu.org/software/texinfo/) version contains only the description and not the full reference.
 
- * [HTML manual](https://www.seamplex.com/feenox/doc/feenox-manual.html)
- * [PDF manual](https://www.seamplex.com/feenox/doc/feenox-manual.pdf)
- * [Texinfo description](https://www.seamplex.com/feenox/doc/feenox-desc.texi)
- * [HTML description](https://www.seamplex.com/feenox/doc/feenox-desc.html)
- * [PDF description](https://www.seamplex.com/feenox/doc/feenox-desc.pdf)
-
-The sources are in the [doc](https://github.com/seamplex/feenox/tree/main/doc) directory of the Git repository.
+ * [FeenoX manual](https://www.seamplex.com/feenox/doc/feenox-manual.html) [PDF](https://www.seamplex.com/feenox/doc/feenox-manual.pdf)
+ * [FeenoX description](https://www.seamplex.com/feenox/doc/feenox-desc.html) [PDF](https://www.seamplex.com/feenox/doc/feenox-desc.pdf) [Texinfo](https://www.seamplex.com/feenox/doc/feenox-desc.texi)
 
 
 ## Design
@@ -40,111 +35,136 @@ The sources are in the [doc](https://github.com/seamplex/feenox/tree/main/doc) d
 > <iframe class="embed-responsive-item" src="https://www.youtube.com/embed/66WvYTb4pSg?rel=0" allowfullscreen></iframe>
 > :::::
 ::::::::: 
-
  
  * [Software Requirements specifications](srs.md) [PDF](https://www.seamplex.com/feenox/doc/srs.pdf) (Fictitious RFQ)
  * [Software Design specifications](sds.md) [PDF](https://www.seamplex.com/feenox/doc/sds.pdf) (Imaginary FeenoX's offer)
- * [Mid-term PhD review presentation slides with the SRS/SDS explanation in PDF](https://www.seamplex.com/feenox/doc/2021-feenox.pdf) (August 2021).
-   - [Video recording of the presentation](https://youtu.be/-RJ5qn7E9uE) (Slides are in English but audio is in Spanish).
-   - Browse the repository with the sources of the examples in the presentation [here](https://github.com/gtheler/2021-presentation)
- * [A free and open source computational tool for solving (nuclear-related) differential equations in the cloud](https://www.seamplex.com/feenox/doc/2021-brasil.pdf) (INAC/ENFIR Congress, December 2021).
-   - [Video recording of the presentation](https://youtu.be/e8kFmFOsbPk) (Turn on the CCs, I am not a native English speaker).
-   - Browse the repository with the sources of the examples in the presentation [here](https://github.com/gtheler/2021-brasil)
-   
-## Frequently Asked Questions
 
[…]
