+++ doc/sds.md
@@ -1,7 +1,7 @@
 ---
-title: FeenoX Software Design Specification
+title: FeenoX Software Design Specifications
 lang: en-US
-abstract: This Software Design Specifications (SDS) document applies to an imaginary Software Requirement Specifications (SRS) document issued by a fictitious agency asking for vendors to offer a free and open source cloud-based computational tool to solve engineering problems. The latter can be seen as a “Request for Quotation” and the former as an offer for the fictitious tender. Each section  of this SDS addresses one section of the SRS. The original text from the SRS is shown quoted at the very beginning before the actual SDS content.
+abstract: This Software Design Specifications (SDS) document applies to an imaginary [Software Requirement Specifications](./srs.md) (SRS) document issued by a fictitious agency asking for vendors to offer a free and open source cloud-based computational tool to solve engineering problems. The latter can be seen as a “Request for Quotation” and the former as an offer for the fictitious tender. Each section  of this SDS addresses one section of the SRS. The original text from the SRS is shown quoted at the very beginning before the actual SDS content.
 number-sections: true
 toc: true
 documentclass: report
@@ -13,11 +13,21 @@ documentclass: report
 > 100-introduction.md
 > ```
 
-[FeenoX](https://www.seamplex.com/feenox/) is a cloud-first computational tool aimed at solving engineering problems.
+[FeenoX](https://www.seamplex.com/feenox/) is a cloud-first computational tool aimed at solving engineering problems with a particular design basis.
+
+### "Cloud first" vs. "cloud friendly" {#cloud-first .unnumbered}
+
+```include
+cloud-first.md
+```
+
+### Unfair advantage {.unnumbered}
 
 ```include
 word-md-tex.md
 ```
+
+### Licensing {.unnumbered}
 
 Besides noting that software being _free_ (regarding freedom, not price) does not imply the same as being _open source_, the requirement is clear in that the tool has to be both _free_ and _open source_, a combination which is usually called [FOSS](https://en.wikipedia.org/wiki/Free_and_open-source_software). 
 
@@ -25,10 +35,11 @@ Besides noting that software being _free_ (regarding freedom, not price) does no
 licensing.md
 ```
 
-To sum up this section, FeenoX is...
+To sum up this introduction, FeenoX is...
 
- 1. a cloud-first computational tool (not just cloud _friendly,_ but cloud **first**).
- 2. both free ([as in freedom](https://en.wikipedia.org/wiki/Free_as_in_Freedom)) and open source.
+ #. a cloud-first computational tool (not just cloud _friendly,_ but cloud **first**).
+ #. to traditional computational software and to specialized libraries what [Markdown](https://en.wikipedia.org/wiki/Markdown) is to [Word](https://en.wikipedia.org/wiki/Microsoft_Word) and [TeX](https://en.wikipedia.org/wiki/TeX), respectively.
+ #. both free ([as in freedom](https://en.wikipedia.org/wiki/Free_as_in_Freedom)) and open source.
 
 
 
@@ -96,7 +107,7 @@ There are other FEA tools that even though born closer in time, also follow the 
 lorenz.md
 ```
 
-As already stated, FeenoX is designed and implemented following the UNIX philosophy in general and Eric Raymond's\ 17\ Unix Rules (@sec:unix) in particular. One of the main ideas is the _rule of separation_ that essentially asks to separate mechanism from policy, that in the computational engineering world translates into separating the frontend from the backend. The usage of FeenoX to compute and of Gnuplot to plot is a clear example of separation. Same idea applies to partial differential equations (PDEs), where the mesh is created with Gmsh and the output can be post-processed with Gmsh, Paraview or any other post-processing system (even a web-based interface) that follows rule of separation.
[…]
