+++ doc/contributing-guidelines.md
@@ -1,0 +1,18 @@
+Contributions from [hackers](http://localhost/milhouse/feenox/README4hackers.html) and/or [academics](http://localhost/milhouse/feenox/README4academics.html) are welcome, especially new types of PDEs and new formulations of existing PDEs.
+For elliptic operators feel free to use the Laplace equation at [`src/pdes/laplace`](https://github.com/seamplex/feenox/tree/main/src/pdes/laplace) as a template.
+
+ 1. Read the [Programming Guide](https://seamplex.com/feenox/doc/programming.html).
+ 2. Browse [Github discussions](https://github.com/seamplex/feenox/discussions) and open a new thread explaining what you want to do and/or asking for help.
+ 3. Fork the [Git repository](https://github.com/seamplex/feenox/) under your Github account
+ 4. Create a pull request, including
+    - code,
+    - documentation, and
+    - tests.
+ 5. Follow up the review procedure.
+
+Note that
+
+ * It is mandatory to observe the [Code of Conduct](https://seamplex.com/feenox/doc/CODE_OF_CONDUCT.html).
+ * The contributed code has to be compatible with the [GPLv3+ license](https://www.seamplex.com/feenox/#sec:licensing).
+ * Each author keeps the copyright of the contribution.
+ * You can [ask](https://github.com/seamplex/feenox/discussions)!
