+++ README.md
@@ -11,13 +11,13 @@ toc: true
 
 # About FeenoX
 
-[FeenoX](https://seamplex.com/feenox) is a [cloud-first](https://seamplex.com/feenox/doc/sds.html#cloud-first) [free and open source](#licensing)  tool to solve [engineering-related problems](#extentes) using a computer (or many computers in parallel) with a [particular design basis](./doc/srs.md).
+[FeenoX](https://seamplex.com/feenox) is a [cloud-first](./doc/sds.md#cloud-first) [free and open source](#licensing)  tool to solve [engineering-related problems](#extents) using a computer (or many computers in parallel) with a [particular design basis](./doc/srs.md).
 
 ::::: {.only-in-format .html }
 ```{=html}
-<a class="btn btn-lg btn-outline-primary"   href="https://www.seamplex.com/feenox/README4engineers.html" role="button">FeenoX for Industry Engineers</a>
-<a class="btn btn-lg btn-outline-secondary" href="https://www.seamplex.com/feenox/README4hackers.html"   role="button">FeenoX for Unix Hackers</a>
-<a class="btn btn-lg btn-outline-info"      href="https://www.seamplex.com/feenox/README4academics.html" role="button">FeenoX for Academic Professors</a>
+<a class="btn btn-lg btn-outline-primary"   href="./README4engineers.md" role="button">FeenoX for Industry Engineers</a>
+<a class="btn btn-lg btn-outline-secondary" href="./README4hackers.md"   role="button">FeenoX for Unix Hackers</a>
+<a class="btn btn-lg btn-outline-info"      href="./README4academics.md" role="button">FeenoX for Academic Professors</a>
 ```
 :::::
 
@@ -32,7 +32,7 @@ Choose your background for further details about the what, how and whys:
 
 ### Examples
 
- - [Basic mathematics](https://seamplex.com/feenox/examples/basic.html)
+ - [Basic mathematics](./examples/basic.md)
  - [Systems of ODEs/DAEs](https://seamplex.com/feenox/examples/daes.html)
  - [Laplace’s equation](https://seamplex.com/feenox/examples/laplace.html)
  - [Heat conduction](https://seamplex.com/feenox/examples/thermal.html)
@@ -137,17 +137,17 @@ The [`tests`](https://github.com/seamplex/feenox/tree/main/tests) directory in t
 
 # Documentation
 
-Browse through the [documentation index](https://seamplex.com/feenox/doc/) and/or the [`doc`](https://github.com/seamplex/feenox/tree/main/doc) subdirectory of the [Github repository](https://github.com/seamplex/feenox/) for
+Browse through the [documentation index](doc) and/or the [`doc`](https://github.com/seamplex/feenox/tree/main/doc) subdirectory of the [Github repository](https://github.com/seamplex/feenox/) for
 
- * [FAQs](https://seamplex.com/feenox/doc/FAQ.html), including the [what FeenoX means](https://seamplex.com/feenox/doc/FAQ.html#what-does-feenox-mean)
- * [Manual](https://www.seamplex.com/feenox/doc/feenox-manual.html)
- * [Description](https://www.seamplex.com/feenox/doc/feenox-desc.html)
- * [Software Design Requirements](https://seamplex.com/feenox/doc/srs.html)
- * [Software Design Specifications](https://seamplex.com/feenox/doc/sds.html)
- * [Unix man page](https://www.seamplex.com/feenox/doc/feenox.1.html) (accesible through `man feenox` after installation)
- * [History](https://seamplex.com/feenox/doc/history.html)
- * [Compilation guide](https://seamplex.com/feenox/doc/compilation.html)
[…]
