+++ doc/md2.sh
@@ -233,7 +233,6 @@ fi
 
 # post
 if [ "x${format}" = "x.html" ]; then
-  echo pistola
   n1=$(grep -n side_toc_start ${out} | cut -d':' -f1)
   n2=$(grep -n side_toc_end   ${out} | cut -d':' -f1)
 
@@ -265,7 +264,6 @@ if [ "x${format}" = "x.html" ]; then
   sed -i 's/.\/\(.*\)\.md/\1.html/g'  ${out}
   
 elif [ "x${format}" = "x.markdown" ]; then
-  echo mongocho
   sed -i 's/.\/\(.*\)\.md/\1.markdown/g'  ${out}
   
 fi
