+++ src/pdes/mechanical/parser.c
@@ -118,113 +118,117 @@ int feenox_parse_linearize_stress(void) {
   
   feenox_linearize_t *linearize = NULL;
   feenox_check_alloc(linearize = calloc(1, sizeof(feenox_linearize_t)));
+
+///kw_mechanical+LINEARIZE_STRESS+desc Compute linearized membrane and/or bending stresses according to ASME\ VIII Div\ 2 Sec\ 5.
+///kw_mechanical+LINEARIZE_STRESS+usage LINEARIZE_STRESS
+  
   
   char *token = NULL;
   while ((token = feenox_get_next_token(NULL)) != NULL) {
-///kw_pde+LINEARIZE_STRESS+usage FROM <x1> <y1> <z1>
-///kw_pde+LINEARIZE_STRESS+detail The stress classification line (SCL) defined by the coordinates of
-///kw_pde+LINEARIZE_STRESS+detail the points\ $[x_1, y_1, z_1]$ and $[x_2, y_2, z_2]$.
-///kw_pde+LINEARIZE_STRESS+detail For two-dimensional problems, the $z$ coordinate has to be given as well.
+///kw_mechanical+LINEARIZE_STRESS+usage FROM <x1> <y1> <z1>
+///kw_mechanical+LINEARIZE_STRESS+detail The stress classification line (SCL) defined by the coordinates of
+///kw_mechanical+LINEARIZE_STRESS+detail the points\ $[x_1, y_1, z_1]$ and $[x_2, y_2, z_2]$.
+///kw_mechanical+LINEARIZE_STRESS+detail For two-dimensional problems, the $z$ coordinate has to be given as well.
     if (strcasecmp(token, "FROM") == 0) {
       feenox_call(feenox_parser_expression(&linearize->x1));
       feenox_call(feenox_parser_expression(&linearize->y1));
       feenox_call(feenox_parser_expression(&linearize->z1));
       
-///kw_pde+LINEARIZE_STRESS+usage TO <x2> <y2> <z2> @
+///kw_mechanical+LINEARIZE_STRESS+usage TO <x2> <y2> <z2> @
     } else if (strcasecmp(token, "TO") == 0) {
       feenox_call(feenox_parser_expression(&linearize->x2));
       feenox_call(feenox_parser_expression(&linearize->y2));
       feenox_call(feenox_parser_expression(&linearize->z2));
 
-///kw_pde+LINEARIZE_STRESS+detail The linearized membrane, membrane plus bending and peak total stresses are 
-///kw_pde+LINEARIZE_STRESS+detail stored in the variables given by the keywords `M`, `MB` and `P`, respectively.
-///kw_pde+LINEARIZE_STRESS+detail These three variables use the von\ Mises stress intensity.
-///kw_pde+LINEARIZE_STRESS+usage [ M <variable> ]
+///kw_mechanical+LINEARIZE_STRESS+detail The linearized membrane, membrane plus bending and peak total stresses are 
+///kw_mechanical+LINEARIZE_STRESS+detail stored in the variables given by the keywords `M`, `MB` and `P`, respectively.
+///kw_mechanical+LINEARIZE_STRESS+detail These three variables use the von\ Mises stress intensity.
+///kw_mechanical+LINEARIZE_STRESS+usage [ M <variable> ]
     } else if (strcasecmp(token, "M") == 0) {
       feenox_call(feenox_parser_get_or_define_variable(&linearize->M));
       mechanical.sigma->used = 1;
 
-///kw_pde+LINEARIZE_STRESS+usage [ MB <variable> ]
[…]
