+++ doc/tests-list.md
@@ -1,0 +1,6 @@
+The [`tests`](https://github.com/seamplex/feenox/tree/main/tests) directory in the repository has hundreds of
+ 
+ - `grep`-able examples,
+ - unit tests,
+ - regression tests, and/or
+ - (crude) mathematical verification tests.
