+++ doc/examples-list.md
@@ -1,0 +1,9 @@
+ - [Basic mathematics](https://seamplex.com/feenox/examples/examples/basic.html)
+ - [Systems of ODEs/DAEs](https://seamplex.com/feenox/examples/daes.html)
+ - [Laplace’s equation](https://seamplex.com/feenox/examples/laplace.html)
+ - [Heat conduction](https://seamplex.com/feenox/examples/thermal.html)
+ - [Linear elasticity](https://seamplex.com/feenox/examples/mechanical.html)
+ - [Modal analysis](https://seamplex.com/feenox/examples/modal.html)
+ - [Neutron diffusion](https://seamplex.com/feenox/examples/neutron_diffusion.html)
+ - [Neutron S<sub>N</sub>](https://seamplex.com/feenox/examples/neutron_sn.html)
+ 
