+++ README4hackers.md
@@ -40,9 +40,7 @@ Unlike other similar FEA tools, FeenoX provides...
    ```
 
 
- * a parser for a [syntactically-sugared](https://seamplex.com/feenox/doc/sds.html#syntactic-sugar-highlighting)) self-explanatory ASCII file (passed as the first non-option argument to the `feenox` executable)  with keywords that completely define the problem without requiring further human actions, allowing a [cloud-first](https://seamplex.com/feenox/doc/sds.html#cloud-first) workflow
- 
- * continuous integration (using Github actions), an issue tracker (using Github issues) and a discussion page (using Github discussinos)
+ * a parser for a [syntactically-sugared](https://seamplex.com/feenox/doc/sds.html#syntactic-sugar-highlighting)) self-explanatory ASCII file (passed as the first non-option argument to the `feenox` executable)  with keywords that completely define the problem without requiring further human actions, allowing a [cloud-first](https://seamplex.com/feenox/doc/sds.html#cloud-first) workflow using containerized images (because there is no need to recompile the binary for each problem)
  
  * a Git repository with GPL sources (and FDL documentation) where contributions are welcome. In particular, each partial differential equation that FeenoX can solve correspondence to one of the subdirectories of `src/pdes`. The `autogen.sh` step (prior to `./configure` and `make`) detects the directory structure and includes all the subdirectories it finds as available [problem types](https://www.seamplex.com/feenox/doc/feenox-manual.html#problem). They can be queried at runtime with the `--pdes` option:
  
@@ -56,7 +54,9 @@ Unlike other similar FEA tools, FeenoX provides...
     thermal
     $ 
     ```
- 
+    
+ * continuous integration (using Github actions), an issue tracker (using Github issues) and a discussion page (using Github discussinos)
+  
  * a mechanism to expand command-line arguments as literal strings in the input file so as to allow parametric (and/or optimization) loops. For instance, if an input file `print.fee` looks like
  
    ```fee
@@ -74,7 +74,7 @@ Unlike other similar FEA tools, FeenoX provides...
    $ 
    ``` 
    
- * can come from `stdin` by passing `-` as the input file path:
+ * the possibility to provide the input from `stdin` (so as to use it as a Unix pipe) by passing `-` as the input file path:
  
    ```terminal
    $ for i in $(seq 1 5); do echo "PRINT 2*\${1}" | feenox - $i; done
@@ -86,27 +86,82 @@ Unlike other similar FEA tools, FeenoX provides...
    $ 
    ``` 
  
- * flexibility to handle many workflows, including [web-based interfaces](https://www.caeplex.com)
+ * flexibility to handle many workflows, including [web-based interfaces](https://www.caeplex.com) and thin command-line clients.
  
 The input file...
 
- - has one-to-one correspondence between computer input file and human description of the problem
- - is Git-traceable (the mesh is defined in a separate file which may or may not be tracked)
+ - has one-to-one correspondence with the human description of the problem
+ - is Git-traceable (the mesh is defined in a separate file created by Gmsh, which may or may not be tracked)
  - allows the user to enter algebraic expressions whenever a numerical value is needed (everything is an expression)
  - understands definitions (nouns) and instructions (verbs). FeenoX has an actual instruction pointer that loops over the instruction set (there might even be conditional blocks).
  - is simple for simple files (but might get more complicated for mor complex problems). Remember Alan Kay's quote: "simple things should be simple and complex things should be possible."
  
  
[…]
