+++ doc/tutorials-list.md
@@ -1,0 +1,4 @@
+ 0. [Setting up your workspace](https://www.seamplex.com/feenox/doc/tutorials/000-setup)
+ 1. [Overview: the tensile test case](https://www.seamplex.com/feenox/doc/tutorials/110-tensile-test)
+ 2. [Fun & games: solving mazes with PDES instead of AI](https://www.seamplex.com/feenox/doc/tutorials/120-mazes)
+ 3. [Heat conduction](https://www.seamplex.com/feenox/doc/tutorials/320-thermal)
