+++ doc/sds.md
@@ -532,9 +532,28 @@ Regarding storage, FeenoX needs space to store the input file (negligible), the 
 > 240-scalability.md
 > ```
 
-The time needed to solve a relatively large problem can be reduced by exploiting the fact that most cloud servers (and even laptop computers) have more than one CPU available. There are some tasks that can be split into several processors sharing a common memory address space that will scale up almost perfectly, such as building the elemental matrices and assembling the global stiffness matrix. There are some other tasks that might not scale perfectly but that nevertheless might (or might not) reduce the overall wall time if split among processors using a common memory space, such as solving the linear system\ $K \cdot \vec{u} = \vec{b}$. The usual scheme to parallelize a problem under these conditions is to use the [OpenMP](https://en.wikipedia.org/wiki/OpenMP) framework.
+When for a fixed problem the mesh is refined over and over, more and more computational resources are needed to solve it (and to obtain more accurate results, of course).
+Parallelization can help to 
 
-Yet, if the problem is large enough, a server might not have enough physical random-access memory to be able to handle the whole problem. The problem now has to be split among different servers which, in turn, might have several processors each. Some of the processors share the same address space but most of them will only have access to a fraction of the whole global problem data. In these cases, there are no tasks that can scale up perfectly since even when building and assembling the matrices, a processor needs some piece of data which is handled by another processor with a different address space and that has to be conveyed specifically from one process to another one. The usual scheme to parallelize a problem under these conditions is to use the [MPI](https://en.wikipedia.org/wiki/Message_Passing_Interface) standard and one of its two most well-known implementations, either [Open\ MPI](https://www.open-mpi.org/) or [MPICH](https://www.mpich.org/).
+ a. reduce the wall time needed to solve a problem by using several processors at the same time
+ b. allow to solve big problems that would not fit into a single computer by splitting them into smaller parts, each of them fitting in a single computer
+
+There are three types of parallelization schemes:
+
+Shared-memory systems (OpenMP)
+
+:   several processing units sharing a single memory address space
+
+Distributed systems (MPI)
+
+:   several computational units, each with their own processing units and memory, inter-connected with high-speed network hardware
+
+Graphical processing units (GPU)
+
+:   used as co-processors to solve numerically-intensive problems
+
+In principle, any of these three schemes can be used to reduce the wall time (a).
+But only the distributed systems scheme allows to solve arbitrarily big problems (b).
 
[…]
