+++ doc/feenox-desc.md
@@ -78,7 +78,7 @@ description.md
 ```
 
 
-# FeenoX & the UNIX Philospohy
+# FeenoX & the Unix Philospohy
 
 ```{.include shift-heading-level-by=1}
 unix.md
