+++ doc/sds.md
@@ -29,8 +29,6 @@ word-md-tex.md
 
 ### Licensing {.unnumbered}
 
-Besides noting that software being _free_ (regarding freedom, not price) does not imply the same as being _open source_, the requirement is clear in that the tool has to be both _free_ and _open source_, a combination which is usually called [FOSS](https://en.wikipedia.org/wiki/Free_and_open-source_software). 
-
 ```include
 licensing.md
 ```
@@ -49,7 +47,8 @@ To sum up this introduction, FeenoX is...
 > 110-objective.md
 > ```
 
-The choice of the initial supported features is based on the types of problem that the FeenoX's precursor codes (namely wasora, Fino and milonga, referred to as "previous versions" from now on) already have been supporting since more than ten years now. It is also a first usable version so scope can be bounded. A subsequent road map and release plans can be designed as requested. FeenoX's first version includes a subset of the required functionality, namely
+The choice of the initial supported features is based on the types of problem that the FeenoX's precursor codes (namely wasora, Fino and milonga, referred to as "previous versions" from now on) already have been supporting since more than ten years now.
+A subsequent road map and release plans can be designed as requested. FeenoX's first version includes a subset of the required functionality, namely
 
  * open and closed-loop dynamical systems
  * Laplace/Poisson/Helmholtz equations
@@ -60,7 +59,10 @@ The choice of the initial supported features is based on the types of problem th
 
 @Sec:extensibility explains the mechanisms that FeenoX provides in order to add (or even remove) other types of problems  to be solved.
  
-Recalling that FeenoX is a “cloud-first” tool, it is designed to be developed and executed primarily on [GNU/Linux](https://www.gnu.org/gnu/linux-and-gnu.html) hosts, which is the architecture of more than 90% of the internet servers which we collectively call “the public cloud.” It should be noted that GNU/Linux is a [POSIX](https://en.wikipedia.org/wiki/POSIX)-compliant operating system which is compatible with [Unix](https://en.wikipedia.org/wiki/Unix), and that FeenoX follows the rules of Unix philosophy (further explained in @sec:unix) for the actual computational implementation. Besides the POSIX standard, as explained further below in @sec:scalability, FeenoX also uses [MPI](https://en.wikipedia.org/wiki/Message_Passing_Interface) which is a well-known industry standard for massive execution of parallel processes, both in multi-core hosts and multi-hosts environments. Finally, if performance and/or scalability are not important issues, FeenoX can be run in a (properly cooled) local PC, laptop or even in embedded systems such as [Raspberry\ Pi](https://en.wikipedia.org/wiki/Raspberry_Pi) (see @sec:architecture).
+Recalling that FeenoX is a “cloud-first” tool as explained in @sec:introduction, it is designed to be developed and executed primarily on [GNU/Linux](https://www.gnu.org/gnu/linux-and-gnu.html) hosts, which is the architecture of more than 90% of the internet servers which we collectively call “the public cloud.”
+It should be noted that GNU/Linux is a [POSIX](https://en.wikipedia.org/wiki/POSIX)-compliant operating system which is compatible with [Unix](https://en.wikipedia.org/wiki/Unix), and that FeenoX was designed and implemented following the rules of Unix philosophy which is further explained in @sec:unix.
+Besides the POSIX standard, as explained below in @sec:scalability, FeenoX also uses [MPI](https://en.wikipedia.org/wiki/Message_Passing_Interface) which is a well-known industry standard for massive execution of parallel processes following the distributed-systems parallelization paradigm.
[…]
