### 2024-01-11 #5846 @jeremyonscale
Hi @Kevin-Mattheus-Moerman .
Everything's fine down here. Thanks.
I hope everything is fine there as well.

Indeed, there have been updates.
The thermal tutorial explaining linear and non-linear and static and transient cases is finished.

I'm about to finish the three separate descriptions (READMEs) for engineers, hackers and academics. I also updated the Software Design Specifications to reflect some explanations as requested by the reviewers which are linked from the READMEs.

I plan to finish the updated SDS this weekend.
I'll try to update the paper ASAP after that.
