+++ tests/mms/README.md
@@ -1,3 +1,6 @@
 # Verification with the Method of Manufactured Solution
 
 Check out this video for an explanation about MMS: <https://youtu.be/mGNwL8TGijg>
+
+
+ * [thermal](Verification of the `thermal` problem)
