+++ README4hackers.md
@@ -42,7 +42,15 @@ Unlike these other FEA tools, FeenoX provides...
 
  * a parser for a [syntactically-sugared](https://seamplex.com/feenox/doc/sds.html#syntactic-sugar-highlighting) self-explanatory ASCII file (passed as the first non-optional argument to the `feenox` executable)  with keywords that completely define the problem without requiring further human actions, allowing a [cloud-first](https://seamplex.com/feenox/doc/sds.html#cloud-first) workflow using containerized images (because there is no need to recompile the binary for each problem).
  
- * a Git repository with GPL sources (and FDL documentation) where contributions are welcome. In particular, each partial differential equation that FeenoX can solve correspondence to one of the subdirectories of `src/pdes`. The `autogen.sh` step (prior to `./configure` and `make`) detects the directory structure and includes all the subdirectories it finds as available [problem types](https://www.seamplex.com/feenox/doc/feenox-manual.html#problem). They can be queried at runtime with the `--pdes` option:
+ * a few supported [`PROBLEM`](https://www.seamplex.com/feenox/doc/feenox-manual.html#problem) types and a mechanism to allow hacker and [academics](./README4academics.md) to add new PDEs, as explained in the next bullet. This bullet is about the fact that a [regular user](./README4engineers.md) wanting to solve heat conduction (even with [multi-material non-uniform conductivities $k$](https://www.seamplex.com/feenox/doc/tutorials/320-thermal/)) just needs to do
+ 
+   ```feenox
+   PROBLEM thermal
+   ```
+   
+   and does not need to know nor write the weak form of the Poisson equation in the input file (as other tools ask for).
+ 
+ * a Git repository with GPL sources (and FDL documentation) where contributions are welcome. In particular, each partial differential equation that FeenoX can solve correspondence to one of the subdirectories of `src/pdes` that provide [C entry points that the main mathematical framework calls as function pointer to build the elemental objects](https://seamplex.com/feenox/doc/sds.html#sec:extensibility). The `autogen.sh` step (prior to `./configure` and `make`) detects the directory structure and includes all the subdirectories it finds as available [problem types](https://www.seamplex.com/feenox/doc/feenox-manual.html#problem). They can be queried at runtime with the `--pdes` option:
  
     ```terminal
     $ feenox --pdes
@@ -54,6 +62,8 @@ Unlike these other FEA tools, FeenoX provides...
     thermal
     $ 
     ```
+    
+   The decision of extensibility through compiled code is, as the choice of making FeenoX a program and not a library, a thoughtful one. See [FeenoX for academics](./README4academics.md) for more details about how the extensibility mechanism works.
     
  * continuous integration (using Github actions), an issue tracker (using Github issues) and a discussion page (using Github discussions)
   
@@ -90,22 +100,22 @@ Unlike these other FEA tools, FeenoX provides...
  
 The input file...
 
- - has one-to-one correspondence with the human description of the problem
- - is Git-traceable (the mesh is defined in a separate file created by Gmsh, which may or may not be tracked)
- - allows the user to enter algebraic expressions whenever a numerical value is needed (everything is an expression)
- - understands definitions (nouns) and instructions (verbs). FeenoX has an actual instruction pointer that loops over the instruction set (there might even be conditional blocks).
- - is simple for simple files (but might get more complicated for mor complex problems). Remember Alan Kay's quote: "simple things should be simple and complex things should be possible."
[…]

[…]
