+++ src/version.c
@@ -1,7 +1,7 @@
 /*------------ -------------- -------- --- ----- ---   --       -            -
  *  FeenoX version and description routines
  *
- *  Copyright (C) 2009--2023 jeremy theler
+ *  Copyright (C) 2009--2024 jeremy theler
  *
  *  This file is part of FeenoX.
  *
@@ -145,7 +145,7 @@ void feenox_copyright(void) {
   printf("Copyright © %d--%d Seamplex, https://seamplex.com/feenox\n\
 GNU General Public License v3+, https://www.gnu.org/licenses/gpl.html. \n\
 FeenoX is free software: you are free to change and redistribute it.\n\
-There is NO WARRANTY, to the extent permitted by law.\n", 2009, 2023); 
+There is NO WARRANTY, to the extent permitted by law.\n", 2009, 2024); 
 }
 
 void feenox_longversion(void) {
