+++ doc/sds.md
@@ -354,8 +354,17 @@ The program can also be executed remotely either
  c. in a [container](https://en.wikipedia.org/wiki/OS-level_virtualization) as part of a provisioning script,
  d. in many other ways.
 
+As explained in the help message, FeenoX can read the input from the standard input if `-` is specified as the input path.
+This is useful in scripts where small calculations are needed, e.g.
+
+```terminal
+$ a=3
+$ echo "PRINT 1/$a" | feenox -
+0.333333
+$ 
+```
  
-FeenoX provides mechanisms to inform its progress by writing certain information to devices or files, which in turn can be monitored remotely or even trigger server actions. Progress can be as simple as an ASCII bar (triggered with `--progress` in the command line or with the keyword `PROGRESS` in the input file) to more complex mechanisms like writing the status in a shared memory segment.
+FeenoX provides mechanisms to inform its progress by writing certain information to devices or files, which in turn can be monitored remotely or even trigger server actions. Progress can be as simple as an ASCII bar (triggered with `--progress` in the command line or with the keyword [`PROGRESS`](https://www.seamplex.com/feenox/doc/feenox-manual.html#problem) in the input file) to more complex mechanisms like writing the status in a shared memory segment.
 @Fig:caeplex-progress shows how the [CAEplex](https://www.caeplex.com) platform shows the progress interactively in its web-based interface.
 
 ![ASCII progress bars parsed and converted into a web-based interface](caeplex-progress.png){#fig:caeplex-progress width_latex=65% width_html=100%}
@@ -1465,9 +1474,19 @@ $
 > ```
 
 FeenoX currently works by reading an input file (which in turn can recursively [`INCLUDE`](https://www.seamplex.com/feenox/doc/feenox-manual.html#include) further input files) with an _ad-hoc_ format, whose rationale is described in this section.
-Therefore, it already does satisfy requirement\ a. but eventually could also satisfy requirement\ b.
+Therefore, it already does satisfy requirement\ a. but, eventually, could also satisfy requirement\ b. by adding a wrapper for high-level languages such as 
 
-As already explained in\ @sec:introduction, the motto is “FeenoX is---in a certain sense---to desktop FEA programs  and libraries what Markdown is to Word and (La)TeX, respectively and _deliberately_.” Hence, the input files act as the Markdown source: instructions about what to do but not how to do it.
+ * Python
+ * Julia
+ * R
+
+that would either
+
+ i. create an input file and run FeenoX in the back, or
+ ii. successively call the FeenoX functions that define definitions and execute instructions (to be done).
+ 
+As already explained in\ @sec:introduction, the motto is “FeenoX is---in a certain sense---to desktop FEA programs  and libraries what Markdown is to Word and (La)TeX, respectively and _deliberately_.”
+Hence, the input files act as the Markdown source: instructions about what to do but not how to do it.
 
 The input files are indeed plain-text ASCII files with English-like keywords that fully define the problem.
 The main features of the input format, thoroughly described below, are:
@@ -1482,28 +1501,20 @@ The main features of the input format, thoroughly described below, are:
  #. It should be possible to read run-time arguments from the command line.
  #. Input files are [distributed version control](https://en.wikipedia.org/wiki/Distributed_version_control)-friendly.
  
-### Syntactic sugar & highlighting
+### Syntactic sugar & highlighting {#sec:sugar}
+
+The ultimate goal of FeenoX is to solve mathematical equations that are hard to solve with pencil and paper.
[…]
