+++ README.md
@@ -132,7 +132,7 @@ Browse through the [documentation index](doc) and/or the [`doc`](https://github.
  * [Description](https://www.seamplex.com/feenox/doc/feenox-desc.html)
  * [Software Design Requirements](https://seamplex.com/feenox/doc/srs.html)
  * [Software Design Specifications](https://seamplex.com/feenox/doc/sds.html)
- * [Unix man page](https://seamplex.com/feenox/doc/feenox.1.html) (accesible through `man feenox` after installation)
+ * [Unix man page](https://seamplex.com/feenox/doc/feenox.1.html) (accessible through `man feenox` after installation)
  * [History](./doc/history.md)
  * [Compilation guide](./doc/compilation.md)
  * [Programming guide](./doc/programming.md)
@@ -185,7 +185,7 @@ Follow us: [YouTube](https://www.youtube.com/channel/UCC6SzVLxO8h6j5rLlfCQPhA)
 
 ---------------------------
 
-FeenoX is copyright ©2009-2025 [Seamplex](https://www.seamplex.com)  
+FeenoX is copyright ©2009-2024 [Seamplex](https://www.seamplex.com)  
 FeenoX is licensed under [GNU GPL version 3](http://www.gnu.org/copyleft/gpl.html) or (at your option) any later version.  
 FeenoX is free software: you are free to change and redistribute it.  
 There is **NO WARRANTY**, to the extent permitted by law.  
