+++ README4academics.md
@@ -11,35 +11,39 @@ toc: true
 
 # What
 
-FeenoX is a [cloud-first](https://seamplex.com/feenox/doc/sds.html#cloud-first) [Unix](./README4hackers.md) stand-alone program (i.e. something your run, not something you have to link existing code against) that reads an [engineering-related problem](./README4engineers.md) in a plain-text file containing definitions and instructions. This program then solves the problem and, eventually, writes the outputs which the input file requests with explicit instructions (and nothing if nothing is asked for) and returns back to the Unix shell.
+FeenoX is a [cloud-first](https://seamplex.com/feenox/doc/sds.html#cloud-first) [Unix](./README4hackers.md) stand-alone program (i.e. something your run, not something you have to link existing code against) that reads an [engineering-related problem](./README4engineers.md) in a [plain-text file containing definitions and instructions](https://seamplex.com/feenox/doc/sds.html#sec:input).
+This program then solves the problem and, eventually, writes the outputs which the input file requests with explicit instructions (and nothing if nothing is asked for) and returns back to the Unix shell.
 It can be seen as a Unix filter (or as a transfer function)
 
 ```include
 doc/transfer.md
 ```
 
-which, when zoomed in, acts as a "glue layer" between a mesher (Gmsh) and a library for solving large sparse problems (PETSc) which for linear elastic looks as follows:
+which, when zoomed in, acts as a "glue layer" between a mesher ([Gmsh](http://gmsh.info/)) and a library for solving large sparse problems ([PETSc](https://petsc.org/release/)) which for linear elastic looks as follows:
 
 ![](doc/transfer-le10-zoom.svg)\  
 
 
 
-The design responds to a [Software Requirement Specifications](https://www.seamplex.com/feenox/doc/srs.html) that acts as a "request for quotations" of a computational engineering tool that should satisfy some fictitious (but plausible) requirements.
-The [Software Design Specifications](https://www.seamplex.com/feenox/doc/sds.html) explains how FeenoX addresses each requirement of the SRS.
+The design responds to a [Software Requirement Specifications](https://www.seamplex.com/feenox/doc/srs.html) documente that acts as a "request for quotations" of a computational engineering tool that should satisfy some fictitious (but plausible) requirements.
+The [Software Design Specifications](https://www.seamplex.com/feenox/doc/sds.html) document explains how FeenoX addresses each requirement of the SRS.
 
-In principle, even though FeenoX can solve [generic numerical problems](https://seamplex.com/feenox/examples/examples/basic.html) and [systems of ordinary differential/algebraic eq1uations](https://seamplex.com/feenox/examples/daes.html), its main focus is on solving partial differential equations using the finite element method---eventually [in parallel using the MPI standard](https://seamplex.com/feenox/doc/sds.html#sec:scalability).
+In principle, even though FeenoX can solve [generic numerical problems](https://seamplex.com/feenox/examples/examples/basic.html) and [systems of ordinary differential/algebraic equations](https://seamplex.com/feenox/examples/daes.html), its main objective is to solve partial differential equations using the finite element method---eventually [in parallel using the MPI standard](https://seamplex.com/feenox/doc/sds.html#sec:scalability).
 The current version can solve
 
 ```include
 doc/examples-list.md
 ```
 
-Two of the supported PDEs are related to neutron transport.
-There is a PhD thesis (in Spanish only) that discusses the design and implementation of FeenoX in view of core-level neutronics.
[…]
