+++ README.md
@@ -36,6 +36,12 @@ Choose your background for further details about the what, how and whys:
 doc/examples-list.md
 ```
 
+```{=html}
+<div id="cast-le10"></div>
+<script>AsciinemaPlayer.create('doc/le10.cast', document.getElementById('cast-le10'), {cols:133,rows:32, poster: 'npt:0:3'});</script>
+```
+
+
 ### Tutorials
 
 ```include
@@ -43,44 +49,45 @@ doc/tutorials-list.md
 ```
 
 
-::::: {.only-in-format .html }
-
-::: {.container .text-center .my-5 .ratio .ratio-16x9}
-<iframe class="embed-responsive-item" src="https://www.youtube.com/embed/b3K2QewI8jE?rel=0" allowfullscreen></iframe>
-:::
-
-:::::
-    
-    
 ### Tests
 
-```include
-doc/tests-list.md
-```
+The [`tests`](https://github.com/seamplex/feenox/tree/main/tests) directory in the repository has hundreds of
+ 
+ - `grep`-able examples
+ - [unit and regression tests](https://seamplex.com/feenox/doc/compilation.html#sec:test-suite), 
+ - (crude) mathematical & code verification tests (as in "are we solving right the equations?"),
+ - subdirectories with further case studies
+    - [`mms`](https://github.com/seamplex/feenox/tree/main/tests/mms)
+      - [`thermal`](https://github.com/seamplex/feenox/tree/main/tests/mms/thermal)
+        - [`2d`](https://github.com/seamplex/feenox/tree/main/tests/mms/thermal/2d)
+        - [`3d`](https://github.com/seamplex/feenox/tree/main/tests/mms/thermal/3d)
+    - [`nafems`](https://github.com/seamplex/feenox/tree/main/tests/nafems)
+      - [`le10`](https://github.com/seamplex/feenox/tree/main/tests/nafems/le10)
  
 ## Capabilities
 
+ * Both _free_ as in "free speech" and in "free beer"
  * The [problem](https://www.seamplex.com/feenox/doc/feenox-manual.html#problem) to solve is defined through a [syntactically-sugared](https://seamplex.com/feenox/doc/sds.html#sec:syntactic) [self-descriptive English-like plain-text input file](https://seamplex.com/feenox/doc/sds.html#sec:input) that should [resemble the original human-friendly problem formulation](https://seamplex.com/feenox/doc/sds.html#sec:matching-formulations) as much as possible:
    - [no need to recompile if the problem changes](https://seamplex.com/feenox/doc/sds.html#sec:introduction) (FeenoX is a _program_, not a library)
    - [nouns are definitions and verbs are instructions](https://seamplex.com/feenox/doc/sds.html#sec:nouns_verbs)
    - [simple problems need simple inputs](https://seamplex.com/feenox/doc/sds.html#sec:simple)
    - [everything is an expression](https://seamplex.com/feenox/doc/sds.html#sec:expression)
    - [100%-defined user output](https://seamplex.com/feenox/doc/sds.html#sec:output) (no [`PRINT`](https://www.seamplex.com/feenox/doc/feenox-manual.html#problem) nor [`WRITE_RESULTS`](https://www.seamplex.com/feenox/doc/feenox-manual.html#write_results) instructions, no output)
-   
- * [General mathematical problems](https://seamplex.com/feenox/examples/basic.html) using [GNU GSL](https://www.gnu.org/software/gsl/)
- * [Sets of ODEs/DAEs](https://seamplex.com/feenox/examples/daes.html) using [SUNDIALS](https://computing.llnl.gov/projects/sundials)
[…]
