+++ doc/joss/paper.md
@@ -20,32 +20,98 @@ syntax-definition: feenox.xml
 
 # Summary
 
-[FeenoX](https://www.seamplex.com/feenox) is a computational tool to solve (currently)
+[FeenoX](https://seamplex.com/feenox) is a
+[cloud-first](https://seamplex.com/feenox/doc/sds.html#cloud-first)
+[free](https://www.seamplex.com/feenox/#licensing)
+[no-X](https://www.seamplex.com/feenox/README4engineers.html#how-feenox-works)
+[uniX-like](https://www.seamplex.com/feenox/doc/sds.html#sec:unix)
+finite-element(ish) computational engineering tool designed to solve
+[engineering-related problems](https://www.seamplex.com/feenox/#extents) using
+[cloud servers in parallel](https://www.seamplex.com/feenox/doc/sds.html#sec:scalability) in such a way that
+[the problem](https://www.seamplex.com/feenox/doc/sds.html#sec:scope) is defined in a plain-text near-English self descriptive
+[input file](https://www.seamplex.com/feenox/doc/sds.html#sec:input) read at
+[run time](https://www.seamplex.com/feenox/doc/sds.html#sec:architecture),
+[without requiring further user intervention](https://www.seamplex.com/feenox/doc/sds.html#cloud-first) after
+[the invocation](https://www.seamplex.com/feenox/doc/feenox-manual.html#running-feenox).
+FeenoX meets a fictitious-yet-plausible [Software Requirement Specifications](https://www.seamplex.com/feenox/doc/srs.html).
+The FeenoX [Software Design Specifications](https://www.seamplex.com/feenox/doc/sds.html) address each requirement of the SRS.
+FeenoX provides a set of common
+[extents](https://www.seamplex.com/feenox/#extents),
+[capabilities](https://www.seamplex.com/feenox/#capabilities) and
+[usefulness](https://www.seamplex.com/feenox/#usefulness) but offers different features (following slightly different spirits) for
+[industry engineers](https://www.seamplex.com/feenox/README4engineers.html),
+[Unix hackers](https://www.seamplex.com/feenox/README4hackers.html) and
+[academic researchers](https://www.seamplex.com/feenox/README4academics.html).
+The main features of this [design basis](https://seamplex.com/feenox/doc/sds.html) are
 
- * [dynamical systems written as sets of differential-algebraic equations](https://seamplex.com/feenox/examples/daes.html), or
- * [steady or transient heat conduction problems](https://seamplex.com/feenox/examples/thermal.html), or
- * [steady or quasi-static thermo-mechanical problems](https://seamplex.com/feenox/examples/mechanical.html), or
- * [modal analysis problems](https://seamplex.com/feenox/examples/modal.html), or
- * [steady-state neutronics](https://seamplex.com/feenox/examples/neutron_diffusion.html)
- 
-in such a way that the input is a near-English text file that defines the problem.
-FeenoX is a cloud-first tool to solve engineering problems, with a particular design basis.
-Two of the main features of this [design basis](https://seamplex.com/feenox/doc/sds.html) are
-
- #. Simple problems ought to have simple inputs.
- #. There should be a one-to-one correspondence between the problem definition and FeenoX's input file (fig. \ref{le10}).
+ * The tool has to be an [already-compiled program](https://www.seamplex.com/feenox/doc/sds.html#sec:architecture) (not a library) so regular users do not have to compile anything to solve a problem.
+ * Simple problems ought to need [simple input files](https://www.seamplex.com/feenox/doc/sds.html#sec:simple).
[…]
