### 2024-01-21 852eadc
minor fixes to the joss paper
