+++ doc/joss/paper.md
@@ -33,7 +33,7 @@ finite-element(ish) computational engineering tool designed to solve
 [run time](https://www.seamplex.com/feenox/doc/sds.html#sec:architecture),
 [without requiring further user intervention](https://www.seamplex.com/feenox/doc/sds.html#cloud-first) after
 [the invocation](https://www.seamplex.com/feenox/doc/feenox-manual.html#running-feenox).
-FeenoX meets a fictitious-yet-plausible [Software Requirement Specifications](https://www.seamplex.com/feenox/doc/srs.html).
+FeenoX meets fictitious-yet-plausible [Software Requirement Specifications](https://www.seamplex.com/feenox/doc/srs.html) (SRS).
 The FeenoX [Software Design Specifications](https://www.seamplex.com/feenox/doc/sds.html) address each requirement of the SRS.
 FeenoX provides a set of common
 [extents](https://www.seamplex.com/feenox/#extents),
@@ -46,7 +46,7 @@ The main features of this [design basis](https://seamplex.com/feenox/doc/sds.htm
 
  * The tool has to be an [already-compiled program](https://www.seamplex.com/feenox/doc/sds.html#sec:architecture) (not a library) so regular users do not have to compile anything to solve a problem.
  * Simple problems ought to need [simple input files](https://www.seamplex.com/feenox/doc/sds.html#sec:simple).
- * There should be a [one-to-one correspondence](https://www.seamplex.com/feenox/doc/sds.html#sec:matching-formulations) between the problem definition and FeenoX's input file (illustrated in fig. \ref{le10}).
+ * There should be a [one-to-one correspondence](https://www.seamplex.com/feenox/doc/sds.html#sec:matching-formulations) between the problem definition and FeenoX's input file, as illustrated in fig. \ref{le10}.
  * There should be an [extension mechanism](https://www.seamplex.com/feenox/doc/sds.html#sec:extensibility) to allow hackers and researchers to add new partial differential equations to the tool.
  
 ![The NAFEMS LE10 problem statement [@national1990nafems] and the corresponding FeenoX input illustrating the one-to-one correspondence between the two.\label{le10}](nafems-le10-problem-input.svg)
@@ -80,28 +80,28 @@ First, it can solve
 
 Second, it is the only free and open-source tool that satisfies the [Software Requirement Specifications](https://www.seamplex.com/feenox/doc/srs.html), including that...
 
- * in order to solve a problem one needs to prepare a (relatively) [simple input file](https://seamplex.com/feenox/doc/sds.html#sec:input) (not a script) which is [read at run-time](https://seamplex.com/feenox/doc/sds.html#sec:execution) (not code which calls a library).
+ * in order to solve a problem one needs to prepare a (relatively) [simple input file](https://seamplex.com/feenox/doc/sds.html#sec:input) (not a script nor a deck) which is [read at run-time](https://seamplex.com/feenox/doc/sds.html#sec:execution) (not code which calls a library).
  For example, considering the [NAFEMS LE10 Benchmark problem](https://seamplex.com/feenox/examples/mechanical.html#nafems-le10-thick-plate-pressure-benchmark) from fig. \ref{le10}, FeenoX works as two ["glue layers,"](https://www.linuxtopia.org/online_books/programming_books/art_of_unix_programming/ch04s03_1.html)
 
    1. between the mesher [Gmsh](http://gmsh.info/) and the [PETSc library](https://petsc.org/release/) [@petsc-user-ref;@petsc-efficient]
-   2. between the [PETSc library](https://petsc.org/release/) and the post-processor [Paraview](https://www.paraview.org/)
+   2. between the [PETSc library](https://petsc.org/release/) and a post-processor such as [Paraview](https://www.paraview.org/)
 
    ![](transfer-le10-zoom.svg)\  
    
[…]
