+++ dist/petsc.sh
@@ -26,7 +26,6 @@ export PETSC_ARCH=${PETSC_ARCH}
 
 # TODO: neither mpi nor mumps work in static
 ./configure \
-  --with-mpi=0 \
   --with-cxx=0 \
   --with-fortran-bindings=0 \
   --with-c2html=0 \
@@ -38,7 +37,10 @@ export PETSC_ARCH=${PETSC_ARCH}
   --download-mpich \
   --COPTFLAGS="-O3" --FOPTFLAGS="-O3"
            
+#   --with-mpi=0 \
+#   --download-mpich \
 
+           
 # export PETSC_ARCH=arch-linux2-serial-mumps-static
 # ./configure --with-mpi=0 --with-cxx=0 --with-fortran-bindings=0 --with-fc=0 --with-c2html=0 --with-x=0 --with-debugging=0 --with-shared-libraries=0 --with-mumps-serial=1 --download-mumps --download-openblas --COPTFLAGS="-O" --FOPTFLAGS="-O3"
 # OJO! hay que sacar a mano el -lgcc_s de ${PETSC_DIR}/${PETSC_ARCH}/lib/petsc/conf/petscvariables
