### 2024-02-09 b0d7c30
doc in bin dist is inside a subdir named feenox
