### 2024-02-12 #5846 @vijaysm
@gtheler Thanks for making all the requested changes. The manuscript reads much better and provides detailed information about the capabilities and how FeenoX differentiates itself from other libraries. My one final comment regarding the paper is that it contains too many links to FeenoX documentation pages. While this is good in general, several of these seem redundant and repetitive. I would suggest linking to documentation only when important as I feel it reduces the readability of the paper.

I reviewed the nonlinear examples that you added as well. I'm satisfied with those changes.

@Kevin-Mattheus-Moerman I am happy with the changes made and recommend the submission for acceptance.
