### 2024-02-12 #5846 @gtheler
> @gtheler Thanks for making all the requested changes. The manuscript reads much better and provides detailed information about the capabilities and how FeenoX differentiates itself from other libraries. My one final comment regarding the paper is that it contains too many links to FeenoX documentation pages. While this is good in general, several of these seem redundant and repetitive. I would suggest linking to documentation only when important as I feel it reduces the readability of the paper.

I cannot agree more. I overshot the link count based on the previous reviews.
It would be very helpful if you could point out a few links should be removed and a few which should stay.
