### 2024-03-03 20c0c23
removed duplicated links in JOSS paper
