+++ doc/joss/paper.md
@@ -22,32 +22,32 @@ syntax-definition: feenox.xml
 
 [FeenoX](https://seamplex.com/feenox) is a
 [cloud-first](https://seamplex.com/feenox/doc/sds.html#cloud-first)
-[free](https://www.seamplex.com/feenox/#licensing)
-[no-X](https://www.seamplex.com/feenox/README4engineers.html#how-feenox-works)
-[uniX-like](https://www.seamplex.com/feenox/doc/sds.html#sec:unix)
+[free](https://seamplex.com/feenox/#licensing)
+[no-X](https://seamplex.com/feenox/README4engineers.html#how-feenox-works)
+[uniX-like](https://seamplex.com/feenox/doc/sds.html#sec:unix)
 finite-element(ish) computational engineering tool designed to solve
-[engineering-related problems](https://www.seamplex.com/feenox/#extents) using
-[cloud servers in parallel](https://www.seamplex.com/feenox/doc/sds.html#sec:scalability) in such a way that
-[the problem](https://www.seamplex.com/feenox/doc/sds.html#sec:scope) is defined in a plain-text near-English self descriptive
-[input file](https://www.seamplex.com/feenox/doc/sds.html#sec:input) read at
-[run time](https://www.seamplex.com/feenox/doc/sds.html#sec:architecture),
-[without requiring further user intervention](https://www.seamplex.com/feenox/doc/sds.html#cloud-first) after
-[the invocation](https://www.seamplex.com/feenox/doc/feenox-manual.html#running-feenox).
-FeenoX meets fictitious-yet-plausible [Software Requirement Specifications](https://www.seamplex.com/feenox/doc/srs.html) (SRS).
-The FeenoX [Software Design Specifications](https://www.seamplex.com/feenox/doc/sds.html) address each requirement of the SRS.
+engineering-related problems using
+[cloud servers in parallel](https://seamplex.com/feenox/doc/sds.html#sec:scalability) in such a way that
+[the problem](https://seamplex.com/feenox/doc/sds.html#sec:scope) is defined in a plain-text near-English self descriptive
+[input file](https://seamplex.com/feenox/doc/sds.html#sec:input) read at
+[run time](https://seamplex.com/feenox/doc/sds.html#sec:architecture),
+without requiring further user intervention after
+[the invocation](https://seamplex.com/feenox/doc/feenox-manual.html#running-feenox).
+FeenoX meets fictitious-yet-plausible [Software Requirement Specifications](https://seamplex.com/feenox/doc/srs.html) (SRS).
+The FeenoX [Software Design Specifications](https://seamplex.com/feenox/doc/sds.html) address each requirement of the SRS.
 FeenoX provides a set of common
-[extents](https://www.seamplex.com/feenox/#extents),
-[capabilities](https://www.seamplex.com/feenox/#capabilities) and
-[usefulness](https://www.seamplex.com/feenox/#usefulness) but offers different features (following slightly different spirits) for
-[industry engineers](https://www.seamplex.com/feenox/README4engineers.html),
-[Unix hackers](https://www.seamplex.com/feenox/README4hackers.html) and
-[academic researchers](https://www.seamplex.com/feenox/README4academics.html).
-The main features of this [design basis](https://seamplex.com/feenox/doc/sds.html) are
+[extents](https://seamplex.com/feenox/#extents),
+[capabilities](https://seamplex.com/feenox/#capabilities) and
+[usefulness](https://seamplex.com/feenox/#usefulness) but offers different features (following slightly different spirits) for
+[industry engineers](https://seamplex.com/feenox/README4engineers.html),
+[Unix hackers](https://seamplex.com/feenox/README4hackers.html) and
+[academic researchers](https://seamplex.com/feenox/README4academics.html).
+The main features of this design basis are
 
[…]
