### 2024-03-03 #5846 @gtheler
> > It would be very helpful if you could point out a few links should be removed and a few which should stay.
>
> As a general rule, I would keep the first reference to any webpage/documentation and leave out further links to the page unless it is **absolutely** necessary and adds more context than what the text already details.

@vijaysm can you confirm if this commit removing the duplicate links improves the readability enough?
https://github.com/seamplex/feenox/commit/20c0c23432908b73a9292b646181999c62fe25f7
