### 2024-03-03 #5846 @chennachaos
Thanks for updating the repo, @gtheler! The documentation looks better now.

* This link to [Basic mathematics](https://seamplex.com/feenox/examples/examples/basic.html) at https://seamplex.com/feenox/ is broken.
* A couple of broken links in **Contributing** section on GitHub page.
* The software paper has far too many links. Suggest removing them at least in the summary.
* For software libraries like FEniCS, MoFEM, PETSc, etc., I suggest adding references with appropriate referencing style instead of as links to websites.
