### 2024-03-04 c047739
fixed broken link to examples/basic
