+++ doc/joss/paper.md
@@ -68,7 +68,7 @@ Open-source finite-element tools are either
 FeenoX sits in the middle.
 First, it can solve
 
- - [Basic mathematics](https://seamplex.com/feenox/examples/examples/basic.html)
+ - [Basic mathematics](https://seamplex.com/feenox/examples/basic.html)
  - [Systems of ODEs/DAEs](https://seamplex.com/feenox/examples/daes.html)
  - [Laplace’s equation](https://seamplex.com/feenox/examples/laplace.html)
  - [Heat conduction](https://seamplex.com/feenox/examples/thermal.html)
