### 2024-03-05 303d7f0
removed links from joss paper and added references
