+++ doc/joss/paper.bib
@@ -81,3 +81,72 @@ articleno = {9},
 numpages = {37},
 keywords = {symbolic differentiation, functional, partial differential equation, AD, index notation, automatic functional differentiation, tensor algebra, embedded language, PDE, DSEL, DSL, mixed element, weak formulation, domain specific language, discretization, weak form, finite element method, algorithmic differentiation, FEM, einstein notation, implicit summation}
 }
+
+@phdthesis{sparselizard,
+  author       = {Alexandre Halbach},
+  title        = {Sparselizard---the user friendly finite element c++ library},
+  school       = {Université de Liège---Dép. d'électric., électron. et informat. (Inst.Montefiore)},
+  year         = 2017,
+}
+
+
+@article{mofem,
+  title = {{MoFEM:} An open source, parallel finite element library},
+  author = {
+    Kaczmarczyk, {\L}ukasz and Ullah, Zahur and Lewandowski, Karol and Meng,
+    Xuan and    Zhou, Xiao-Yi and Athanasiadis, Ignatios and Nguyen,
+    Hoang  and  Chalons-Mouriesse, Christophe-Alexandre and Richardson,
+    Euan and Miur, Euan and Shvarts, Andrei and Wakeni,Mebratu and Pearce, Chris},
+  journal={The Journal of Open Source Software},
+  year={2020},
+  note={http://mofem.eng.gla.ac.uk},
+  doi={10.21105/joss.01441},
+  url={https://joss.theoj.org/papers/10.21105/joss.01441}
+}
+
+@article{mfem,
+  title = {{MFEM}: A Modular Finite Element Methods Library},
+  author = {R. Anderson and J. Andrej and A. Barker and J. Bramwell and J.-S. Camier and
+            J. Cerveny and V. Dobrev and Y. Dudouit and A. Fisher and Tz. Kolev and W. Pazner and
+            M. Stowell and V. Tomov and I. Akkerman and J. Dahm and D. Medina and S. Zampini},
+  journal = {Computers \& Mathematics with Applications},
+  doi = {10.1016/j.camwa.2020.06.009},
+  volume = {81},
+  pages = {42---74},
+  year = {2021}
+}
+
+@misc{fenics,
+  title     = {{DOLFINx}: the next generation {FEniCS} problem solving environment},
+  author    = {Baratta, Igor A. and Dean, Joseph P. and Dokken, J{\o}rgen S. and Habera, Michal and Hale, Jack S. and Richardson, Chris N. and Rognes, Marie E. and Scroggs, Matthew W. and Sime, Nathan and Wells, Garth N.},
+  doi       = {10.5281/zenodo.10447666},
+  year      = {2023},
+  howpublished = {preprint}
+}
+
+
+@article{gmsh,
+  title = {{Gmsh}: a three-dimensional finite element mesh generator with built-in pre- and post-processing facilities},
+  author = {C. Geuzaine and J.-F. Remacle},
+  journal = {International Journal for Numerical Methods in Engineering},
+  doi = {10.1016/j.camwa.2020.06.009},
+  volume = {79},
+  number = {11},
+  pages = {1309---1331},
+  year = {2009}
+}
+
+@book{paraview,
+  title={The ParaView Guide: A Parallel Visualization Application},
+  author={Utkarsh Ayachit},
+  year={2015},
+  publisher={Kitware},
+  isbn={9781930934306}
+}
+
+@book{raymond,
+ title = "The Art of Unix Programming",
+ author = "Eric S. Raymond",
+ publisher = "Addison-Wesley",
+ year="2003",
+}
