+++ doc/joss/paper.md
@@ -20,34 +20,34 @@ syntax-definition: feenox.xml
 
 # Summary
 
-[FeenoX](https://seamplex.com/feenox) is a
-[cloud-first](https://seamplex.com/feenox/doc/sds.html#cloud-first)
-[free](https://seamplex.com/feenox/#licensing)
-[no-X](https://seamplex.com/feenox/README4engineers.html#how-feenox-works)
-[uniX-like](https://seamplex.com/feenox/doc/sds.html#sec:unix)
+FeenoX is a
+cloud-first
+free
+no-X
+uniX-like
 finite-element(ish) computational engineering tool designed to solve
 engineering-related problems using
-[cloud servers in parallel](https://seamplex.com/feenox/doc/sds.html#sec:scalability) in such a way that
-[the problem](https://seamplex.com/feenox/doc/sds.html#sec:scope) is defined in a plain-text near-English self descriptive
-[input file](https://seamplex.com/feenox/doc/sds.html#sec:input) read at
-[run time](https://seamplex.com/feenox/doc/sds.html#sec:architecture),
+cloud servers in parallel in such a way that
+the problem is defined in a plain-text near-English self descriptive
+input file read at
+run time,
 without requiring further user intervention after
-[the invocation](https://seamplex.com/feenox/doc/feenox-manual.html#running-feenox).
-FeenoX meets fictitious-yet-plausible [Software Requirement Specifications](https://seamplex.com/feenox/doc/srs.html) (SRS).
-The FeenoX [Software Design Specifications](https://seamplex.com/feenox/doc/sds.html) address each requirement of the SRS.
+the invocation.
+FeenoX meets fictitious-yet-plausible Software Requirement Specifications (SRS).
+The FeenoX Software Design Specifications address each requirement of the SRS.
 FeenoX provides a set of common
-[extents](https://seamplex.com/feenox/#extents),
-[capabilities](https://seamplex.com/feenox/#capabilities) and
-[usefulness](https://seamplex.com/feenox/#usefulness) but offers different features (following slightly different spirits) for
-[industry engineers](https://seamplex.com/feenox/README4engineers.html),
-[Unix hackers](https://seamplex.com/feenox/README4hackers.html) and
-[academic researchers](https://seamplex.com/feenox/README4academics.html).
+extents,
+capabilities and
+usefulness but offers different features (following slightly different spirits) for
+industry engineers,
+Unix hackers and
+academic researchers.
 The main features of this design basis are
 
  * The tool has to be an already-compiled program (not a library) so regular users do not have to compile anything to solve a problem.
- * Simple problems ought to need [simple input files](https://seamplex.com/feenox/doc/sds.html#sec:simple).
- * There should be a [one-to-one correspondence](https://seamplex.com/feenox/doc/sds.html#sec:matching-formulations) between the problem definition and FeenoX's input file, as illustrated in fig. \ref{le10}.
- * There should be an [extension mechanism](https://seamplex.com/feenox/doc/sds.html#sec:extensibility) to allow hackers and researchers to add new partial differential equations to the tool.
+ * Simple problems ought to need simple input files.
+ * There should be a one-to-one correspondence between the problem definition and FeenoX's input file, as illustrated in fig. \ref{le10}.
+ * There should be an extension mechanism to allow hackers and researchers to add new partial differential equations to the tool.
  
 ![The NAFEMS LE10 problem statement [@national1990nafems] and the corresponding FeenoX input illustrating the one-to-one correspondence between the two.\label{le10}](nafems-le10-problem-input.svg)
 
@@ -57,10 +57,10 @@ The main features of this design basis are
[…]
