+++ doc/joss/paper.bib
@@ -145,7 +145,7 @@ keywords = {symbolic differentiation, functional, partial differential equation,
 }
 
 @book{raymond,
- title = "The Art of Unix Programming",
+ title = "The Art of {Unix} Programming",
  author = "Eric S. Raymond",
  publisher = "Addison-Wesley",
  year="2003",
