### 2024-03-06 #5846 @gtheler
Dear @chennachaos , thank you very much for your time.

>   * This link to [Basic mathematics](https://seamplex.com/feenox/examples/examples/basic.html) at https://seamplex.com/feenox/ is broken.

Good catch.
Should be fixed here https://github.com/seamplex/feenox/commit/c047739f23109b0eea94949ae7a7c88a5d7b6f26

Note that that commit fixes the markdown source. The web site is generated from these sources but I am on vacations now. I will re-compile the website next week.

>   * A couple of broken links in **Contributing** section on GitHub page.

Another good catch: https://github.com/seamplex/feenox/commit/f2b2b93bf2c0094783cf2aba52b2ebec859444e0

>   * The software paper has far too many links. Suggest removing them at least in the summary.
>   * For software libraries like FEniCS, MoFEM, PETSc, etc., I suggest adding references with appropriate referencing style instead of as links to websites.

I removed a lot of links in the latest version and cited academic software as bibtex citations.
Can you please verify if the latest version fulfills the JOSS expectations?
