+++ doc/help-options-base.txt
@@ -1,6 +1,7 @@
   -h, --help         display options and detailed explanations of commmand-line usage
   -v, --version      display brief version information and exit
   -V, --versions     display detailed version information
+  -c, --check        validates if the input file is sane or not
   --pdes             list the types of PROBLEMs that FeenoX can solve, one per line
   --elements_info    output a document with information about the supported element types
   --linear           force FeenoX to solve the PDE problem as linear
