### 2024-03-13 #5846 @Kevin-Mattheus-Moerman
@gtheler congratulations, after the 3 positive reviews we can now process the final steps towards acceptance in JOSS. Can  you look at the above post-review action points :point_up: , let me know if you have any questions.
