+++ doc/joss/paper.bib
@@ -84,7 +84,7 @@ keywords = {symbolic differentiation, functional, partial differential equation,
 
 @phdthesis{sparselizard,
   author       = {Alexandre Halbach},
-  title        = {Sparselizard---the user friendly finite element c++ library},
+  title        = {Sparselizard---the user friendly finite element {C++} library},
   school       = {Université de Liège---Dép. d'électric., électron. et informat. (Inst.Montefiore)},
   year         = 2017,
 }
@@ -144,8 +144,20 @@ keywords = {symbolic differentiation, functional, partial differential equation,
   isbn={9781930934306}
 }
 
+
+@book{10.5555/2789330,
+author = {Ayachit, Utkarsh},
+title = {The ParaView Guide: A Parallel Visualization Application},
+year = {2015},
+isbn = {1930934300},
+publisher = {Kitware, Inc.},
+address = {Clifton Park, NY, USA},
+abstract = {ParaView is an open-source, multi-platform data analysis and visualization application. ParaView users can quickly build visualizations to analyze their data using qualitative and quantitative techniques. The data exploration can be done interactively in 3D or programmatically using ParaViews batch processing capabilities. ParaView was developed to analyze extremely large datasets using distributed memory computing resources. It can be run on supercomputers to analyze datasets of petascale size as well as on laptops for smaller data, has become an integral tool in many national laboratories, universities and industry, and has won several awards related to high performance computation. ParaView is open-source (BSD licensed, commercial software friendly). As with any of successful open-source project, ParaView is supported by an active user and developer community.}
+}
+
+
 @book{raymond,
- title = "The Art of {Unix} Programming",
+ title = "The Art of {UNIX} Programming",
  author = "Eric S. Raymond",
  publisher = "Addison-Wesley",
  year="2003",
