### 2024-03-14 e73e17c
do not add texi in src tarball
