+++ tests/mms/README.md
@@ -3,4 +3,4 @@
 Check out this video for an explanation about MMS: <https://youtu.be/mGNwL8TGijg>
 
 
- * [thermal](Verification of the `thermal` problem)
+ * [Verification of the `thermal` problem](thermal)
