+++ doc/unix.md
@@ -1,20 +1,42 @@
 
+In 1978, Doug McIlroy---the inventor of Unix pipes and one of the founders of the Unix tradition---stated:
+
+ i. Make each program do one thing well. To do a new job, build afresh rather than complicate old programs by adding new features.
+
+ ii. Expect the output of every program to become the input to another, as yet unknown, program. Don't clutter output with extraneous information. Avoid stringently columnar or binary input formats. Don't insist on interactive input.
+
+ iii. Design and build software, even operating systems, to be tried early, ideally within weeks. Don't hesitate to throw away the clumsy parts and rebuild them.
+
+ iv. Use tools in preference to unskilled help to lighten a programming task, even if you have to detour to build the tools and expect to throw some of them out after you've finished using them.
+
+He later summarized it this way:
+
+ > This is the Unix philosophy: Write programs that do one thing and do it well. Write programs to work together. Write programs to handle text streams, because that is a universal interface.
+
+ 
+FeenoX explicitly followed the above ideas from scratch, especially the for sentences in bullet ii.
+It is even, like Unix itself, a third-system effect where clumsy parts of previous attempts were thrown away and rebuilt from scratch.
+The following sections explain how each of the seventeen rules was taken into account when designing and implementing FeenoX.
+ 
 # Rule of Modularity {#sec:unix-modularity}
 
 > Developers should build a program out of simple parts connected by well defined interfaces, so problems are local, and parts of the program can be replaced in future versions to support new features. This rule aims to save time on debugging code that is complex, long, and unreadable.
 
 FeenoX is designed to be as lightweight as possible.
-On the one hand, it relies on third-party high-quality libraries to do the heavy mathematical weightlifting
+On the one hand, it relies on third-party high-quality libraries to do the heavy mathematical weightlifting such as
 
-   - GNU Scientific Library
-   - SUNDIALS
-   - PETSc
-   - SLEPc
+ * [GNU Scientific Library](https://www.gnu.org/software/gsl/) for general mathematics,
+ * [SUNDIALS IDA](https://computing.llnl.gov/projects/sundials/ida) for ODEs and DAEs,
+ * [PETSc](https://petsc.org/) for linear, non-linear and transient PDEs, and
+ * [SLEPc](http://slepc.upv.es/) for PDEs involving eigen problems
 
-But also the extensibility feature (@sec:unix-extensibility) of having each PDE in separate directories which can be added or removed at compile time without changing any line of the source code goes into this direction as well.
+because these libraries were written by professional programmers using algorithms designed by professional mathematicians.
+Yet-to-be-discovered improved mathematical schemes and/or coding algorithms can be eventually used by FeenoX by just updating those dependencies, which for sure will keep their well-defined interfaces (because they are programmed by professional programmers).
+
+Moreover, the extensibility feature (@sec:unix-extensibility) of having each PDE in separate directories which can be added or removed at compile time without changing any line of the source code goes into this direction as well.
 Relying of C function pointers allows (in principle) to replace these "virtual" methods with other ones using the same interface.
 
-> Note that our language in general and words in particular shape and limit the way we think.
+> Note that our (human) languages in general and words in particular shape and limit the way we think.
 > Fortran's concept of "modules" is _not_ the same as Unix's concept of "modularity."
 > I wish two different words had been used.
 
@@ -22,8 +44,9 @@ Relying of C function pointers allows (in principle) to replace these "virtual" 
 
 > Developers should write programs as if the most important communication is to the developer who will read and maintain the program, rather than the computer. This rule aims to make code as readable and comprehensible as possible for whoever works on the code in the future.
 
[…]
