### 2024-03-14 ef5312d
fix for texinfo
