### 2024-03-14 d6eb002
do not include texi in dist
