+++ autogen.sh
@@ -61,17 +61,17 @@ for i in tests/*.sh tests/*.fee tests/*.geo tests/*.msh tests/*.dat tests/*.ref 
  echo "AC_CONFIG_LINKS([${i}:${i}])" >> config_links.m4
 done
 
-# do not add doc to src unless they ask for it
-echo -n "creating Makefile.am... "
-if [ "x${1}" = "x--doc" ]; then
-  cp Makefile-doc.am Makefile.am
-else
-  echo "SUBDIRS = src" > Makefile.am
-fi
+# let's revisit this when understanding texinfo figures
+# echo -n "creating Makefile.am... "
+# if [ "x${1}" = "x--doc" ]; then
+#   cp Makefile-doc.am Makefile.am
+# else
+#   echo "SUBDIRS = src" > Makefile.am
+# fi
 
-cat Makefile-base.am >> Makefile.am
-touch doc/feenox-desc.texi
-echo "ok"
+# cat Makefile-base.am >> Makefile.am
+# touch doc/feenox-desc.texi
+# echo "ok"
 
 
 # detect the sources in pdes
