### 2024-03-14 #5846 @gtheler
> @gtheler congratulations, after the 3 positive reviews we can now process the final steps towards acceptance in JOSS. Can you look at the above post-review action points ☝️ , let me know if you have any questions.

* Double check authors and affiliations (including ORCIDs)

Checked

*   Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

To celebrate this milestone, I added tag `v1.0` to the repository: https://github.com/seamplex/feenox/releases/tag/v1.0
The version is thus 1.0.

*  Archive the release on Zenodo/figshare/etc and post the DOI here.

10.5281/zenodo.10819606

https://zenodo.org/records/10819606

* Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Checked

* Make sure that the license listed for the archive is the same as the software license.

Checked
