### 2024-03-15 #5846 @gtheler
> @gtheler please edit the archive listed title to include the `v` in `v1.0`.

Like this?

![image](https://github.com/openjournals/joss-reviews/assets/9962631/ae6d27e1-025f-4994-9b0e-dad0bbc9188d)
