### 2024-03-15 #5846 @Kevin-Mattheus-Moerman
@gtheler no, please make the title match the paper title, then edit the version field to be `v1.0`.

See the thing on the right here, which should end up saying `version v1.0`:
![Screenshot from 2024-03-15 15-26-46](https://github.com/openjournals/joss-reviews/assets/8392709/15e19f2d-10c3-4441-8126-d9f14751f491)
