### 2024-03-16 #5846 @xuanxu
The error is caused by a latex formula (`$_N$`) inside a link text, here:
https://github.com/seamplex/feenox/blob/main/doc/joss/paper.md?plain=1#L78
That's generating an invalid JATS metadata file
