### 2024-03-16 #5846 @gtheler
Hi Juanjo @xuanxu . Thanks for takign a look. First time pandoc (and even old-plain TeX) complains about that syntax.
Fixed it as `$\text{S}_N$`:

https://github.com/seamplex/feenox/commit/a4ff64b248094539f4d778484361a59dbee363ad

I did not write it like that to begin with because I though the `S` would be rendered with a Serif font, but I previewed the PDF with a local INARA and it is rendered correctly as Sans Serif.

@Kevin-Mattheus-Moerman can we take another shot?
