### 2024-03-16 #5846 @xuanxu
@gtheler it's not really pandoc, but the [generated JATS file](https://github.com/openjournals/joss-papers/pull/5137/commits/c7ffc7e65adf80be1302cca020a9915b7be4cd86#diff-7fa6cb9143ddbc39dd7cc90e8f2fdda5285530a97f8391a2037defb49b6a4cffR188-R193) used for depositing that is not liking the formula code inside the link and not passing the validation.
