### 2024-03-16 #5846 @gtheler
ok, I now removed the inline math: https://github.com/seamplex/feenox/commit/353a3a30659ba67fe914be469b69fb500b18fac8
one final attempt @Kevin-Mattheus-Moerman ?
