### 2024-03-18 48c4cf6
added citation for joss paper
