### 2024-03-19 9b98430
added link to joss paper in readme
