+++ README.md
@@ -28,8 +28,10 @@ Choose your background for further details about the what, how and whys:
  - [Academic Professor](./README4academics.md)
 
 See the following paper for a brief summary (and to [cite FeenoX in your work](./doc/FAQ.md#how-should-i-cite-feenox)):
- 
- * Theler, J. (2024). FeenoX: a cloud-first finite-element(ish) computational engineering tool. Journal of Open Source Software, 9(95), 5846. <https://doi.org/10.21105/joss.05846> 
+
+ * Theler J, “FeenoX: a cloud-first finite-element(ish) computational engineering tool,” _Journal of Open Source Software_, vol. 9,
+no. 95, p. 5846, Mar. 2024, doi: [10.21105/joss.05846](https://doi.org/10.21105/joss.05846).
+
 
 ## Extents
 
