+++ doc/sds.md
@@ -1074,7 +1074,7 @@ The following input file computes the natural frequencies of oscillation of a ca
 ```terminal
 $ gmsh -3 wire-hex.geo 
 [...]
-$ $ feenox wire.fee | pandoc  
+$ feenox wire.fee | pandoc  
 <table>
 <caption>Comparison of analytical and numerical frequencies, in Hz</caption>
 <thead>
@@ -2094,8 +2094,8 @@ The output in FeenoX is 100% user defined, i.e. everything that FeenoX writes co
  * [`WRITE_RESULTS`](https://www.seamplex.com/feenox/doc/feenox-manual.html#write_results)
  * [`DUMP`](https://www.seamplex.com/feenox/doc/feenox-manual.html#dump)
 
-In the absence of any of these instructions, FeenoX will not write anything.
-Not in the standard output, not in any other file.
+In the absence of any of these instructions, FeenoX _will not_ write anything.
+Not in the standard output, not in any other file. Nothing (Unix rule of silence, @sec:unix-silence).
 
  Computer | Monthly Rental  | Relative Speed | First Delivery 
 :----------------|:------------------------:|:-----------:|:-----------:
@@ -2130,15 +2130,64 @@ For example, if the input file does not reference the principal stress `sigma1` 
 
 ### Output formats {#sec:output-formats}
 
-JSON/YAML, state of the art open post-processing formats.
-Mobile & web-friendly.
+With the ASCII output to standard output (and other text files) controlled with `PRINT`-like instructions, YAML or JSON outputs can be easily implemented within the input file itself.
+For example,
 
-Common and preferably open-source formats.
+```feenox
+DEFAULT_ARGUMENT_VALUE 1 "hello world"
+phi = (1+sqrt(5))/2 
 
+PRINTF "a: %.3f" 1/3
+PRINT TEXT "phi:" phi SEP " "
+PRINT message: ${1}   SEP " "
+```
 
+would give
 
-yaml/json friendly outputs
-vtk (vtu), gmsh, frd?
+```terminal
+$ feenox yaml.fee | tee test.yaml | yq .
+{
+  "a": 0.333,
+  "phi": 1.61803,
+  "message": "hello world"
+}
+$ cat test.yaml 
+a: 0.333
+phi: 1.61803
+message: hello world
+$ 
+```
+
+Now, JSON is more picky and care with quoted characters is needed:
+
+ 1. Curly brackets `{` and `}` are used for multi-line input in FeenoX so they have to be quoted as `\{` and `\}`.
+ 2. Double quotes `"`  are used to delimit keywords with blanks, so they also have to be quoted `\"` when appearing verbatim in an output token.
+
+```feenox
+DEFAULT_ARGUMENT_VALUE 1 "hello world"
+phi = (1+sqrt(5))/2 
+
+PRINTF "\{ \"a\": %.3f," 1/3
+PRINT  TEXT "\"phi\":" phi ,
+PRINT  "\"message\": \"${1}\" \}"
+```
+
+```terminal
+$ feenox json.fee | jq .
+{
+  "a": 0.333,
+  "phi": 1.61803,
+  "message": "hello world"
+}
+$
+```
+
+In the same sense, in principle any ASCII-based format can be implemented this way.
+Markdown output, which can then be converted to other formats as well (such as LaTeX which can then create professionally-looking tables as in @fig:latex-tables), has been already covered in @sec:interoperability.
+
+Current version can write space and time-dependent distributions into Gmsh's `.msh` and VTK's `.vtk` formats.
+Both of them are open standard and have open-source readers.
[…]
