+++ doc/sds.md
@@ -2418,11 +2418,96 @@ The Github Issues feature at <https://github.com/seamplex/feenox/issues> is used
 > 460-documentation.md
 > ```
 
-it's not compact, but almost!
-Compactness is the property that a design can fit inside a human being's head. A good practical test for compactness is this: Does an experienced user normally need a manual? If not, then the design (or at least the subset of it that covers normal use) is compact.
-unix man page
-markdown + pandoc = html, pdf, texinfo
+According to Eric Raymond's book "The Art of Unix Programming":
 
+> Compactness is the property that a design can fit inside a human being's head. A good practical test for compactness is this: Does an experienced user normally need a manual? If not, then the design (or at least the subset of it that covers normal use) is compact.
+
+Following to 20-80 rule, we could say that FeenoX is compact for 80% of its usage. But the most complex 20% of the cases might need users (even the author) to look up the syntax of the definition and instruciones in the manual page (illustrated in @fig:sds-manpage), which is accessible with `man feenox` after installing  with `make install`:
+
+```terminal
+$ man -k feenox
+feenox (1)           - a cloud-first free no-X uniX-like finite-element(ish) computational engineering tool
+$ man feenox
+$
+```
+
+::: {#fig:sds-manpage layout="[0.925,1.075]"}
+
+![Gnome Terminal](manpage-gnome.png){#fig:sds-manpage1}
+
+![Konsole](manpage.png){#fig:sds-manpage2}
+
+The FeenoX Unix manpage in section 1 when running `man feenox`
+:::
+
+This man page is compiled into troff from a markdown source, which in turn has some sections involving the syntax and reference of the
+
+ * definitinos and instructions
+ * special variables
+ * internal built-in functions and functionals
+ 
+generated by a script that parses the actual source code.
+For instance, the code that parses the `INTEGRATE` function has three-forward-slash comments that tell this script that it has to prepare documentation:
+
+```c
+int feenox_parse_integrate(void) {
+  
+  mesh_integrate_t *mesh_integrate = NULL;
+  feenox_check_alloc(mesh_integrate = calloc(1, sizeof(mesh_integrate_t)));
+  
+///kw_pde+INTEGRATE+usage { <expression> | <function> }
+///kw_pde+INTEGRATE+detail Either an expression or a function of space $x$, $y$ and/or $z$ should be given.
+///kw_pde+INTEGRATE+detail If the integrand is a function, do not include the arguments, i.e. instead of `f(x,y,z)` just write `f`.
+///kw_pde+INTEGRATE+detail The results should be the same but efficiency will be different (faster for pure functions).
+  char *token = feenox_get_next_token(NULL);
+  if ((mesh_integrate->function = feenox_get_function_ptr(token)) == NULL) {
+    feenox_call(feenox_expression_parse(&mesh_integrate->expr, token));
+  }
+  
+  char *name_mesh = NULL;
+  char *name_physical_group = NULL;
+  char *name_result = NULL;
+  
+  while ((token = feenox_get_next_token(NULL)) != NULL) {
+///kw_pde+INTEGRATE+usage [ OVER <physical_group> ]
+///kw_pde+INTEGRATE+detail By default the integration is performed over the highest-dimensional elements of the mesh,
+///kw_pde+INTEGRATE+detail i.e. over the whole volume, area or length for three, two and one-dimensional meshes, respectively.
+///kw_pde+INTEGRATE+detail If the integration is to be carried out over just a physical group, it has to be given in `OVER`.
+
+    if (strcasecmp(token, "OVER") == 0) {
[…]
