+++ doc/programming.md
@@ -14,7 +14,7 @@ The main objective is to comply with the Software Requirements Specification. Wi
  2. FeenoX is Open Source as defined by The Open Source initiative. Stick to open standards, formats, practices, etc. Make sure you understand what the difference between "free software" and "open source" is.
  3. Only after making sure every piece of code meets the two criteria above, the technical and/or efficiency aspects mentioned in the SRS are to be considered.
 
-# Compiling and debuging
+# Compiling and debugging
 
 See the [compilation instructions](compilation.md).
 
@@ -33,7 +33,7 @@ See the [compilation instructions](compilation.md).
 
 > Rule of Simplicity: Design for simplicity; add complexity only where you must.
 
-C++ is a great language when used to solve problems it solves naturally well without being forced to solve it (see [LibreBlackjack](https://seamplex.com/blackjack/)). C++ is a terrible language whenever else. It is messy, overwhelming, complicated, not rubust and hard to debug (especially when using templates and this nightmare called smart pointers). The cons are far more than the pros. There is no need to add complexity.
+C++ is a great language when used to solve problems it solves naturally well without being forced to solve it (see [LibreBlackjack](https://seamplex.com/blackjack/)). C++ is a terrible language whenever else. It is messy, overwhelming, complicated, not robust and hard to debug (especially when using templates and this nightmare called smart pointers). The cons are far more than the pros. There is no need to add complexity.
 
 Fortran is a terrible language. Fortran90+ is a patched FORTRAN\ 77. The assumptions that the language had over the kind of computers it has to run on worked in the early days but are now widely outdated. That's it.
 
@@ -73,7 +73,7 @@ Therefore:
  
    Libraries written in C++ are allowed as long as they provide C wrappers and do not meddle with stuff we do not need nor want.
    
- * Keep Fortran even farther away. It is allowed only for existing libaries dating from ancient well-tested and optimized code, but for nothing else.
+ * Keep Fortran even farther away. It is allowed only for existing libraries dating from ancient well-tested and optimized code, but for nothing else.
  * Modern high-level languages like Python or Julia are targets and not sources of FeenoX.
  * For documentation and comments within the code, American English should be used.
 
@@ -121,7 +121,7 @@ Then create a new C/C++ project File->New Project... choose C/C++ Project with E
    }
    ```
  
- * Do not worry about long lines. Only wrap lines when it makes sense to from a logic point of view not becuase "it looks bad on screen" because “screen” can be anything from a mobile phone to a desktop with many 24" LED monitors.
+ * Do not worry about long lines. Only wrap lines when it makes sense to from a logic point of view not because "it looks bad on screen" because “screen” can be anything from a mobile phone to a desktop with many 24" LED monitors.
  * Make sure you understand and follow the 17 rules of Unix philosophy <https://en.wikipedia.org/wiki/Unix_philosophy>
  
     1. modularity
@@ -151,12 +151,12 @@ Reading and writing mesh/post-processing files work the same way. Each format ha
 ## Memory management
 
  * Check all `malloc()` calls for `NULL`. You can use the `feenox_check_alloc()/feenox_check_alloc_null()` macros.
- * Use the macro `feenox_free()` instead of plain `free()`. The former explictly makes the pointer equal to `NULL` after freeing it, which is handy.
+ * Use the macro `feenox_free()` instead of plain `free()`. The former explicitly makes the pointer equal to `NULL` after freeing it, which is handy.
  * Use `valgrind` to check for invalid memory access and leaks
  
[…]
