+++ doc/debian.md
@@ -27,15 +27,19 @@ Now the command `feenox` should be globally available:
 
 ```terminal
 $ feenox
-FeenoX v0.2.219-g9e9ef22 
+FeenoX v0.3
 a cloud-first free no-fee no-X uniX-like finite-element(ish) computational engineering tool
 
-usage: feenox [options] inputfile [replacement arguments] [petsc options]
+usage: ./feenox [options] inputfile [replacement arguments] [petsc options]
 
   -h, --help         display options and detailed explanations of commmand-line usage
   -v, --version      display brief version information and exit
   -V, --versions     display detailed version information
+  -c, --check        validates if the input file is sane or not
   --pdes             list the types of PROBLEMs that FeenoX can solve, one per line
+  --elements_info    output a document with information about the supported element types
+  --linear           force FeenoX to solve the PDE problem as linear
+  --non-linear       force FeenoX to solve the PDE problem as non-linear
 
 Run with --help for further explanations.
 $
