+++ doc/README.md
@@ -24,11 +24,11 @@ As per the [GNU Coding Standards](https://www.gnu.org/prep/standards/standards.h
 
  1. The FeenoX project starts as an offer to an imaginary “request for quotations” that defines [software requirements specifications](https://seamplex.com/feenox/doc/srs.html) for an open source computational tool.
 
-     * [Software Requirements specifications](https://seamplex.com/feenox/doc/srs.html) (Fictitious RFQ) [[PDF]](https://www.seamplex.com/feenox/doc/srs.pdf)
+     * [Software Requirements Specification](https://seamplex.com/feenox/doc/srs.html) (Fictitious RFQ) [[PDF]](https://www.seamplex.com/feenox/doc/srs.pdf)
 
  2. Then a fictitious “offer” to the above tender is given in a [software design specifications](https://seamplex.com/feenox/doc/srs.html) document that explains the design decisions and features included in FeenoX.
 
-     * [Software Design specifications](https://seamplex.com/feenox/doc/sds.html) (Imaginary FeenoX's offer) [[PDF]](https://www.seamplex.com/feenox/doc/sds.pdf)
+     * [Software Design Specification](https://seamplex.com/feenox/doc/sds.html) (Imaginary FeenoX's offer) [[PDF]](https://www.seamplex.com/feenox/doc/sds.pdf)
 
  
 ::::::::: {.only-in-format .html .markdown}
