### 2024-05-02 #5846 @kyleniemeyer
Looks like we missed closing this one - congratulations @gtheler on your article's publication in JOSS!

And many thanks to @vijaysm, @AnjaliSandip, and @chennachaos for reviewing this.
